%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from ar.R
%%
\name{inla.ar}
\alias{inla.ar.pacf2phi}
\alias{ar.pacf2phi}
\alias{pacf2phi}
\alias{inla.ar.phi2pacf}
\alias{ar.phi2pacf}
\alias{phi2pacf}

\title{Convert between parameterizations for the AR(p) model}

\description{These functions convert between the AR(p) coefficients \code{phi} and
             the partial autorcorrelation coefficients \code{pacf}.
             The \code{phi}-parameterization is the same as used for \code{arima}-models in \code{R}; see \code{?arima}
             and the parameter-vector \code{a} in \code{Details}.}
\usage{
   phi = inla.ar.pacf2phi(pac)
   pac = inla.ar.phi2pacf(phi)
}

\arguments{
  \item{pac}{The partial autorcorrelation coefficients}
  \item{phi}{The AR(p) parameters \code{phi}}
}
\value{
  \code{inla.ar.pacf2phi}  returns \code{phi} for given \code{pacf}.
  \code{inla.ar.phi2pacf}  returns \code{pac} for given \code{phi}.
}
\author{Havard Rue \email{hrue@math.ntnu.no}}
\examples{
 pac = runif(5)
 phi = inla.ar.pacf2phi(pac)
 pac2 = inla.ar.phi2pac(phi)
 print(paste("Error:", max(abs(pac2-pac))))
}
