% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.install.R
\name{inla.binary.install}
\alias{inla.binary.install}
\alias{binary.install}
\title{Install alternative binary builds. (OBSOLETE: use 'inla.stiles.install()' instead)}
\usage{
inla.binary.install(
  os = c("CentOS Linux-6", "CentOS Linux-7", "CentOS Linux-8", "CentOS Stream-8",
    "Rocky Linux-8", "Rocky Linux-9", "Rocky Linux-10", "Manjaro Linux-", "Fedora-33",
    "Fedora-34", "Fedora Linux-35", "Fedora Linux-36", "Fedora Linux-37",
    "Fedora Linux-38", "Fedora Linux-39", "Fedora Linux-40", "Fedora Linux-41",
    "Fedora Linux-42", "Fedora Linux-43", "Fedora Linux-44", "Ubuntu-16.04",
    "Ubuntu-18.04", "Ubuntu-20.04", "Ubuntu-22.04", "Ubuntu-24.04", "Ubuntu-25.04",
    "Ubuntu-26.04"),
  path = NULL,
  verbose = TRUE,
  md5.check = TRUE,
  secure.http = TRUE,
  list.only = FALSE,
  version = NULL
)
}
\arguments{
\item{os}{If \code{os} is given, install binary build for this \code{os}.
If \code{os} is not given, chose \code{os} interactively among available
builds.}

\item{path}{character. The install path. If \code{NULL} the path is derived
from the installed \code{INLA} package}

\item{verbose}{Logical. Verbose output if \code{TRUE}}

\item{md5.check}{Logical. If \code{TRUE}, stop if md5-checksum-file is not
present or md5-checksum fail. If \code{FALSE}, ignore md5-checksum check.}

\item{secure.http}{Logical. Use secure http (ie \verb{https://}) or
\verb{http://}}

\item{list.only}{Logical. If \code{TRUE}, only return character vector of
available builds and exit. Default is \code{FALSE}.}

\item{version}{character. The version of INLA binaries to install. If \code{NULL}
(default) then the version matching the currently installed INLA package
is used.}
}
\value{
By default, return \code{TRUE} if installation was successful and \code{FALSE}
if not. If \code{list.only = TRUE} return character vector of available builds
(invisibly).
}
\description{
Install a new binary for \code{os} unless \code{missing(os)}, for which the
\code{os} is chosen interactively among the available builds.
}
\examples{

   \dontrun{
     inla.binary.install()
     inla.binary.install(os = "CentOS Linux-7")
     inla.binary.install(os = "CentOS Linux-7",  path = "~/local/bin/inla.binary")
     inla.binary.install(list.only = TRUE)
   }
 
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
