% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{BivMetaAnalysis}
\alias{BivMetaAnalysis}
\title{Bivariate Meta Analysis}
\format{
A data frame with 92 observations on the following 9 variables.
\describe{
\item{N}{a numeric vector}
\item{Y}{a numeric vector}
\item{diid}{a numeric vector}
\item{lag.tp}{a numeric vector}
\item{lag.tn}{a numeric vector}
\item{ct.tp}{a numeric vector}
\item{ct.tn}{a numeric vector}
\item{mr.tp}{a numeric vector}
\item{mr.tn}{a numeric vector}
}
}
\usage{
BivMetaAnalysis
}
\description{
Data are taken from a meta-analysis to compare the utility of three types of
diagnostic imaging - lymphangiography (LAG), computed tomography (CT) and
magnetic resonance (MR) - to detect lymph node metastases in patients with
cervical cancer. The dataset consists of a total of 46 studies: the first 17
for LAG, the following 19 for CT and the last 10 for MR.
}
\examples{

data(BivMetaAnalysis)
}
\references{
J. Scheidler and H. Hricak and K. K. Yu and L. Subak and M. R.
Segal,"Radiological evaluation of lymph node metastases in patients with
cervical cancer: a meta-analysis",JAMA 1997
}
\keyword{datasets}
