% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changelog.R
\name{inla.changelog}
\alias{inla.changelog}
\alias{changelog}
\alias{inla.changes}
\title{inla.changelog}
\usage{
inla.changelog()
}
\description{
List the recent changes in the inla-program and its R-interface
}
\seealso{
\code{\link{inla}}
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
