% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloglike.R
\name{cloglike}
\alias{cloglike}
\alias{inla.cloglike.define}
\alias{cloglike.define}
\title{cloglike models}
\usage{
inla.cloglike.define(model = NULL, shlib = NULL, debug = FALSE, ...)
}
\arguments{
\item{model}{The name of the loglike function}

\item{shlib}{Name of the compiled object-file with \code{model}}

\item{debug}{Logical. Turn on/off debugging}

\item{...}{Additional arguments, required by \code{inla.cloglike.define()} to be named arguments}
}
\description{
A framework for defining likelihood models in C
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
