\name{inla.collect.fixed}
\alias{inla.collect.fixed}
\alias{inla.collect.cpo}
\alias{inla.collect.dic}
\alias{inla.collect.fixed}
\alias{inla.collect.hyperpar}
\alias{inla.collect.mlik}
\alias{inla.collect.predictor}
\alias{inla.collect.random} 

\title{ Auxiliary function used to collect results }

\description{ Auxiliary functions used in \code{inla.collect.results()}
to collect the different elements in the results directory }

\author{Sara Martino \email{martino@math.ntnu.no} }

\keyword{ internal }
