% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.results.R
\name{inla.collect.results}
\alias{inla.collect.results}
\alias{collect.results}
\title{Collect results from a inla-call}
\usage{
inla.collect.results(
  results.dir,
  debug = FALSE,
  only.hyperparam = FALSE,
  file.log = NULL,
  file.log2 = NULL,
  silent = inla.getOption("silent")
)
}
\arguments{
\item{results.dir}{The directory where the results of the inla run are
stored}

\item{debug}{Logical. If \code{TRUE} some debugging information are printed}

\item{only.hyperparam}{Binary variable indicating wheather only the results
for the hyperparameters should be collected}

\item{file.log}{Character. The filename, if any, of the logfile for the
internal calculations}

\item{file.log2}{Character. The filename, if any, of the logfile2 for the
internal calculations}

\item{silent}{Internal use only}
}
\value{
The function returns an object of class \code{"inla"}, see the help
file for \code{inla} for details.
}
\description{
\code{inla.collect.results} collect results from a inla-call
}
\details{
This function is mainly used inside \code{inla} to collect results after
running the inla function. It can also be used to collect results into R
after having run an inla section outside R.
}
