\name{common.mean.section}
\alias{common.mean.section}
\alias{data.section}
\alias{discretize.vector}
\alias{ffield.section}
\alias{inla.section}
\alias{linear.section}
\alias{predictor.section}
\alias{problem.section}

\title{Build the ini file}
\description{
  These are internal functions used by \code{inla} to build the .ini
  file.

  Normally not called directly
}




\author{ Sara Martino \email{martino@math.ntnu.no}}
\keyword{internal }

