% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.results.R
\name{inla.compare.results}
\alias{inla.compare.results}
\alias{compare.results}
\title{Compare INLA and MCMC results}
\usage{
inla.compare.results(dir.inla = NULL, dir.mcmc = NULL)
}
\arguments{
\item{dir.inla}{The directory with the INLA results}

\item{dir.mcmc}{The directory with the MCMC results}
}
\value{
Return nothing. This is an interactive function.

This function is OBSOLETE
}
\description{
A small utility to compare INLA and MCMC results (OBSOLETE)
}
\examples{

## See demo("Tokyo-compare")

}
\author{
Havard Rue \email{hrue@r-inla.org}
}
\keyword{internal}
