%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from contributed-1.R
%%
\name{contrib.sd}
\alias{contrib.sd}
\alias{inla.contrib.sd}
\title{Computes the standard deviation for the structured (random) effects in an INLA model}
\description{Computes the posterior distribution of the standard deviations for the structured 
             (random) effects in an INLA model, starting from the default output based on the precisions}
\usage{
    inla.contrib.sd(model, nsamples=1000)
}
\arguments{
    \item{model}{An INLA model, fitted calling the \code{inla()}-function. The formula specified for the model 
            should include at least one structured (random) effect in the form \code{f(variable, model="iid")}.}
    \item{nsamples}{The number of simulations from the posterior distribution of the standard deviations
                    used to compute the summary statistics}
}
\value{
     \code{inla.contrib.sd} returns a matrix \code{samples} including the simulated values from the
                            posterior distributions as well as a summary table \code{hyper} reporting 
                            the mean, sd and 95\% credible interval for the posterior distributions of each random effect.
}
\author{Gianluca Baio \email{gianluca@stats.ucl.ac.uk}}
\seealso{
    \code{\link{inla}}
}
\examples{
# Data generation
n=12
Ntrials = sample(c(80:100), size=n, replace=TRUE)
eta = rnorm(n,0,0.5)
prob = exp(eta)/(1 + exp(eta))
y = rbinom(n, size=Ntrials, prob = prob)
data=data.frame(y=y,z=1:n)
formula=y~f(z,model="iid")
m=inla(formula,data=data,family="binomial",Ntrials=Ntrials)
summary(m)
s=inla.contrib.sd(m)
s$hyper
hist(s$samples,xlab="standard deviation for z",main="")
}
