%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.bgev.default}
\alias{control.bgev.default}
\alias{inla.control.bgev.default}
\alias{inla.set.control.bgev.default.default}
\alias{set.control.bgev.default.default}
\alias{control.bgev.default.default}
\title{Control variables in control.bgev.default}
\description{Control variables in \code{control.bgev.default} for use in \code{inla}}
\usage{
inla.set.control.bgev.default.default(...)
control.bgev.default(beta.ab, q.location, q.mix, q.spread)
}
\arguments{
\item{...}{Possible arguments}
\item{q.location}{   The quantile level for the location parameter}
\item{q.spread}{   The quantile level for the spread parameter (must be < 0.5)}
\item{q.mix}{   The lower and upper quantile level for the mixing function}
\item{beta.ab}{   The parameters a and b in the Beta mixing function}
}
\value{
   The \code{control.bgev}-list is set within the corresponding \code{control.family}-list as control parameters to the \code{family="bgev"}
The function \code{control.bgev.default} is used to TAB-complete arguments and returns a list of given arguments.
The function \code{inla.set.control.bgev.default.default} returns a list with all the default values of all parameters within this control statement.
}
\seealso{
\code{\link{control.update}}, \code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.scopy}}, \code{\link{control.mix}}, \code{\link{control.pom}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.lp.scale}}, \code{\link{control.pardiso.default}}, \code{\link{control.bgev.default}}, \code{\link{control.family}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
