% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.expert}
\alias{control.expert}
\alias{inla.set.control.expert.default}
\title{control.expert}
\usage{
control.expert(
  cpo.manual = FALSE,
  cpo.idx = -1,
  disable.gaussian.check = FALSE,
  jp = NULL,
  globalconstr = list(A = NULL, e = NULL),
  opt.solve = FALSE,
  opt.storage = FALSE,
  opt.num.threads = TRUE,
  memory.alignment = -1,
  blas.num.threads = NULL
)

inla.set.control.expert.default(...)
}
\arguments{
\item{cpo.manual}{A boolean variable to decide if the inla-program is to be
runned in a manual-cpo-mode. (EXPERT OPTION: DO NOT USE)}

\item{cpo.idx}{The index/indices of the data point(s) to remove. (EXPERIMENTAL
OPTION: DO NOT USE)}

\item{disable.gaussian.check}{Disable the check for fast computations with a
Gaussian likelihood and identity link (default \code{FALSE})}

\item{jp}{An object of class \code{inla.jp} defining a joint prior}

\item{globalconstr}{Add a global constraint (see \code{?f} and argument
\code{extraconstr}). Note that a global constraint does NOT
correct the normalisation constant.
(EXPERIMENTAL OPTION)}

\item{opt.solve}{Store also \code{L^T} to optimize linear solves (TAUCS only).
Might benefite larger models. (EXPERIMENTAL OPTION)}

\item{opt.storage}{Make sure \code{L} is stored sequencially (TAUCS only).
Might benefite larger models. (EXPERIMENTAL OPTION)}

\item{opt.num.threads}{Reduce the number of threads adaptively if needed at
critical places to optimize wall-clock time}

\item{memory.alignment}{Memory alignment for alloc (Linux and Mac
only). One of -1 (no alignment correction done),  0 (chose automatically),
16, 32 or 64 bytes (EXPERIMENTAL OPTION)}

\item{blas.num.threads}{Allow setting a fixed value for the number of threads
for BLAS functions, most often 1. If set to 0, then it will be determined
adaptively. If set to 'NULL' then its value it will be determined by
'inla.getOption("blas.num.threads")' (which defaults to 0).}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\seealso{
Other control:
\code{\link[=control.bgev]{control.bgev()}},
\code{\link[=control.compute]{control.compute()}},
\code{\link[=control.family]{control.family()}},
\code{\link[=control.fixed]{control.fixed()}},
\code{\link[=control.gcpo]{control.gcpo()}},
\code{\link[=control.group]{control.group()}},
\code{\link[=control.hazard]{control.hazard()}},
\code{\link[=control.inla]{control.inla()}},
\code{\link[=control.lincomb]{control.lincomb()}},
\code{\link[=control.link]{control.link()}},
\code{\link[=control.lp.scale]{control.lp.scale()}},
\code{\link[=control.mix]{control.mix()}},
\code{\link[=control.mode]{control.mode()}},
\code{\link[=control.numa]{control.numa()}},
\code{\link[=control.pom]{control.pom()}},
\code{\link[=control.predictor]{control.predictor()}},
\code{\link[=control.scopy]{control.scopy()}},
\code{\link[=control.sem]{control.sem()}},
\code{\link[=control.stiles]{control.stiles()}},
\code{\link[=control.taucs]{control.taucs()}},
\code{\link[=control.update]{control.update()}},
\code{\link[=control.vb]{control.vb()}}
}
\concept{control}
