% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.hazard}
\alias{control.hazard}
\alias{inla.set.control.hazard.default}
\title{control.hazard}
\usage{
control.hazard(
  model = "rw1",
  hyper = NULL,
  fixed = FALSE,
  initial = NULL,
  prior = NULL,
  param = NULL,
  constr = TRUE,
  diagonal = NULL,
  n.intervals = 15,
  cutpoints = NULL,
  strata.name = NULL,
  scale.model = NULL
)

inla.set.control.hazard.default(...)
}
\arguments{
\item{model}{The model for the baseline hazard model. One of 'rw1', 'rw2' or
'iid'. (Default 'rw1'.)}

\item{hyper}{The definition of the hyperparameters.}

\item{fixed}{(OBSOLETE!) A boolean variable; is the precision for 'model'
fixed? (Default FALSE.)}

\item{initial}{(OBSOLETE!) The initial value for the precision.}

\item{prior}{(OBSOLETE!) The prior distribution for the precision for
'model'}

\item{param}{(OBSOLETE!) The parameters in the prior distribution}

\item{constr}{A boolean variable; shall the  'model' be constrained to sum
to zero?}

\item{diagonal}{An extra constant added to the diagonal of the precision
matrix}

\item{n.intervals}{Number of intervals in the baseline hazard. (Default 15)}

\item{cutpoints}{The cutpoints to use. If not specified the they are compute
from 'n.intervals' and the maximum length of the interval. (Default NULL)}

\item{strata.name}{The name of the stratefication variable for the baseline
hazard in the data.frame}

\item{scale.model}{Scale the baseline hazard model (RW1, RW2) so the
generalized variance is 1. (Default
\code{inla.getOption("scale.model.default")}.)}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\seealso{
Other control:
\code{\link[=control.bgev]{control.bgev()}},
\code{\link[=control.compute]{control.compute()}},
\code{\link[=control.expert]{control.expert()}},
\code{\link[=control.family]{control.family()}},
\code{\link[=control.fixed]{control.fixed()}},
\code{\link[=control.gcpo]{control.gcpo()}},
\code{\link[=control.group]{control.group()}},
\code{\link[=control.inla]{control.inla()}},
\code{\link[=control.lincomb]{control.lincomb()}},
\code{\link[=control.link]{control.link()}},
\code{\link[=control.lp.scale]{control.lp.scale()}},
\code{\link[=control.mix]{control.mix()}},
\code{\link[=control.mode]{control.mode()}},
\code{\link[=control.numa]{control.numa()}},
\code{\link[=control.pom]{control.pom()}},
\code{\link[=control.predictor]{control.predictor()}},
\code{\link[=control.scopy]{control.scopy()}},
\code{\link[=control.sem]{control.sem()}},
\code{\link[=control.stiles]{control.stiles()}},
\code{\link[=control.taucs]{control.taucs()}},
\code{\link[=control.update]{control.update()}},
\code{\link[=control.vb]{control.vb()}}
}
\concept{control}
