
\name{control.lincomb}
\alias{control.lincomb}
\alias{inla.control.lincomb}
\title{Control variables in control.lincomb}
\description{Control variables in \code{control.lincomb} for use in \code{inla}}
\usage{
inla(formula, ..., control.lincomb=list( <arg>=<value>[, <arg>=<value>]))
}
\arguments{
\item{precision}{  The precision for the artificial tiny noise. Default 1e09.}
\item{verbose}{  Use verbose mode for linear combinations if verbose model is set globally. (Default TRUE)}
}
\seealso{
\code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.re}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.family}}, \code{\link{control.data}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.results}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
