% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.mix}
\alias{control.mix}
\alias{inla.set.control.mix.default}
\title{control.mix}
\usage{
control.mix(
  model = NULL,
  hyper = NULL,
  initial = NULL,
  fixed = NULL,
  prior = NULL,
  param = NULL,
  npoints = 101,
  integrator = "default"
)

inla.set.control.mix.default(...)
}
\arguments{
\item{model}{The model for the random effect. Currently, only
\code{model='gaussian'} is implemented}

\item{hyper}{Definition of the hyperparameter(s) for the random effect model
chosen}

\item{initial}{(OBSOLETE!) The initial value(s) for the hyperparameter(s)}

\item{fixed}{(OBSOLETE!) A boolean variable if hyperparmater(s) is/are fixed
or random}

\item{prior}{(OBSOLETE!) The name of the prior distribution(s) for the
hyperparmater(s)}

\item{param}{(OBSOLETE!) The parameters for the prior distribution(s) for the
hyperparmater(s)}

\item{npoints}{Number of points used to do the numerical integration (default
101)}

\item{integrator}{The integration scheme to use (\code{default},
\code{quadrature}, \code{simpson})}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\details{
The \code{control.mix} list is set within the corresponding \code{control.family}-list a the mixture of the likelihood is likelihood specific. (This option is EXPERIMENTAL.)
}
\seealso{
Other control:
\code{\link[=control.bgev]{control.bgev()}},
\code{\link[=control.compute]{control.compute()}},
\code{\link[=control.expert]{control.expert()}},
\code{\link[=control.family]{control.family()}},
\code{\link[=control.fixed]{control.fixed()}},
\code{\link[=control.gcpo]{control.gcpo()}},
\code{\link[=control.group]{control.group()}},
\code{\link[=control.hazard]{control.hazard()}},
\code{\link[=control.inla]{control.inla()}},
\code{\link[=control.lincomb]{control.lincomb()}},
\code{\link[=control.link]{control.link()}},
\code{\link[=control.lp.scale]{control.lp.scale()}},
\code{\link[=control.mode]{control.mode()}},
\code{\link[=control.numa]{control.numa()}},
\code{\link[=control.pom]{control.pom()}},
\code{\link[=control.predictor]{control.predictor()}},
\code{\link[=control.scopy]{control.scopy()}},
\code{\link[=control.sem]{control.sem()}},
\code{\link[=control.stiles]{control.stiles()}},
\code{\link[=control.taucs]{control.taucs()}},
\code{\link[=control.update]{control.update()}},
\code{\link[=control.vb]{control.vb()}}
}
\concept{control}
