%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.pardiso.default}
\alias{control.pardiso.default}
\alias{inla.control.pardiso.default}
\alias{inla.set.control.pardiso.default.default}
\alias{set.control.pardiso.default.default}
\alias{control.pardiso.default.default}
\title{Control variables in control.pardiso.default}
\description{Control variables in \code{control.pardiso.default} for use in \code{inla}}
\usage{
inla.set.control.pardiso.default.default(...)
control.pardiso.default(debug, nrhs, parallel.reordering, verbose)
}
\arguments{
\item{...}{Possible arguments}
\item{verbose}{   Show detailed output (default FALSE)}
\item{debug}{   Show internal debug output (default FALSE)}
\item{parallel.reordering}{   Do reordering in parallel (default TRUE)}
\item{nrhs}{   Number of right-hand sides to solve for in parallel (\code{-1}  will determine this adapative)}

}
\value{
   Extra options controlling the PARDISO library
The function \code{control.pardiso.default} is used to TAB-complete arguments and returns a list of given arguments.
The function \code{inla.set.control.pardiso.default.default} returns a list with all the default values of all parameters within this control statement.
}
\seealso{
\code{\link{control.update}}, \code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.scopy}}, \code{\link{control.mix}}, \code{\link{control.pom}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.lp.scale}}, \code{\link{control.pardiso.default}}, \code{\link{control.bgev.default}}, \code{\link{control.family}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
