% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.taucs}
\alias{control.taucs}
\alias{inla.set.control.taucs.default}
\title{control.taucs}
\usage{
control.taucs(min.block.size = 4, block.size = 64)

inla.set.control.taucs.default(...)
}
\arguments{
\item{block.size}{Preferred number of rhs's in each parallel solve. Used to
split many rhs's into threads (on the outer level)}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\details{
Extra options controlling the TAUCS library
}
\seealso{
Other control:
\code{\link[=control.bgev]{control.bgev()}},
\code{\link[=control.compute]{control.compute()}},
\code{\link[=control.expert]{control.expert()}},
\code{\link[=control.family]{control.family()}},
\code{\link[=control.fixed]{control.fixed()}},
\code{\link[=control.gcpo]{control.gcpo()}},
\code{\link[=control.group]{control.group()}},
\code{\link[=control.hazard]{control.hazard()}},
\code{\link[=control.inla]{control.inla()}},
\code{\link[=control.lincomb]{control.lincomb()}},
\code{\link[=control.link]{control.link()}},
\code{\link[=control.lp.scale]{control.lp.scale()}},
\code{\link[=control.mix]{control.mix()}},
\code{\link[=control.mode]{control.mode()}},
\code{\link[=control.numa]{control.numa()}},
\code{\link[=control.pom]{control.pom()}},
\code{\link[=control.predictor]{control.predictor()}},
\code{\link[=control.scopy]{control.scopy()}},
\code{\link[=control.sem]{control.sem()}},
\code{\link[=control.stiles]{control.stiles()}},
\code{\link[=control.update]{control.update()}},
\code{\link[=control.vb]{control.vb()}}
}
\concept{control}
