%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from coxph.R
%%
 Export: inla.coxph
 \name{inla.coxph}
 \alias{inla.coxph}
 \alias{coxph}
 
 \title{Convert a Cox proportional hazard model into Poisson regression}
 
 \description{This function is used to convert manually a Cox proportional hazard model into Poisson regression}
 \usage{
     inla.coxph(formula, data, control.hazard = list())
 }
 
 \arguments{
   \item{formula}{The formula for the coxph model where the reponse must be a \code{inla.surv}-object.}
   \item{data}{All the data used in the formula,  as a list.}
   \item{control.hazard}{Control the model for the baseline-hazard.}
}
\value{
      A list of new expanded variables to be used in the \code{inla}-call.
}
\author{Havard Rue \email{hrue@math.ntnu.no}}
 
\examples{
n = 1000
x = runif(n)
lambda = exp(1+x)
y = rexp(n, rate=lambda)
event = rep(1,n)
data = list(y=y, event=event, x=x)
y.surv = inla.surv(y, event)
p = inla.coxph(yy.surv ~ x,
        list(y.surv = y.surv,  x=x))
model = inla(p$formula, 
        family = p$family, 
        data=p$data,
        E = p$E,
        .internal = p$.internal)
summary(model)
}
