% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.new.R
\name{inla.dev.new}
\alias{inla.dev.new}
\title{Opens a new device}
\usage{
inla.dev.new(...)
}
\arguments{
\item{...}{Optional arguments to \code{\link[=dev.new]{dev.new()}}}
}
\value{
The value of \code{\link[=dev.new]{dev.new()}} if not running RStudio, otherwise
\code{NULL}
}
\description{
Open a new device using \code{\link[=dev.new]{dev.new()}} unless using RStudio
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
