% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{Drivers}
\alias{Drivers}
\title{Time series with seasonal effect}
\format{
A data frame with 204 observations on the following 4 variables.
\describe{
\item{y}{Number of deaths}
\item{belt}{Indicator of weather the belt was compulsory to use
(1) or not (0)}
\item{trend}{time (in months)}
\item{seasonal}{time (in months)}
}
}
\description{
Montly total of car drivers killed or several injuried in England from
January 1969 to December 1984
}
\details{
NB: The last 12 lines of the data set have the first column set to
\code{NULL} since these data where not observed but we want to predict them.
}
\examples{

data(Drivers)
}
\references{
Rue, H and Held, L. (2005) \emph{Gaussian Markov Random Fields -
Theory and Applications} Chapman and Hall
}
\keyword{datasets}
