% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{Epil}
\alias{Epil}
\title{Repeated measures on Poisson counts}
\format{
A data frame with 236 observations on the following 7 variables.
\describe{
\item{y}{Number of seizures}
\item{Trt}{indicator for the presence of treatment}
\item{Base}{8-week baseline seizure counts}
\item{Age}{Age of the patient}
\item{V4}{indicator variable for the 4th visit.}
\item{rand}{a numeric vector}
\item{Ind}{indicator for the specific patient}
}
}
\source{
WinBUGS/OpenBUGS Manual Examples Vol I
}
\description{
Seizure counts in a randomised trial of anti-convulsant therpay in epilepsy
for 59 patients.
}
\examples{

data(Epil)
}
\keyword{datasets}
