%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from fgn.R
%%
 \name{fgn}
 \alias{inla.fgn}
 \alias{fgn}

 \title{Return the coefficients in the 3-component AR(1) mixture representing FGN(H)}

 \description{This function will return the coefficients in the 3-component AR(1)
              mixture representing FGN(H)}

 \usage{
     inla.fgn(H, K=3L, lag.max = NULL)
 } 

 \arguments{
   \item{H}{The Hurst coeffcient (0<H<1),  or a vector of those}
   \item{K}{The number of components in representation,  must be 3L or 4L}
   \item{lag.max}{If positive integer, return the coeffcients implicitely as the ACF
                  from 0 to \code{lag.max}}
  }
 \value{
  If \code{is.null(lag.max)},  then return a
  named matrix with 2*K+1 columns, where the first column is \code{H},
  columns \code{1+1:K} are lag one correlations (or phi's),
  and columns \code{1+K+1:K} are the weights.
  If \code{lag.max > 0},  return a matrix with the ACFs. One row
  for each H.

  This function is EXPERIMENTAL!!!
 }
 \author{Havard Rue \email{hrue@r-inla.org}}

 \examples{
     r = c(inla.fgn(0.7))
     r_m = inla.fgn(seq(0.6, 0.8, by=0.01))
 } 
