% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{Germany}
\alias{Germany}
\title{Disease Mapping}
\format{
A data frame with 544 observations on the following 4 variables.
\describe{
\item{region}{Region of Germany}
\item{E}{Fixed quantity which accounts for number of people
in the district (offset)}
\item{Y}{Number of cases}
\item{x}{covariate measuring smoking consumption}
}
}
\description{
Cases of Oral cavity cancer in Germany from 1986-1990
}
\examples{

data(Germany)
}
\references{
Rue, H and Held, L. (2005) \emph{Gaussian Markov Random Fields -
Theory and Applications} Chapman and Hall
}
\keyword{datasets}
