%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from get.R
%%
\name{inla.get}
\alias{inla.get}
\alias{summary.inla.get}
\alias{print.inla.get}
\title{Get and view submitted jobs}
\description{
Functions to get and view submitted jobs
}
\usage{
inla.get(id = NULL, remove = TRUE)
\method{summary}{inla.get}(object,...)
\method{print}{inla.get}(x,...)
}
\arguments{
  \item{id}{The job-id which is the output from \code{inla} when the job is submitted,  the
            job-number or job-name}
  \item{remove}{Logical If FALSE, leave the job on the server after getting-it,  otherwise remove it.}
  \item{x}{An \code{inla.get}-object,  which is the output from \code{inla.get()}}
  \item{object}{An \code{inla.get}-object,  which is the output from \code{inla.get()}}
  \item{...}{ other arguments.}
}
\details{
\code{inla.get} returns a list of all current jobs if \code{id=NULL},  otherwise it will get the results
from the job with job-id/number/name \code{id}.
}
\value{
  \code{inla.get()} returns an \code{inla.get}-object with information about current jobs.
}
\author{Havard Rue}
\seealso{ \code{\link{inla}} }
\examples{
\dontrun{
r = inla(y~1, data = data.frame(y=rnorm(10)), inla.call="submit")
inla.get()
r = inla.get(r)
}
}
