\name{get.results.fixed}
\alias{get.results.fixed}
\alias{get.results.fixed.binary}
\alias{get.results.hyperpar}
\alias{get.results.hyperpar.binary}
\alias{get.results.inverse.predictor}
\alias{get.results.inverse.predictor.binary}
\alias{get.results.linear.predictor}
\alias{get.results.linear.predictor.binary}
\alias{get.results.random}
\alias{get.results.random.binary}

\title{Read results files }
\description{Function used internally by the \code{inla} function to
  recover data after running the inla program.}

\author{sara martino \email{martino@math.ntnu.no}}


\keyword{internal}
