% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.convert.R
\name{graph.convert}
\alias{graph.convert}
\alias{inla.geobugs2inla}
\alias{geobugs2inla}
\title{INLA utility functions}
\usage{
inla.geobugs2inla(adj, num, graph.file = "graph.dat")
}
\arguments{
\item{adj}{A vector listing the ID numbers of the adjacent areas for each
area. This is a sparse representation of the full adjacency matrix for the
study region, and can be generated using the Adjacency Tool from the Map
menu in GeoBUGS.}

\item{num}{A vector of length N (the total number of areas) giving the
number of neighbours n.i for each area.}

\item{graph.file}{Name of the file of the new graph in the INLA format.}
}
\value{
The return value is the name of the graph-file created.
}
\description{
Various utility functions for INLA
}
\note{
These are all the same function, and the two different names are due
to backward-compatibility
}
\seealso{
\code{\link[=inla]{inla()}}, \code{\link[=inla.surv]{inla.surv()}},
\code{\link[=hyperpar.inla]{hyperpar.inla()}}
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
\keyword{graph}
