%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from graph.matrix.R
%%
\name{graph2matrix}
\alias{graph2matrix}
\alias{inla.graph2matrix}
\alias{inla.matrix2graph}
\alias{spy}
\alias{inla.spy}
\title{Convert between a (possible) sparse symmetric matrix and an \code{inla.graph}-object}
\description{Convert between a (possible) sparse symmetric matrix and an \code{inla.graph}-object}
\usage{
Q = inla.graph2matrix(graph)
g = inla.matrix2graph(Q)
inla.spy(graph, ...)
}
\arguments{
    \item{graph}{An \code{inla.graph}-object, a filename containing an \code{inla.graph}-object,  or a (sparse) symmetric matrix}
    \item{g}{An \code{inla.graph}-object}
    \item{Q}{An (possible) sparse symmtric matrix}
}
\value{
     \code{inla.graph2matrix} returns a sparse symmetric matrix where the non-zero pattern is defined by the \code{graph}.
     The \code{inla.spy} function, plots a binary image of a \code{graph}.

     The function\code{inla.read.graph}-function is the same as the function \code{inla.matrix2graph}.
}
\author{Havard Rue \email{hrue@math.ntnu.no}}
\seealso{
    \code{\link{inla.read.graph}}
}
\examples{
n = 50
Q = matrix(0, n, n)
idx = sample(1:n, 2*n, replace=TRUE)
Q[idx, idx] = 1
diag(Q) = 1
inla.dev.new()
inla.spy(Q)

g = inla.matrix2graph(Q)
inla.dev.new()
inla.spy(g)

g.file = inla.write.graph(g)
inla.dev.new()
inla.spy(g.file)
}
