%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from iidkd.R
%%
\name{inla.iidkd.sample}
\alias{inla.iidkd.sample}
\alias{iidkd.sample}

\title{Provide samples from the \code{iidkd} component}
\description{
 This function provide samples of the \code{iidkd} component 
 using more interpretable parameters
}
\usage{
inla.iidkd.sample(n = 10^4, result, name)
}

\arguments{
\item{n}{Number of samples to use}
\item{result}{An object of class \code{inla}, ie a result of a call to \code{inla()}}
\item{name}{The name of the \code{iidkd} component}
}
\value{A list of sampled matrices,  with correlations on the off-diagonal and
standard-deviations on the diagonal}
\seealso{\code{inla.doc("iidkd")}}
\author{Havard Rue \email{hrue@r-inla.org}}
