% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{package}
\name{INLA-package}
\alias{INLA-package}
\alias{INLA}
\title{INLA}
\usage{
INLA()
}
\description{
Package to perform full Bayesian analysis on latent Gaussian models using
Integrated Nested Laplace Approximations.

See \url{https://www.r-inla.org/} for further details.
}
\seealso{
\code{\link[=inla]{inla()}}
}
\author{
\strong{Maintainer}: Havard Rue \email{hrue@r-inla.org}

Authors:
\itemize{
  \item Havard Rue \email{hrue@r-inla.org}
  \item Finn Lindgren \email{finn.lindgren@gmail.com}
  \item Elias Teixeira Krainski \email{elias@r-inla.org}
  \item Esmail Abdul Fattah \email{esmail.abdulfattah@kaust.edu.sa}
  \item Sara Martino \email{saramartino0@gmail.com}
}

Other contributors:
\itemize{
  \item Haakon Bakka \email{bakka@r-inla.org} [contributor]
  \item Daniel Simpson \email{dp.simpson@gmail.com} [contributor]
  \item Andrea Riebler \email{andrea.riebler@math.ntnu.no} [contributor]
  \item Geir-Arne Fuglstad \email{geirarne.fulgstad@gmail.com} [contributor]
  \item Cristian Chiuchiolo \email{cristian.chiuchiolo@kaust.edu.sa} [contributor]
}

}
