%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from inla.call.R
%%
\name{inla.call.builtin}
\alias{inla.call.builtin}

\title{Return the path to the (default) \code{inla}-program}
\description{
 Return the path to the (default) \code{inla}-program. This is
 also the default value for the argument \code{inla.call}
}
\usage{
inla.call.builtin()
}
\value{This function returns the path to the default \code{inla}-program.}
\author{Havard Rue \email{hrue@r-inla.org}}
