% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde.common.R
\name{inla.dBind}
\alias{inla.dBind}
\alias{inla.dBind-deprecated}
\title{Build a block-diagonal sparse matrix.}
\usage{
inla.dBind(...)
}
\arguments{
\item{\dots}{A list of square or rectangular matrices.}
}
\value{
A sparse matrix.
}
\description{
Build a block-diagonal sparse matrix.  Obsolete wrapper for \code{bdiag()}.
}
\examples{
\dontrun{
  inla.dBind(Matrix(1, 2, 1), Matrix(2, 1, 2))
}
bdiag(Matrix(1, 2, 1), Matrix(2, 1, 2))
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
\keyword{internal}
