% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde.common.R
\name{inla.extract.el}
\alias{inla.extract.el}
\alias{inla.extract.el.data.frame}
\alias{inla.extract.el.list}
\alias{inla.extract.el.matrix}
\title{Extract elements by matching name from container objects.}
\usage{
inla.extract.el(M, ...)

\method{inla.extract.el}{matrix}(M, match, by.row = TRUE, ...)

\method{inla.extract.el}{data.frame}(M, match, by.row = TRUE, ...)

\method{inla.extract.el}{list}(M, match, ...)
}
\arguments{
\item{M}{A container object.}

\item{\dots}{Additional arguments, not used.}

\item{match}{A regex defining the matching criterion.}

\item{by.row}{If \code{TRUE}, extract data by row, otherwise by column.}
}
\description{
Extract elements by wildcard name matching from a \code{data.frame},
\code{list}, or \code{matrix}.
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
