% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{inla.generate.colors}
\alias{inla.generate.colors}
\title{Generate text RGB color specifications.}
\usage{
inla.generate.colors(
  color,
  color.axis = NULL,
  color.n = 512,
  color.palette = cm.colors,
  color.truncate = FALSE,
  alpha = NULL
)
}
\arguments{
\item{color}{\code{character}, \code{matrix} or \code{vector}}

\item{color.axis}{The min/max limit values for the color mapping.}

\item{color.n}{The number of colors to use in the color palette.}

\item{color.palette}{A color palette function.}

\item{color.truncate}{If \code{TRUE}, truncate the colors at the color axis
limits.}

\item{alpha}{Transparency/opaqueness values.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[fmesher:fm_generate_colors]{fmesher::fm_generate_colors()}}
instead.

Generates a tex RGB color specification matrix based on a color palette.
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
