% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmesh.R
\name{inla.has_PROJ6}
\alias{inla.has_PROJ6}
\alias{inla.not_for_PROJ6}
\alias{inla.not_for_PROJ4}
\alias{inla.fallback_PROJ6}
\alias{inla.requires_PROJ6}
\title{PROJ6 detection}
\usage{
inla.has_PROJ6()

inla.not_for_PROJ6(fun)

inla.not_for_PROJ4(fun)

inla.fallback_PROJ6(fun)

inla.requires_PROJ6(fun)
}
\value{
For \code{inla.has_PROJ6}, logical;
`TRUE` if PROJ6 is available, `FALSE` otherwise
}
\description{
Detect whether PROJ6 is available for INLA
}
\details{
\code{inla.not_for_PROJ6} is called to warn about using old PROJ4
features even though PROJ6 is available

\code{inla.not_for_PROJ4} is called to give an error when
calling methods that are only available for PROJ6

\code{inla.fallback_PROJ6} is called to warn about falling back
to using old PROJ4 methods when a RPOJ6 method hasn't been implemented

\code{inla.requires_PROJ6} is called to give an error when PROJ6
is required but not available
}
\examples{
inla.has_PROJ6()
}
