% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.components.R
\name{inla.mesh.components}
\alias{inla.mesh.components}
\title{Compute connected mesh subsets}
\usage{
inla.mesh.components(mesh)
}
\arguments{
\item{mesh}{An \code{fm_mesh_2d} object}
}
\value{
A list with elements \code{vertex} and \code{triangle}, vectors of
integer labels for which connected component they belong, and \code{info}, a
\code{data.frame} with columns
\item{component}{Connected component integer label.}
\item{nV}{The number of vertices in the component.}
\item{nT}{The number of triangles in the component.}
\item{area}{The surface area associated with the component.
Component labels are not comparable across
different meshes, but some ordering stability is guaranteed by initiating
each component from the lowest numbered triangle whenever a new component is
initiated.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Compute subsets of vertices and triangles
in an inla.mesh object that are
connected by edges. This function is deprecated from INLA \verb{25.4.10} when
fmesher version \verb{0.3.0.9005} or later is installed, which has
\code{fm_mesh_components()}.
}
\examples{

# Construct two simple meshes:
library(fmesher)
loc <- matrix(c(0, 1, 0, 1), 2, 2)
mesh1 <- fm_mesh_2d(loc = loc, max.edge = 0.1)
bnd <- fm_nonconvex_hull_inla(loc, 0.3)
mesh2 <- fm_mesh_2d(boundary = bnd, max.edge = 0.1)

# Compute connectivity information:
conn1 <- inla.mesh.components(mesh1)
conn2 <- inla.mesh.components(mesh2)
# One component, simply connected mesh
conn1$info
# Two disconnected components
conn2$info
}
\seealso{
\code{\link[fmesher:fm_mesh_2d]{fmesher::fm_mesh_2d()}}, \code{\link[fmesher:fm_rcdt_2d]{fmesher::fm_rcdt_2d()}}
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
