% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{inla.mesh.deriv}
\alias{inla.mesh.deriv}
\title{Directional derivative matrices for functions on meshes.}
\usage{
inla.mesh.deriv(mesh, loc)
}
\arguments{
\item{mesh}{An \code{\link[=inla.mesh]{inla.mesh()}} object.}

\item{loc}{Coordinates where the derivatives should be evaluated.}
}
\value{
\item{A }{The projection matrix, \verb{u(loc_i)=sum_j A_ij w_i}}
\item{dx, dy, dz }{Derivative weight matrices, \verb{du/dx(loc_i)=sum_j dx_ij w_i}, etc.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[fmesher:fm_basis]{fmesher::fm_basis()}} instead.

Calculates directional derivative matrices for functions on
\code{\link[=inla.mesh]{inla.mesh()}} objects.
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
