\name{inla.mesh.query}
\alias{inla.mesh.query}
\title{
  High-quality triangulations
}
\description{
  Query information about an inla.mesh object.
}
\usage{
inla.mesh.query(mesh, ...)
}
\arguments{
  \item{mesh}{
    An \code{inla.mesh} object.
  }
}
\details{
}
\value{
  A list of query results.
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
}
\author{
  Finn Lindgren \email{finn.lindgren@gmail.com}
}
\note{
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
  \code{\link{inla.mesh.create}}, 
  \code{\link{inla.mesh.segment}}, \code{\link{inla.mesh.lattice}}
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ fmesher }
