% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin.R
\name{inla.pin}
\alias{inla.pin}
\alias{inla.unpin}
\title{Control main thread pinning for INLA (experimental)}
\usage{
inla.pin()

inla.unpin()
}
\value{
No value is returned.
}
\description{
Control main thread pinning for INLA (experimental)
}
\details{
\itemize{
\item \code{inla.pin} set OMP variables for pinning
}

\itemize{
\item \code{inla.unpin} unset OMP variables for pinning
}
}
\examples{
inla.pin()
inla.unpin()
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
