%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from inla.remote.R
%%
\name{inla.ssh.copy.id}
\alias{inla.ssh.copy.id}
\alias{ssh.copy.id}

\title{Setup remote-access}

\description{Print the path to the internal script to transfer ssh-keys}

\usage{
inla.ssh.copy.id()
}
\arguments{
}

\value{%%
  The path to the internal script to transfer ssh-keys
}
%%

\author{Havard Rue \email{hrue@math.ntnu.no}}

\examples{
##TODO
}
