% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.R
\name{inla.run.many}
\alias{inla.run.many}
\title{inla.run.many}
\usage{
inla.run.many(
  n.models = 1,
  working.directory = NULL,
  verbose = inla.getOption("verbose"),
  num.threads = inla.getOption("num.threads"),
  cleanup = FALSE
)
}
\arguments{
\item{n.models}{Number of models to run}

\item{working.directory}{Where to store results}

\item{verbose}{Run in verbose mode or not}

\item{num.threads}{Number of threads to be used}

\item{cleanup}{Cleanup result files or leave them behind?}
}
\description{
inla.run.many
}
\details{
A special \code{inla()}-function to run many models (of a particular type)
used by package \code{INLAjoint} (for details,  check the \code{INLAjoint} implementation)
}
