\name{inla.spde.query}
\title{
  SPDE model objects for INLA
}
\description{
  Query an \code{inla.spde} model object.
}
\usage{
## Query properties of an SPDE model:
inla.spde.query(spde, ...)
}
\arguments{
  \item{spde}{
    The \code{inla.spde} object to query.
  }
  \item{...}{
    
  }
}
%\details{
%}
\value{
  An list of query results.
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
}
\author{
  Finn Lindgren \email{finn.lindgren@gmail.com}
}
\note{
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
  \code{\link{inla.spde}}
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ fmesher }
