% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde.common.R
\name{inla.spde.sample}
\alias{inla.spde.sample}
\alias{inla.spde.sample.default}
\alias{inla.spde.sample.inla.spde}
\title{Sample from SPDE models}
\usage{
inla.spde.sample(...)

\method{inla.spde.sample}{default}(precision, seed = NULL, ...)

\method{inla.spde.sample}{inla.spde}(spde, seed = NULL, ...)
}
\arguments{
\item{\dots}{Parameters passed on to other methods.}

\item{precision}{A precision matrix.}

\item{seed}{The seed for the pseudo-random generator.}

\item{spde}{An \code{inla.spde} object.}
}
\description{
Old methods fo sampling from a SPDE model.  For new code, use
\code{\link[=inla.spde.precision]{inla.spde.precision()}} and \code{\link[=inla.qsample]{inla.qsample()}} instead.
}
\seealso{
\code{\link[=inla.spde.precision]{inla.spde.precision()}}, \code{\link[=inla.qsample]{inla.qsample()}}
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
