% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{inla.stack.mexpand}
\alias{inla.stack.mexpand}
\title{Expand observation vectors/matrices in stacks into to a multicolumn matrix for multiple likelihoods}
\usage{
inla.stack.mexpand(..., old.names = "response", new.name = "response")
}
\arguments{
\item{...}{List of stacks that contain vector observations
(existing multilikelihood observation matrices are also permitted)}

\item{old.names}{A vector of strings with the names of the observation vector/matrix for each stack.
If a single string, this is assumed for all the stacks. (default "response")}

\item{new.name}{The name to be used for the expanded observation matrix,
possibly the same as an old name. (default "response")}
}
\value{
a list of modified stacks with multicolumn observations
}
\description{
Internal helper method for \code{inla.stack.join}.
}
\author{
Fabian E. Bachl \email{f.e.bachl@bath.ac.uk} and Finn Lindgren \email{finn.lindgren@gmail.com}
}
\keyword{internal}
