% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inla.tempdir}
\alias{inla.tempdir}
\title{inla.tempdir}
\usage{
inla.tempdir()
}
\description{
inla.tempdir
}
\details{
Return the name, and create, a temporary directory for storing intermediate results.
This function is exported as it is also used by package \code{INLAjoint} (for details, check the \code{INLAjoint}
implementation)
}
