%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from inlaEnv.R
%%
\name{inla.get.inlaEnv}
\alias{inla.get.inlaEnv}
\alias{get.inlaEnv}
\title{Return the internal environment used by INLA}
\description{A function which return the internal environment used by INLA}
\usage{
inla.get.inlaEnv()
}
\arguments{
}

\value{%%
  This function returns the internal environment used by INLA to
  keep internal variables.
}
\author{Havard Rue \email{hrue@r-inla.org}}
