% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{Kidney}
\alias{Kidney}
\title{Kidney infection data}
\format{
A data frame with 76 observations on the following 9 variables.
\describe{
\item{time}{a numeric vector. Time to infection from the
insertion of catheter}
\item{event}{a numeric vector. 1: time of
infection 0: time of censuring }
\item{age}{a numeric vector. Age of
the patient at the time of infection}
\item{sex}{a numeric vector.
Sex of the patient 0: male 1:female}
\item{disease}{a numeric
vector. Type of disease}
\item{dis1}{a numeric vector. Dummy
variable to codify the disease type.}
\item{dis2}{a numeric vector.
Dummy variable to codify the disease type.}
\item{dis3}{a numeric
vector. Dummy variable to codify the disease type.}
\item{ID}{a
numeric vector. Patient code.}
}
}
\description{
Times of infection from the time to insertion of the catheter for 38 kindey
patients using portable dialysis equipment
}
\references{
McGilchrist and C.W. Aisbett (1991), Regression with frailty in
survival analysis, Biometrics,vol.47,pages 461--166.

D.J. Spiegelhalter and A. Thomas and N.G. Best and W.R. Gilks (1995) BUGS:
Bayesian Inference Using Gibbs sampling, Version 0.50., MRC Biostatistics
Unit, Cambridre, England.
}
\keyword{datasets}
