%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from knmodels.R
%%
 \name{inla.knmodels}
 \alias{inla.knmodels}
 \alias{inla.knmodels.sample}
 \title{Spacetime interaction models}
 \description{
    It implements the models in Knorr-Held, L. (2000) 
    with three different constraint approaches: 
    sum-to-zero, contrast or diagonal add.
 } 
 \usage{
inla.knmodels(
 formula,
 data,
 graph=NULL,
 progress=FALSE, 
 type=c(paste(1:4), paste0(2:4, 'c'), paste0(2:4, 'd')), 
 control.prec,
 ...) 
}
 \usage{
inla.knmodels.sample(
  graph,
  m,
  type=4,
  intercept=0, 
  tau.t=1,
  phi.t=0.7,
  tau.s=1,
  phi.s=0.7,
  tau.st=1,
  ev.t=NULL,
  ev.s=NULL)
}
 \arguments{
 \item{formula}{The formula specifying the other 
   model components, without the spacetime 
   interaction term. The spacetime interaction term 
   will be added accordly to the specification in 
   the \code{type} argument. See \code{inla}}
 \item{data}{A data frame or list containing all 
   the variables in the model. It has to contain 
   an index named \code{st} for the spacetime 
   interaction term. The data MUST be provided.}
 \item{graph}{The graph for the spatial neighbor 
   structure to be used in a \code{\link{f}} term 
   for the main spatial random effect term or for 
   building the spacetime interaction model.}
 \item{progress}{If it is to be shown the model 
   fitting progress. Useful if more than one 
   interaction type is being fitted.}
 \item{type}{The spacetime interaction type.  
   \code{0} to \code{4} corresponds to the four 
   interaction types in Knorr-Held, L. (2000) with 
   all the needed sum-to-zero constraints. 
   \code{2c}, \code{3c} and \code{4c} are 
   the contrast version considering the first time  
   or space constrained to be equal to zero. 
   \code{2d}, \code{3d} and \code{4d} are the 
   corresponding versions when considering the 
   diagonal add approach.}  
 \item{diagonal}{The value to be added to the 
   diagonal when using the diagonal add approach.}
 \item{control.st}{Specification of the hyperparameter, 
   fixed or random, initial value, prior and its 
   parameters for the spacetime interaction. See 
   \code{?inla.models} and look for \code{generic0}. 
   By default we scale it and use the PC-prior to set 
   the prior using the \code{pc.prec} prior with 
   \code{param = c(0.5, 0.5)}. See documentation with 
   \code{?inla.doc("pc.prec")}}
 \item{...}{Arguments to be passed to the 
   \code{\link{inla}} function.}
 }
 \value{
  \code{inla.knmodels} returns an object of class \code{"inla"}. 
    or a list of objects of this class if it is asked to compute 
    more than one interaction type at once. 
 Note: when the model type is 2c, 3c, 4c, 2d, 3d or 4d, it also 
   includes linear combinations summary.
 }
 \author{Elias T. Krainski}
 \seealso{
     \code{\link{inla}}
 }
 \examples{
### space
n2 <- n1 <- 5
grid <- SpatialGrid(GridTopology(c(0,0), c(1, 1), c(n1, n2)))
(n <- nrow(xy <- coordinates(grid)))

if (FALSE)
    plot(grid)
  
jj <- lapply(1:n, function(i) 
    which(sqrt((xy[i,1]-xy[,1])^2 + (xy[i,2]-xy[,2])^2)==1))

### the spatial adjacency matrix
graph <- sparseMatrix(rep(1:n, sapply(jj, length)),
                      unlist(jj), x=1, dims=c(n, n))

### time
m <- 10 

### some random data
dat <- inla.knmodels.sample(graph=graph, m=m, type=4, tau.t=2, tau.s=2, tau.st=3)
str(dat)
sapply(dat$x, summary)

nd <- length(dat$x$eta)
dat$e <- runif(nd, 0.9, 1.1)*rgamma(n, 40, 2)
dat$y <- rpois(nd, dat$e*exp(dat$x$eta-3))
summary(dat$y)

### fit the models
res <- inla.knmodels(y~1, dat, graph=graph,
                     type=c(4, '4c', '4d'), control.st=list(), progress=TRUE,
                     control.compute=list(dic=TRUE, waic=TRUE, cpo=TRUE))
tab <- sapply(res, function(x) c(dic=x$dic$dic, waic=x$waic$waic, cpo=-sum(log(x$cpo$cpo))))
round(tab-apply(tab, 1, min), 1)
}
 \arguments{
 \item{graph}{}
 \item{m}{Time dimention.}
 \item{type}{Integer from 1 to 4 to identify one
    of the four interaction type.}
 \item{intercept}{A constant to be added to the linear predictor}
 \item{tau.t}{Precision parameter for the main temporal effect.}
 \item{phi.t}{Mixing parameter in the \code{bym2} model 
    assumed for the main temporal effect.}
 \item{tau.s}{Precision parameter for the main spatial effect.}
 \item{phi.s}{Mixing parameter in the \code{bym2} model
    assumed for the main spatial effect.}
 \item{tau.st}{Precision parameter for the spacetime effect.}
 \item{ev.t}{Eigenvalues and eigenvectors of the temporal
    precision matrix structure.}
 \item{ev.s}{Eigenvalues and eigenvectors of the spatial 
    precision matrix structure.}
 }
 \value{
  A list with the following elements
   \item{t}{The time index for each obervation, with length equals m*n.} 
   \item{s}{The spatial index for each obervation, with length equals m*n.} 
   \item{st}{The spacetime index for each obervation, with length equals m*n.} 
   \item{x}{A list with the following elements}
   \item{t.iid}{The unstructured main temporal effect part.}
   \item{t.str}{The structured main temporal effect part.}
   \item{t}{The main temporal effect with length equals 2m.}
   \item{s.iid}{The unstructured main spatial effect part.}
   \item{s.str}{The structured main spatial effect part.}
   \item{s}{The main spatial effect with length equals 2n.}
   \item{st}{The spacetime interaction effect with length equals m*n.}
   \item{eta}{The linear predictor with length equals n*m.}
 }
 \author{Elias T. Krainski}
 \seealso{
     \code{\link{inla.qsample}}
 }
