%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from knmodels.sample.R
%%
 \name{inla.knmodels.sample}
 \alias{inla.knmodels.sample}
 \title{Spacetime interaction models sampler function}
 \description{
    It implements the sampling method for the models 
    in Knorr-Held, L. (2000) considering the algorithm 
    3.1 in Rue & Held (2005) book.
 } 
 \usage{
inla.knmodels.sample(
  graph,
  m,
  type=4,
  intercept=0, 
  tau.t=1,
  phi.t=0.7,
  tau.s=1,
  phi.s=0.7,
  tau.st=1,
  ev.t=NULL,
  ev.s=NULL)
}
 \arguments{
 \item{graph}{}
 \item{m}{Time dimention.}
 \item{type}{Integer from 1 to 4 to identify one
    of the four interaction type.}
 \item{intercept}{A constant to be added to the linear predictor}
 \item{tau.t}{Precision parameter for the main temporal effect.}
 \item{phi.t}{Mixing parameter in the \code{bym2} model 
    assumed for the main temporal effect.}
 \item{tau.s}{Precision parameter for the main spatial effect.}
 \item{phi.s}{Mixing parameter in the \code{bym2} model
    assumed for the main spatial effect.}
 \item{tau.st}{Precision parameter for the spacetime effect.}
 \item{ev.t}{Eigenvalues and eigenvectors of the temporal
    precision matrix structure.}
 \item{ev.s}{Eigenvalues and eigenvectors of the spatial 
    precision matrix structure.}
 }
 \value{
  A list with the following elements
   \item{time}{The time index for each obervation, 
       with length equals m*n.} 
   \item{space}{The spatial index for each obervation, 
       with length equals m*n.} 
   \item{spacetime}{The spacetime index for each obervation, 
       with length equals m*n.} 
   \item{x}{A list with the following elements}
   \item{t.iid}{The unstructured main temporal effect part.}
   \item{t.str}{The structured main temporal effect part.}
   \item{t}{The main temporal effect with length equals 2m.}
   \item{s.iid}{The unstructured main spatial effect part.}
   \item{s.str}{The structured main spatial effect part.}
   \item{s}{The main spatial effect with length equals 2n.}
   \item{st}{The spacetime interaction effect with length equals m*n.}
   \item{eta}{The linear predictor with length equals n*m.}
 }
 \author{Elias T. Krainski}
 \seealso{
     \code{\link{inla.knmodels}} for model fitting
 }
