\name{Leuk}
\alias{Leukemia}
\alias{Leuk}
\docType{data}
\title{The Leukemia data}
\description{
This the Leukemia data from  Henderson et al (2003); see source.
}
\usage{data(Leuk)}
\format{
  A data frame with 1043 observations on the following 9 variables.
  \describe{
    \item{time}{TODO}
    \item{cens}{TODO}
    \item{xcoord}{TODO}
    \item{ycoord}{TODO}
    \item{age}{TODO}
    \item{sex}{TODO}
    \item{wbc}{TODO}
    \item{tpi}{TODO}
    \item{district}{TODO}
  }
}

\source{
This is the dataset from 

Henderson, R. and Shimakura, S. and Gorst, D., 2002, Modeling spatial
variation in leukemia survival data, JASA, 97, 460, 965--972.
}

\examples{
data(Leuk)
}
\keyword{datasets}
