% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{inla.likelihood}
\alias{inla.likelihood}
\alias{likelihood}
\title{Providing functions for sampling new data, evaluating pdf, cdf, and
quantiles for new data.}
\usage{
inla.likelihood(type = c("d", "p", "r", "q", "s"), args)
}
\arguments{
\item{type}{The returned function type. The definition is similar to
"rnorm","dnorm","pnorm",and "dnorm".}

\item{args}{It is usually a return value from "inla.likelihood.parser",
which specifies parameters, link function and transformation function of
hyperparameters.}
}
\value{
value goes here
}
\description{
This function return function to compute the pdf,cdf,quantiles, or samples
for new data using the likelihood from a inla-object.
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
