% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{lines.inla.mesh.segment}
\alias{lines.inla.mesh.segment}
\title{Draw \code{inla.mesh.segment} objects.}
\usage{
\method{lines}{inla.mesh.segment}(
  x,
  loc = NULL,
  col = NULL,
  colors = c("black", "blue", "red", "green"),
  add = TRUE,
  xlim = NULL,
  ylim = NULL,
  rgl = FALSE,
  ...
)
}
\arguments{
\item{x}{An \code{\link[=inla.mesh.segment]{inla.mesh.segment()}} object.}

\item{loc}{Point locations to be used if \code{x$loc} is \code{NULL}.}

\item{col}{Segment color specification.}

\item{colors}{Colors to cycle through if \code{col} is \code{NULL}.}

\item{add}{If \code{TRUE}, add to the current plot, otherwise start a new
plot.}

\item{xlim}{X axis limits for a new plot.}

\item{ylim}{Y axis limits for a new plot.}

\item{rgl}{If \code{TRUE}, use \code{rgl} for plotting.}

\item{\dots}{Additional parameters, passed on to graphics methods.}
}
\description{
Draws a \code{\link[=inla.mesh.segment]{inla.mesh.segment()}} object with generic or \code{rgl}
graphics.
}
\seealso{
\code{\link[=inla.mesh.segment]{inla.mesh.segment()}}
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
