% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-models.R
\name{inla.list.models}
\alias{inla.list.models}
\alias{list.models}
\title{List available model components, likelihoods, priors, etc}
\usage{
inla.list.models(section = names(inla.models()), ...)
}
\arguments{
\item{section}{The section(s) to list, missing \code{section} will list all
sections.  \code{names(inla.models())} lists available sections.}

\item{...}{Additional argument to \code{cat}}
}
\value{
Nothing is returned
}
\description{
List available model components, likelihoods, priors, etc.  To read specific
documentation for the individual elements, use \code{\link{inla.doc}}.

The list is \code{cat}'ed with \code{...} arguments.

This function is EXPERIMENTAL.
}
\examples{

\dontrun{
inla.list.models("likelihood")
inla.list.models(c("prior", "group"))
inla.list.models(file=file("everything.txt"))

#Show detailed doc for a specific prior/likelihood/latent model
inla.doc("binomial")
}

}
\author{
Havard Rue
}
