% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmesher-transition.R
\name{local_inla_options}
\alias{local_inla_options}
\title{Helper for locally setting inla options}
\usage{
local_inla_options(..., .envir = parent.frame(), .save_only = FALSE)
}
\arguments{
\item{\dots}{Named arguments to be passed to \code{\link[=inla.setOption]{inla.setOption()}}.}

\item{.envir}{The environment in which the options should be set. Defaults to
the calling environment.}

\item{.save_only}{If \code{TRUE}, the options are not set, but the current values
are saved and will be restored when the calling function terminates, even if
other code sets them directly by calling \code{\link[=inla.setOption]{inla.setOption()}}. This can be
used to protect specific options against local changes that should not have
global effect. Default is \code{FALSE}, meaning that the options are set.}
}
\value{
logical; If the option setting was successful, \code{TRUE} is returned,
otherwise \code{FALSE}.
}
\description{
Locally sets INLA options with \code{\link[=inla.setOption]{inla.setOption()}} and restores
when the calling function terminates, like \code{\link[withr:with_options]{withr::local_options()}} does for
global R options.
}
\examples{
print(paste0("Before: ", INLA::inla.getOption("num.threads")))
local({
  fun1 <- function() { INLA::inla.setOption("num.threads" = "1:1") }
  fun2 <- function() {
    local_inla_options(num.threads = NULL, .save_only = TRUE)
    print(paste0("Local value: ", INLA::inla.getOption("num.threads")))
    fun1()
    print(paste0("Local changed value: ",
                 INLA::inla.getOption("num.threads")))
    invisible()
  }
  
  fun2()
  print(paste0("Restored: ", INLA::inla.getOption("num.threads")))
  local_inla_options(num.threads = "2:1")
  print(paste0("Change: ", INLA::inla.getOption("num.threads")))
  fun2()
  print(paste0("Restored: ",
               INLA::inla.getOption("num.threads")))
})
print(paste0("After: ", INLA::inla.getOption("num.threads")))
}
