% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshassessment.R
\name{inla.mesh.assessment}
\alias{inla.mesh.assessment}
\title{Interactive mesh building and diagnostics}
\usage{
inla.mesh.assessment(mesh, spatial.range, alpha = 2, dims = c(500, 500))
}
\arguments{
\item{mesh}{An \code{inla.mesh}}

\item{spatial.range}{numeric; the spatial range parameter to use for the assessment}

\item{alpha}{numeric; A valid \code{inla.spde2.pcmatern} \code{alpha} parameter}

\item{dims}{2-numeric; the grid size}
}
\description{
Assess the finite element approximation errors in a mesh for interactive R
sessions.  More detailed assessment tools are in \code{\link[=meshbuilder]{meshbuilder()}}.
}
\examples{

library(fmesher)
bnd <- fm_segm(cbind(
    c(0, 10, 10, 0, 0),
    c(0, 0, 10, 10, 0)
), is.bnd = TRUE)
mesh <- fm_mesh_2d_inla(boundary = bnd, max.edge = 1)
out <- inla.mesh.assessment(mesh, spatial.range = 3, alpha = 2)

}
\seealso{
\code{\link[fmesher:fm_mesh_2d]{fmesher::fm_mesh_2d()}}, \code{\link[fmesher:fm_rcdt_2d]{fmesher::fm_rcdt_2d()}}, meshbuilder
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
