% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshbuilder.R
\name{meshbuilder}
\alias{meshbuilder}
\title{Interactive mesh building and diagnostics}
\usage{
meshbuilder()
}
\description{
Interactively design and build a triangle mesh
for use with SPDE models, and assess the finite element
approximation errors. The R code needed to recreate the mesh
outside the interactive Shiny app is also generated. Spatial
objects can be imported from the global workspace.
}
\examples{
\dontrun{
meshbuilder()
}
}
\seealso{
\code{\link[=inla.mesh.2d]{inla.mesh.2d()}}, \code{\link[=inla.mesh.create]{inla.mesh.create()}}
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
