% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{Munich}
\alias{Munich}
\title{The Munich rent data}
\format{
A data frame with 2035 observations on the following 17 variables.
\describe{

\item{rent}{Net rent per square meter.}
\item{floor.size}{Size of the flat in square meters.}
\item{year}{Year of construction of the building in which the flat
is located.} \item{location}{Location index (in terms of
subquarters).} \item{Gute.Wohnlage}{Dummy variable for good
locations / good neighborhoods.} \item{Beste.Wohnlage}{Dummy
variable for very good locations / very good neighborhoods.}
\item{Keine.Wwv}{Dummy for absence of warm water supply.}
\item{Keine.Zh}{Dummy for absence of central heating system.}
\item{Kein.Badkach}{Dummy for absence of flagging in the bathroom.}
\item{Besond.Bad}{Dummy for special features of the bathroom.}
\item{Gehobene.Kueche}{Dummy for more refined kitchen equipment.}
\item{zim1}{Dummy for a flat with 1 room.} \item{zim2}{Dummy
for a flat with 2 rooms.} \item{zim3}{Dummy for a flat with 3
rooms.} \item{zim4}{Dummy for a flat with 4 rooms.}
\item{zim5}{Dummy for a flat with 5 rooms.}
\item{zim6}{Dummy for a flat with 6 rooms.}

}
}
\source{
See Rue and Held (2005), Chapter 4.
}
\description{
The Munich rent data
}
\references{
Rue, H and Held, L. (2005) \emph{Gaussian Markov Random Fields -
Theory and Applications} Chapman and Hall
}
\keyword{datasets}
