\name{my.interpret.vector}
\alias{my.interpret.vector}
\alias{my.read.binary.file}
\alias{my.interpret.vector1}


\title{ Read files }
\description{
  Functions used internally by the \code{inla} function to read binary
  files.

  Normally not used directly.
}


\author{ sara martino \email{martino@math.ntnu.no} }

\keyword{internal }

