\name{nwEngland}
\alias{nwEngland}
\alias{NewEngland}
\docType{data}
\title{The New England map}
\description{
This map is used in association to the Leukemia data from  Henderson et al (2003); see source.
}
\usage{data(Leuk)}
\format{
  A SpatialPolygons object.
}

\source{
This map are used to analyse the Leukaemia dataset from 

Henderson, R. and Shimakura, S. and Gorst, D., 2002, Modeling spatial
variation in leukemia survival data, JASA, 97, 460, 965--972.
}

\examples{
data(Leuk)
plot(nwEngland)
}
\keyword{datasets}
