% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardiso.R
\name{pardiso}
\alias{pardiso}
\alias{inla.pardiso}
\alias{inla.pardiso.check}
\title{Describe and check the PARDISO support in R-INLA}
\usage{
inla.pardiso()

inla.pardiso.check()
}
\description{
\code{inla.pardiso()} describes the \code{PARDISO} support in R-INLA, how to
get the license key and enable it in the \code{R-INLA} package.
\code{inla.pardiso.check()} check if the \code{PARDISO} support is working.
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
