% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-ar.R
\name{pc.ar}
\alias{pc.ar}
\alias{inla.pc.ar.rpacf}
\alias{inla.pc.ar}
\alias{ar.rpacf}
\alias{inla.pc.ar.dpacf}
\alias{ar.dpacf}
\title{Utility functions for the PC prior for a an AR(p) model}
\usage{
inla.pc.ar.rpacf(n = 1, p, lambda = 1)

inla.pc.ar.dpacf(pac, lambda = 1, log = TRUE)
}
\arguments{
\item{n}{Number of observations}

\item{p}{The order of the AR-model}

\item{lambda}{The rate parameter in the prior}

\item{pac}{A vector of partial autocorrelation coefficients}

\item{log}{Logical. Return the density in natural or log-scale.}
}
\value{
\code{inla.pc.ar.rpac} generate samples from the prior, returning
a matrix where each row is a sample of \code{theta}.  \code{inla.pc.ar.dpac}
evaluates the density of \code{pac}.  Use \code{inla.ar.pacf2phi},
\code{inla.ar.phi2pacf}, \code{inla.ar.pacf2acf} and \code{inla.ar.acf2pacf}
to convert between various parameterisations.
}
\description{
Functions to evaluate and sample from the PC prior for an AR(p) model
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
