%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from pc-corrmat.R
%%
 \name{pc.rcormat}
 \alias{inla.pc.rcormat}
 \alias{pc.rcormat}
 \alias{pc.dcormat}
 \alias{inla.pc.dcormat}
 
 \title{Utility functions for the PC prior for a correlation matrix}
 
 \description{Functions to evaluate and sample from the 
              PC prior for a correlation matrix.}
 \usage{
 inla.pc.rcormat(n=1, p, lambda=1, permute = TRUE)
 inla.pc.dcormat(cormat, lambda=1, permute = TRUE, log=FALSE)
 }
 \arguments{
   \item{n}{Number of observations}
   \item{p}{The dimension of the matrix}
   \item{lambda}{The rate parameter}
   \item{permute}{Make the prior exchangable.}
   \item{cormat}{A correlation matrix}
   \item{log}{Logical. Return the density in natural or log-scale.}
 }
 \details{
     The permute-option can have a high cost as the number of terms in the
     sum is \code{factorial(p)}.
 }
\value{%%
  \code{inla.pc.dcormat} gives the density and 
  \code{inla.pc.rcormat} generates random deviates (a
        list of correlation matrices).
 }
 \author{Havard Rue \email{hrue@math.ntnu.no}}
 \examples{
  A = inla.pc.rcormat(n=1, p=3, lambda = 1, permute=FALSE)[[1]]
  print(inla.pc.dcormat(A))
  A = inla.pc.rcormat(n=10, p=3, lambda = 1, permute=TRUE)
  lapply(A, inla.pc.dcormat)
 }
