% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-gev.R
\name{pc.gevtail}
\alias{pc.gevtail}
\alias{inla.pc.gevtail}
\alias{pc.rgevtail}
\alias{inla.pc.rgevtail}
\alias{pc.dgevtail}
\alias{inla.pc.dgevtail}
\alias{pc.pgevtail}
\alias{inla.pc.pgevtail}
\alias{pc.qgevtail}
\alias{inla.pc.qgevtail}
\title{Utility functions for the PC prior for the \code{tail} parameter in the GEV
likelihood}
\usage{
inla.pc.rgevtail(n, lambda = 7)

inla.pc.dgevtail(xi, lambda = 7, log = FALSE)

inla.pc.qgevtail(p, lambda = 7)

inla.pc.pgevtail(q, lambda = 7)
}
\arguments{
\item{n}{Number of observations}

\item{lambda}{The rate parameter in the PC-prior}

\item{xi}{Vector of evaluation points, where \verb{1>xi>0}.}

\item{log}{Logical. Return the density in natural or log-scale.}

\item{p}{Vector of probabilities}

\item{q}{Vector of quantiles}
}
\value{
\code{inla.pc.dgevtail} gives the density,
\code{inla.pc.pgevtail} gives the distribution function,
\code{inla.pc.qgevtail} gives the quantile function, and
\code{inla.pc.rgevtail} generates random deviates.
}
\description{
Functions to evaluate, sample, compute quantiles and percentiles of the PC
prior for the \code{tail} parameter in the GEV likelihood
}
\details{
This gives the PC prior for the \code{tail} parameter for the GEV
likelihood, where \code{xi=0} is the base model.
}
\examples{

 xi = inla.pc.rgevtail(100,  lambda = 7)
 d = inla.pc.dgevtail(xi, lambda = 7)
 xi = inla.pc.qgevtail(0.5, lambda = 7)
 inla.pc.pgevtail(xi, lambda = 7)
 
}
\seealso{
inla.doc("pc.gevtail")
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
