% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-vm0.R
\name{pc.vm0}
\alias{pc.vm0}
\alias{inla.pc.vm0}
\alias{pc.rvm0}
\alias{inla.pc.rvm0}
\alias{pc.dvm0}
\alias{inla.pc.dvm0}
\alias{pc.pvm0}
\alias{inla.pc.pvm0}
\alias{pc.qvm0}
\alias{inla.pc.qvm0}
\title{Utility functions for the PC prior for the precision}
\usage{
inla.pc.rvm0(n, u, alpha, lambda)

inla.pc.dvm0(k, u, alpha, lambda, log = FALSE)

inla.pc.qvm0(p, u, alpha, lambda, len = 2048L)

inla.pc.pvm0(q, u, alpha, lambda, log = FALSE)
}
\arguments{
\item{n}{Number of observations}

\item{u}{The upper limit (0 < u < 2*pi). The small values of u indicate a
high concentration to a point mass, whilst large values of u mean that the
user believes the data spread widely.}

\item{alpha}{The probability going above the upper limit (the probability
assigned to the event Prob(2*pi/(1+k) > u)).}

\item{lambda}{The rate parameter (see Details)}

\item{k}{The concentration of von Mises distribution}

\item{log}{Logical. Return the density in natural or log-scale.}

\item{p}{Vector of probabilities}

\item{q}{Vector of quantiles}
}
\value{
\code{inla.pc.dvm0} gives the density, \code{inla.pc.pvm0}
gives the distribution function, \code{inla.pc.qvm0} gives the quantile
function, and \code{inla.pc.rvm0} generates random deviates.
}
\description{
Functions to evaluate, sample, compute quantiles and percentiles of the PC prior for the
precision in the Gaussian distribution.
}
\details{
The statement \code{Prob(2*pi/(1+k) > u) = alpha} is used to determine
\code{lambda} unless \code{lambda} is given.  Either \code{lambda} must be
given, or \code{u} AND \code{alpha}.

Due to limitations in handling extreme values for special functions, the output of
these functions may exhibit bias when the input parameter values are either excessively
large or very close to zero.
}
\examples{

 k = inla.pc.rvm0(100,  lambda = 1)
 d = inla.pc.dvm0(1, lambda = 1)
 k = inla.pc.qvm0(0.5, u = 1, alpha=0.01)
 inla.pc.pvm0(5, u = 1, alpha=0.01)

}
\seealso{
inla.doc("pc.vm0")
}
\author{
Xiang Ye \email{xiang.ye@kaust.edu.sa}
}
