% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{plot.inla.marginal}
\alias{plot.inla.marginal}
\title{Not yet in use}
\usage{
\method{plot}{inla.marginal}(x, y, ...)
}
\arguments{
\item{x, y, ...}{Not yet in use}
}
\description{
Method not yet in use
}
\keyword{internal}
