%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from posterior.sample.R
%%
 \name{inla.sample}
 \alias{inla.posterior.sample}
 \alias{posterior.sample}
 
 \title{Generate samples from an approximated posterior of a fitted model}
 
 \description{This function generate samples from an approximated posterior of a fitted model,  ie an inla-object}
 \usage{
     inla.posterior.sample(n = 1L, result, hyper.user.scale = TRUE)
 }
 
 \arguments{
   \item{n}{Number of samples.}
   \item{result}{The inla-object, ie the output from an \code{inla}-call}
   \item{hyper.user.scale}{Use values of hyperparameters in the user scale, not in the internal representation}
}
\value{
  A list of samples,  where each samples is a list with names \code{hyperpar} and \code{latent}}.
  THIS IS AN EXPERIMENTAL FUNCTION AND CHANGES MAY APPEAR AT ANY TIME!
}
\author{Havard Rue \email{hrue@math.ntnu.no}}
 
\examples{
}
