% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{PRborder}
\alias{PRborder}
\title{The PRborder data}
\format{
\describe{ \item{Longtiude}{The Longtiude coordinate}
\item{Latitude}{The Latitude coordinate} }
}
\description{
A data matrix with Longitude and Latitude coordinates for the boundary of
Parana State.
}
\seealso{
PRprec
}
\keyword{datasets}
