\name{print.inla}
\alias{print.inla}
\alias{print.surv.inla}


\title{ INLA default print statement }
\description{
  This is the default print statement for a \code{inla} or \code{surv.inla} object. If you need
  a list of everything that is part the object see \code{inla} of \code{surv.inla}, or
  use \code{names()}.
  The integration strategy is printed together with the number of
  hyperparameters and the likelihood model.
  Also the random effect are listed with their specific model.
}

\author{ sara martino \email{martino@math.ntnu.no} }

\keyword{ print }
