% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.inla}
\alias{print.inla}
\title{Print an INLA fit}
\usage{
\method{print}{inla}(x, digits = 3L, ...)
}
\arguments{
\item{x}{An inla-object (output from an \code{\link[=inla]{inla()}}-call).}

\item{digits}{Number of digits to print}

\item{...}{other arguments.}
}
\value{
None
}
\description{
Print an INLA fit
}
\seealso{
\code{\link[=inla]{inla()}}
}
\author{
Havard Rue
}
