%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from prune.R
%%
\name{inla.prune}
\alias{inla.prune}

\title{Prune the INLA-package}

\description{Prune the INLA-package by removing binaries not supported by the running OS}

\usage{
inla.prune()
}
\arguments{
}
\value{No value is returned.}
\author{Havard Rue \email{hrue@r-inla.org}}
