% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc-prw2-range.R
\name{pc.prw2.range}
\alias{pc.prw2.range}
\alias{prw2.d}
\alias{inla.pc.prw2.drange}
\alias{inla.pc.prw2.rrange}
\alias{inla.pc.prw2.qrange}
\alias{inla.pc.prw2.prange}
\title{Utility functions for the PC prior for range in PRW2}
\usage{
prw2.d(rho)
}
\arguments{
\item{n}{Number of observations}

\item{param}{Vector of parameters (see 'inla.doc("pc.prw2.range")')}

\item{q}{Vector of quantiles}

\item{x}{Vector of quantiles}

\item{p}{Vector of probabilities}
}
\value{
\code{inla.prw2.drange} gives the density, \code{inla.prw2.prange} gives the distribution
function, \code{inla.prw2.qrange} gives the quantile function, and \code{inla.prw2.rrange}
generates random deviates.
}
\description{
Functions to evaluate, sample, compute quantiles and percentiles of the PC prior for the
range parameter in the PRW2 model.
}
\details{
The statement \code{Prob(range > r0) = alpha} is used to determine
\code{lambda} unless \code{lambda} is given.
}
\note{
The 'r' and 'q' functions are quite slow as the quantile-function needs to be computed.
}
\examples{

    param <- c(10, 0.5, 1, 0)
    p <- inla.prw2.prange(10, param)
    q <- inla.prw2.qrange(p, param)
    print(c(10, q))
 
}
\seealso{
inla.doc("pc.prw2.range")
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
