%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from q.R
%%
\name{inla.qstat}
\alias{inla.qstat}
\alias{inla.qget}
\alias{inla.qdel}
\alias{inla.qnuke}
\alias{summary.inla.q}
\alias{print.inla.q}
\title{Control and view a remote inla-queue}
\description{
Control and view a remote inla-queue of submitte jobs
}
\usage{
inla.qget(id, remove = TRUE)
inla.qdel(id)
inla.qstat()
inla.qnuke()
\method{summary}{inla.q}(object,...)
\method{print}{inla.q}(x,...)
}
\arguments{
  \item{id}{The job-id which is the output from \code{inla} when the job is submitted,  the
            job-number or job-name}
  \item{remove}{Logical If FALSE, leave the job on the server after getting-it, otherwise remove it.}
  \item{x}{An \code{inla.get}-object,  which is the output from \code{inla.get()}}
  \item{object}{An \code{inla.get}-object,  which is the output from \code{inla.get()}}
  \item{...}{ other arguments.}
}
\details{
\code{inla.qstat} shows jobs on the server,
\code{inla.qget} fetch the results (and by default remove
the files on the server),  \code{inla.qdel} removes 
a job on the server and \code{inla.qnuke} remove all the jobs on the server.
}
\value{
  \code{inla.qstat} returns an \code{inla.q}-object with information about current jobs.
}
\author{Havard Rue}
\seealso{ \code{\link{inla}} }
\examples{
\dontrun{
r = inla(y~1, data = data.frame(y=rnorm(10)), inla.call="submit")
inla.qstat()
r = inla.get(r, remove=FALSE)
inla.qdel(1)
inla.qnuke()
}
}
