%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from qreordering.R
%%
 \name{qreordering}
 \alias{inla.qreordering}
 \alias{qreordering}
 
 \title{Compute the reordering using the GMRFLib implementation}
 
 \description{This function compute the reordering (or find the best reordering)
              using the GMRFLib implementation}
 \usage{
     inla.qreordering(graph, reordering)
 }
 
 \arguments{
   \item{graph}{A (inla-)graph object, a filename containing the graph (or matrix) or a matrix defining it.}
   \item{reordering}{The type of reordering algorithm,  one of
        "auto", "default", "identity", "band", "metis", "genmmd", "amd", "amdbar", "md", "mmd", "amdc" and "amdbarc".
        The default is "auto" which try several and find the best one.}
}
\value{
  \code{inla.qreordering} returns a list with the name of the reordering algorithm used or found, the reordering code for the reordering algorithm,
                          the actual reordering and its inverse.
}
\author{Havard Rue \email{hrue@math.ntnu.no}}
 
\examples{
 g = system.file("demodata/germany.graph", package="INLA")
 r = inla.qreordering(g)
 m = inla.graph2matrix(g)
 r = inla.qreordering(m)
 m.file = inla.write.fmesher.file(m)
 r = inla.qreordering(m.file)
 unlink(m.file)
}
