% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde.common.R
\name{rbind.inla.data.stack.info}
\alias{rbind.inla.data.stack.info}
\title{Internal function for merging raw stack information}
\usage{
\method{rbind}{inla.data.stack.info}(...)
}
\description{
Internal function for merging raw stack information
}
