% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{inla.ssh.copy.id}
\alias{inla.ssh.copy.id}
\alias{ssh.copy.id}
\alias{inla.remote}
\title{Setup remote computing}
\usage{
inla.ssh.copy.id()

inla.remote()
}
\value{
\code{inla.remote} is used once to setup the remote host
information file (definition file) in the users home directory; see the FAQ
entry on this issue for more information.  \code{inla.ssh.copy.id} will
return the path to the internal script to transfer ssh-keys.
}
\description{
Initialize the definition file and print the path to the internal script to
transfer ssh-keys
}
\examples{

##See the FAQ entry on this issue on r-inla.org.

}
\author{
Havard Rue \email{hrue@r-inla.org}
}
