% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderings.R
\name{inla.reorderings}
\alias{inla.reorderings}
\alias{reorderings}
\title{Reorderings methods for sparse matrices}
\usage{
inla.reorderings()
}
\value{
The names of all available reorderings
}
\description{
Provide the names of all implemented reordering schemes
}
\examples{

inla.reorderings()

}
\author{
Havard Rue \email{hrue@r-inla.org}
}
