% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprior.R
\name{rprior}
\alias{rprior}
\alias{inla.rprior.define}
\alias{rprior.define}
\title{Define prior in R}
\usage{
inla.rprior.define(rprior = NULL, ...)
}
\arguments{
\item{rprior}{An R-function returning the log-prior evaluated at its argument}

\item{...}{Named list of variables that will be in the environment of \code{rprior}}
}
\value{
An \code{inla.rprior}-object which can be used as a prior
}
\description{
A(n experimental) framework for defining a prior in R
}
\examples{

 ## see example in inla.doc("rprior")

}
\author{
Havard Rue \email{hrue@r-inla.org}
}
