% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{Salm}
\alias{Salm}
\title{Extra-Poisson variation in dose-response study}
\format{
A data frame with 18 observations on the following 3 variables.
\describe{ \item{y}{number of salmonella bacteria}
\item{dose}{dose of quinoline (mg per plate)}
\item{rand}{indicator} }
}
\source{
WinBUGS/OpenBUGS manual Examples VOl.I
}
\description{
Breslow (1984) analyses some mutagenicity assay data (shown below) on
salmonella in which three plates have been processed at each dose i of
quinoline and the number of revertant colonies of TA98 Salmonella measured
}
\examples{

data(Salm)
}
\keyword{datasets}
