%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from sampler.R
%%
\name{inla.hyperpar.sample}
\alias{inla.hyperpar.sample}
\alias{inla.hyperpar.sampler}
\alias{hyperpar.sample}
\alias{hyperpar.sampler}

\title{Produce samples from the approximated joint posterior for the hyperparameters}

\description{Produce samples from the approximated joint posterior for the hyperparameters}
\usage{
inla.hyperpar.sample(n, result, intern=FALSE)
}

\arguments{
  \item{n}{Integer. Number of samples required.}
  \item{result}{An \code{inla}-object,  f.ex the output from an \code{inla}-call.}
  \item{intern}{Logical. If \code{TRUE} then produce samples in the
  intern scale for the hyperparmater, if \code{FALSE} then produce
  samples in the user-scale. (For example log-precision (intern)
  and precision (user-scale))}
}

\value{%%
 A matrix where each sample is a row. The contents of the column
 is described in the rownames.
}
%%

\author{Havard Rue \email{hrue@math.ntnu.no}}

\examples{
r = inla(y ~ 1, data = data.frame(y=1:10), family = "t")
x = inla.hyperpar.sample(10,r)
str(x)
}
