% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopy.define.R
\name{scopy.define}
\alias{scopy.define}
\alias{inla.scopy.define}
\title{Extract the parameterisation of the \code{scopy} model}
\usage{
inla.scopy.define(n = 5L)
}
\arguments{
\item{n}{Number of parameters, either 2 (intercept and slope)
or >= 5 (intercept, slope and a spline representing the deviation from it)}
}
\value{
A list with the number of parameters and matrix defining basis vectors for the scopy model
}
\description{
This function provide access to the parameterisation of the \code{scopy} model
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
