% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopy.summary.R
\name{scopy.summary}
\alias{scopy.summary}
\alias{inla.scopy.summary}
\title{Computes the mean and stdev for the spline from \code{scopy}}
\usage{
inla.scopy.summary(
  result,
  name,
  mean.value = NULL,
  slope.value = NULL,
  by = 0.01,
  range = c(0, 1),
  debug = FALSE
)
}
\arguments{
\item{result}{An \code{inla}-object, ie the output from an \code{inla()}
call}

\item{name}{The name of the \code{scopy} model component see \code{?INLA::f}
and argument \code{scopy}}

\item{mean.value}{In case where the mean of the spline is fixed
and not estimated, you have to give it here}

\item{slope.value}{In case where the slope of the spline is fixed
and not estimated, you have to give it here}

\item{by}{The resolution of the results, in the scale where
\code{diff(range(locations))} is 1}

\item{range}{The range of the locations, as \code{c(from, to)}}

\item{debug}{If \code{TRUE} then enable some debug output}
}
\value{
A \code{data.frame} with locations, mean and stdev.  If \code{name}
is not found, NULL is returned.
}
\description{
This function computes the mean and stdev for the spline function that is
implicite from an \code{scopy} model component
}
\examples{

 ## see example in inla.doc("scopy")
 
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
