% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{Scotland}
\alias{Scotland}
\title{Conditional Autoregressive (CAR) model for disease mapping}
\format{
A data frame with 56 observations on the following 4 variables.
\describe{ \item{Counts}{The number of lip cancer registered}
\item{E}{The expected number of lip cancer } \item{X}{The
percentage of the population engaged in agricolture, fishing or forestry }
\item{Region}{The county} }
}
\source{
OpenBUGS Example manual, GeoBUGS
}
\description{
The rate of lip cancer in 56 counties in Scotland is recorder. The data set
includes the observed and expected cases (based on the population and its
age and sex distribution in the country), a covariate measuring the
percentage of the population engaged in agricolture, fishing or forestry and
the "position" of each county expressed as a list of adjacent counties
}
\examples{

data(Scotland)
}
\references{
Clayton and Kaldor (1987) and Breslow and Clayton (1993)
}
\keyword{datasets}
