% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{Seeds}
\alias{Seeds}
\title{Factorial design}
\format{
A data frame with 21 observations on the following 5 variables.
\describe{ \item{r}{number of germinated seeds per plate}
\item{n}{number of total seeds per plate} \item{x1}{seed
type} \item{x2}{root extracted} \item{plate}{indicator for
the plate} }
}
\source{
WinBUGS/OpenBUGS Manual Example, Vol. I
}
\description{
Proportion of seeds that germinated on each of 21 plates arranged according
to a 2 by 2 factorial layout by seed and type of root extract
}
\examples{

data(Seeds)
}
\keyword{datasets}
