\name{set.control.inla.default}
\alias{set.control.inla.default}
\alias{set.control.compute.default}
\alias{set.control.fixed.default}
\alias{set.control.predictor.default}
\alias{set.control.data.default}
\alias{set.control.results.default}

\title{ Set default values }
\description{
  Functions used internally by the \code{inla} function to set default values.

  Normally not used directly.
}

\author{ sara martino \email{martino@math.ntnu.no} }
\keyword{internal}
