%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from show.hyperspec.R
%%
\name{inla.show.hyperspec}
\alias{inla.show.hyperspec}
\alias{show.hyperspec}

\title{List all priors for the hyperparameters used}

\description{List all priors for the hyperparameters used in a an \code{inla}-object}

\usage{
inla.show.hyperspec(result)
}
\arguments{
 \item{result}{An \code {inla} -object} 
}

\value{%%
 Nothing. The used priors for the hyperparameters are displayed only.
}
\author{Havard Rue \email{hrue@math.ntnu.no}}
\examples{
r = inla(y~1, data = data.frame(y=1:10))
inla.show.hyperspec(r)
}
