% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{SPDEtoy}
\alias{SPDEtoy}
\title{toy simulated data set for the SPDE tutorial}
\format{
A data frame with 200 observations on the following 3 variables.
\describe{ \item{s1}{First element of the coordinates}
\item{s2}{Second element of the coordinates} \item{y}{data
simulated at the locations} }
}
\source{
SPDE tutorial
}
\description{
Simulated data set on 200 location points.  The simulation process is made
at the introduction of the SPDE tutorial.
}
\examples{

data(SPDEtoy)
}
\keyword{datasets}
