% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stiles.R
\name{stiles}
\alias{stiles}
\alias{inla.stiles}
\title{Add sTiles support in R-INLA}
\usage{
inla.stiles()
}
\description{
\code{inla.stiles()} opens the web-page with
description how to enable \code{stiles} support in R-INLA
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
