% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.R
\name{summary.inla.marginal}
\alias{summary.inla.marginal}
\title{Not yet in use}
\usage{
\method{summary}{inla.marginal}(object, ...)
}
\arguments{
\item{object, ...}{Not yet in use}
}
\description{
Method not yet in use
}
\keyword{internal}
