% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{summary.inla.mesh}
\alias{summary.inla.mesh}
\alias{print.summary.inla.mesh}
\title{Summarizing triangular mesh objects}
\usage{
\method{summary}{inla.mesh}(object, verbose = FALSE, ...)

\method{print}{summary.inla.mesh}(x, ...)
}
\arguments{
\item{object}{an object of class \code{"inla.mesh"}, usually a result of a
call to \code{\link[=inla.mesh.create]{inla.mesh.create()}} or \code{\link[=inla.mesh.2d]{inla.mesh.2d()}}.}

\item{verbose}{If \code{TRUE}, produce a more detailed output.}

\item{\dots}{further arguments passed to or from other methods.}

\item{x}{an object of class \code{"summary.inla.mesh"}, usually a result of
a call to \code{\link[=summary.inla.mesh]{summary.inla.mesh()}}.}
}
\description{
Construct and print \code{inla.mesh} object summaries
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
