% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.inla}
\alias{summary.inla}
\alias{summary.surv.inla}
\alias{print.summary.inla}
\title{Summary for a INLA fit}
\usage{
\method{summary}{inla}(object, digits = 3L, include.lincomb = TRUE, ...)

\method{print}{summary.inla}(x, digits = 3L, ...)
}
\arguments{
\item{object}{a fitted \code{inla} object as produced by \code{inla}.}

\item{digits}{Integer Number of digits}

\item{include.lincomb}{Logcial Include the summary for the the linear
combinations or not}

\item{...}{other arguments.}

\item{x}{a \code{summary.inla} object produced by \code{summary.inla}}
}
\value{
\code{summary.inla} returns an object of class \code{summary.inla},
a list of components to print.
}
\description{
Takes a fitted \code{inla} or \code{surv.inla} object produced by
\code{inla} or \code{surv.inla} and produces a summary from it.
}
\details{
Posterior mean and standard deviation (together with quantiles or cdf) are
printed for the fixed effects in the model.

For the random effects the function \code{summary()} prints the posterior
mean and standard deviations for the hyperparameters

If the option \code{short.summary} is set to \code{TRUE} using
\code{inla.setOption}, then a less verbose summary variant will be used,
which might be more suitable for Markdown documents.
}
\seealso{
\code{\link{inla}}
}
\author{
Sara Martino and Havard Rue
}
