% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{Surg}
\alias{Surg}
\title{Surgical: Institutional ranking}
\format{
A data frame with 12 observations on the following 3 variables.
\describe{ \item{n}{Number of deaths} \item{r}{Total number
of cases} \item{hospital}{a factor with levels \code{A} \code{B}
\code{C} \code{D} \code{E} \code{F} \code{G} \code{H} \code{I} \code{J}
\code{K} \code{L}} }
}
\source{
WinBUGS/OpenBUGS Manual Examples Vol. I
}
\description{
This example considers mortality rates in 12 hospitals performing cardiac
surgery in babies
}
\examples{

data(Surg)
}
\keyword{datasets}
