% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{SurvSim}
\alias{SurvSim}
\title{Survival data}
\format{
A data frame with 100 observations on the following 3 variables.
\describe{ \item{y}{a numeric vector of survival times}
\item{cens}{a numeric vector of event indicator (0=censured
1=failure)} \item{x}{a numeric vector of covariate } }
}
\description{
Simulated data set for Weibull survival model
}
\keyword{datasets}
