% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{Tokyo}
\alias{Tokyo}
\title{Binomial time series}
\format{
A data frame with 366 observations on the following 3 variables.
\describe{
\item{y}{number of days with rain}
\item{n}{total number of days}
\item{time}{day of the year}
}
}
\source{
\url{http://www.math.ntnu.no/~hrue/GMRF-book/tokyo.rainfall.data.dat}
}
\description{
Recorded days of rain above 1 mm in Tokyo for 2 years, 1983:84
}
\examples{

data(Tokyo)
}
\references{
Rue, H and Held, L. (2005) \emph{Gaussian Markov Random Fields -
Theory and Applications} Chapman and Hall
}
\keyword{datasets}
