% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{inla.update}
\alias{inla.update}
\alias{inla.upgrade}
\title{Upgrade the INLA-package}
\usage{
inla.update(...)

inla.upgrade(...)
}
\arguments{
\item{...}{Arguments not used}
}
\value{
\code{inla.upgrade} returns nothing
}
\description{
Function to upgrade the \code{INLA}-package to the most recent version
}
\seealso{
\code{update.packages}
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
