% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmesh.R
\name{inla.as.wkt_tree.wkt}
\alias{inla.as.wkt_tree.wkt}
\alias{inla.as.wkt.wkt_tree}
\alias{inla.wkt_tree_get_item}
\alias{inla.wkt_tree_set_item}
\title{Internal WKT handling}
\usage{
inla.as.wkt_tree.wkt(x, ...)

inla.as.wkt.wkt_tree(x, pretty = FALSE, ...)

inla.wkt_tree_get_item(x, item, duplicate = 1)

inla.wkt_tree_set_item(x, item_tree)
}
\arguments{
\item{x}{A WKT2 string, or a \code{wkt_tree} list structure}

\item{\dots}{Unused}

\item{item}{An item label identifying a parameter item entry}

\item{duplicate}{For items that have more than one version, \code{duplicate}
indicates the index number of the desired version. Default: 1}
}
\description{
Conversion between WKT and a tree representation
}
