% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{Zambia}
\alias{Zambia}
\title{Semiparametric regression}
\format{
A data frame with 4847 observations on the following 10 variables.
\describe{ \item{hazstd}{standardised Z score of stunting}
\item{bmi}{body mass index of the mother} \item{agc}{age of
the child in months} \item{district}{district where the child lives}
\item{rcw}{mother employment status with categories "working" (1)
and "not working" (-1)} \item{edu1}{mother's educations status with
categories "complete primary but incomplete secondary " (\verb{edu1=1)},
"complete secondary or higher" (\code{edu2=1}) and "no education or
incomplete primary" (\code{edu1=edu2=-1})} \item{edu2}{see above}
\item{tpr}{locality of the domicile with categories "urban" (1) and
"rural" (-1)} \item{sex}{gender of the child with categories "male"
(1) and "female" (-1)} \item{edu}{DO NOT KNOW; check source} }
}
\source{
BayesX Manual
\url{http://www.stat.uni-muenchen.de/~bayesx/bayesx.html}
}
\description{
Undernutrition of children in each region of Zambia is measured through a
score computed on the basis of some anthropometric measures. The data set
contains also other infomation about each child.
}
\examples{

data(Zambia)
}
\keyword{datasets}
