% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmesh.R
\name{as.inla.mesh.segment}
\alias{as.inla.mesh.segment}
\alias{inla.sp2segment}
\title{Convert \code{sp} objects to \code{inla.mesh.segment} objects.}
\usage{
as.inla.mesh.segment(sp, ...)

inla.sp2segment(sp, ...)
}
\arguments{
\item{sp}{An \code{sp} polygon object of class \code{Polygon},
\code{Polygons}, \code{SpatialPolygons}, or \code{SpatialPolygonsDataFrame}.}

\item{\dots}{Additional arguments passed on to other methods.}

\item{join}{If \code{TRUE}, join multiple polygons into a single segment
(possibly non-simply connected).}

\item{grp}{Group ID specification for each polygon, as used by
\code{\link[=inla.mesh.segment]{inla.mesh.segment()}}, one ID per polygon.}

\item{reverse}{Logical, indicating if the line sequence should be traversed
backwards.}

\item{crs}{An optional \code{CRS} or \code{inla.CRS} object}
}
\value{
A \code{\link[=inla.mesh.segment]{inla.mesh.segment()}} object, or a list of
\code{\link[=inla.mesh.segment]{inla.mesh.segment()}} objects.
}
\description{
Wrapper for \code{inlabru::fm_as_inla_mesh_segment}
}
\seealso{
\code{\link[=inla.mesh.segment]{inla.mesh.segment()}}
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
