% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmesh.R
\name{as.inla.mesh.segment}
\alias{as.inla.mesh.segment}
\alias{inla.sp2segment}
\title{Convert \code{sp} objects to \code{inla.mesh.segment} objects.}
\usage{
as.inla.mesh.segment(sp, ...)

inla.sp2segment(sp, ...)
}
\arguments{
\item{sp}{An \code{sp} polygon object of class \code{Polygon},
\code{Polygons}, \code{SpatialPolygons}, or \code{SpatialPolygonsDataFrame}.}

\item{\dots}{Additional arguments passed on to \code{inlabru::fm_as_inla_mesh_segment()}.}
}
\value{
A \code{\link[=inla.mesh.segment]{inla.mesh.segment()}} object, or a list of
\code{\link[=inla.mesh.segment]{inla.mesh.segment()}} objects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} by \code{\link[inlabru:inlabru-deprecated]{inlabru::fm_as_inla_mesh_segment()}}
}
\section{Functions}{
\itemize{
\item \code{inla.sp2segment()}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} by \code{\link[inlabru:inlabru-deprecated]{inlabru::fm_as_inla_mesh_segment()}}

}}
\seealso{
\code{\link[=inla.mesh.segment]{inla.mesh.segment()}}
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
