%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from binary.install.R
%%
 \name{inla.binary.install}
 \alias{inla.binary.install}
 \alias{binary.install}
 \title{Install alternative binary builds}
 \description{
    Install alternative binary builds.
 }
 \usage{
    inla.binary.install(os = c("CentOS Linux-6", "CentOS Linux-7", "CentOS Linux-8",
                               "CentOS Stream-8", "Fedora-33", "Manjaro Linux",
                               "Ubuntu-16.04", "Ubuntu-18.04", "Ubuntu-20.04"), 
                        verbose = TRUE)
 }
 \arguments{
    \item{os}{If \code{os} is given, install binary build for this \code{os}.
              If \code{os} is not given, chose \code{os} interactively
              among available builds.}
    \item{verbose}{Logical. Verbose output if \code{TRUE}}
 }
 \details{
 Install a new binary for \code{os} unless
 \code{missing(os)}, for which the \code{os} is chosen
 interactively among the available builds.
 }
 \value{
 No value returned.
 }
 \author{Havard Rue \email{hrue@r-inla.org}}
 \examples{
   \dontrun{
     inla.binary.install()
     inla.binary.install(os = "CentOS Linux-7")
   }
 }
