%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from binary.install.R
%%
 \name{inla.binary.install}
 \alias{inla.binary.install}
 \alias{binary.install}
 \title{Install alternative binary builds}
 \description{
    Install alternative binary builds.
 }
 \usage{
    inla.binary.install(os = c("CentOS Linux-6", "CentOS Linux-7", "CentOS Linux-8",
                               "CentOS Stream-8",
                               "Fedora-33", "Fedora-34", "Fedora Linux-35",
                               "Fedora Linux-36", Fedora Linux-37",
                               "Manjaro Linux",
                               "Ubuntu-16.04", "Ubuntu-18.04", "Ubuntu-20.04", "Ubuntu-22.04"), 
                        path = NULL, verbose = TRUE, md5.check = TRUE)
 }
 \arguments{
    \item{os}{If \code{os} is given, install binary build for this \code{os}.
              If \code{os} is not given, chose \code{os} interactively
              among available builds.}
    \item{path}{character. The install path. If \code{NULL} the path is derived from \code{INLA} package}
    \item{verbose}{Logical. Verbose output if \code{TRUE}}
    \item{md5.check}{Logical. If \code{TRUE}, stop if md5-checksum-file is not present
                     or md5-checksum fail. If \code{FALSE}, ignore md5-checksum check.}
 }
 \details{
 Install a new binary for \code{os} unless
 \code{missing(os)}, for which the \code{os} is chosen
 interactively among the available builds.
 }
 \value{
 Return \code{TRUE} if installation was sucessful and \code{FALSE} if not.
 }
 \author{Havard Rue \email{hrue@r-inla.org}}
 \examples{
   \dontrun{
     inla.binary.install()
     inla.binary.install(os = "CentOS Linux-7")
     inla.binary.install(os = "CentOS Linux-7",  path = "~/local/bin/inla.binary")
   }
 }
