%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from binary.install.R
%%
 \name{inla.binary.install}
 \alias{inla.binary.install}
 \alias{binary.install}
 \title{Install alternative binary builds}
 \description{
    Install alternative binary builds.
 }
 \usage{
    inla.binary.install(debug = TRUE)
 } 
 \arguments{
    \item{debug}{Logical. Turn on debugging messages if \code{TRUE}}
 }
 \details{
 \code{inla.binary.install()} will offer a menu of alternative 
 (Linux) binary builds to be installed. Currently offered,  are builds
 for Ubuntu1804, Ubuntu1604,  CentOS6, CentOS7, and CentOS8.
 }
 \value{
 No value returned.
 }
 \author{Havard Rue \email{hrue@r-inla.org}}
 \examples{
   \dontrun{
     inla.binary.install()
   }
 }
