% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgeneric.R
\name{cgeneric}
\alias{cgeneric}
\alias{inla.cgeneric.define}
\alias{cgeneric.define}
\alias{cgeneric.q}
\alias{inla.cgeneric.q}
\title{cgeneric models}
\usage{
inla.cgeneric.define(model = NULL, shlib = NULL, n = 0L, debug = FALSE, ...)

inla.cgeneric.q(cmodel = NULL)
}
\arguments{
\item{model}{The name of the model function}

\item{shlib}{Name of the compiled object-file with \code{model}}

\item{n}{The size of the model}

\item{debug}{Logical. Turn on/off debugging}

\item{...}{Additional arguments, required by \code{inla.cgeneric.define()} to be named arguments}

\item{cmodel}{The name of a cgeneric model-object (output from
\code{inla.cgeneric.define}}
}
\description{
A framework for defining latent models in C
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
