%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from compare.results.R
%%
\name{inla.compare.results}
\alias{inla.compare.results}
\alias{compare.results}

\title{Compare INLA and MCMC results}

\description{A small utility to compare INLA and MCMC results (OBSOLETE)}
\usage{
 inla.compare.results(dir.inla = NULL, dir.mcmc = NULL)
}

\arguments{
  \item{dir.inla}{The directory with the INLA results}
  \item{dir.mcmc}{The directory with the MCMC results}
}

\value{%%
  Return nothing. This is an interactive function. 

  This function is OBSOLETE
}
%%
\author{Havard Rue \email{hrue@r-inla.org}}
\examples{
## See demo("Tokyo-compare")
}
