% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.bgev}
\alias{control.bgev}
\alias{inla.set.control.bgev.default}
\title{control.bgev}
\usage{
control.bgev(
  q.location = 0.5,
  q.spread = 0.25,
  q.mix = c(0.1, 0.2),
  beta.ab = 5L
)

inla.set.control.bgev.default(...)
}
\arguments{
\item{q.location}{The quantile level for the location parameter}

\item{q.spread}{The quantile level for the spread parameter (must be < 0.5)}

\item{q.mix}{The lower and upper quantile level for the mixing function}

\item{beta.ab}{The parameters a and b in the Beta mixing function}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\details{
The \code{control.bgev}-list is set within the corresponding \code{control.family}-list as control parameters to the \code{family="bgev"}
}
\seealso{
Other control: 
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.family}()},
\code{\link{control.fixed}()},
\code{\link{control.gcpo}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.mode}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.scopy}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
