% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.compute}
\alias{control.compute}
\alias{inla.set.control.compute.default}
\title{control.compute}
\usage{
control.compute(
  openmp.strategy = "default",
  hyperpar = TRUE,
  return.marginals = TRUE,
  return.marginals.predictor = FALSE,
  dic = FALSE,
  mlik = TRUE,
  cpo = FALSE,
  po = FALSE,
  waic = FALSE,
  residuals = FALSE,
  q = FALSE,
  config = FALSE,
  smtp = NULL,
  graph = FALSE,
  internal.opt = NULL,
  save.memory = NULL,
  control.gcpo = INLA::control.gcpo()
)

inla.set.control.compute.default(...)
}
\arguments{
\item{openmp.strategy}{The computational strategy to use: 'small', 'medium',
'large', 'huge' and 'default'.}

\item{hyperpar}{A boolean variable if the marginal for the hyperparameters
should be computed. Default TRUE.}

\item{return.marginals}{A boolean variable if the marginals for the latent
field should be returned (although it is computed). Default TRUE}

\item{return.marginals.predictor}{A boolean variable if the marginals for
the linear predictor should be returned (although it is computed).
Default FALSE}

\item{dic}{A boolean variable if the DIC-value should be computed. Default
FALSE.}

\item{mlik}{A boolean variable if the marginal likelihood should be computed.
Default \code{TRUE}.}

\item{cpo}{A boolean variable if the cross-validated predictive measures
(cpo, pit) should be computed (default \code{FALSE})}

\item{po}{A boolean variable if the predictive ordinate should be computed
(default \code{FALSE})}

\item{waic}{A boolean variable if the Watanabe-Akaike information criteria
should be computed (default \code{FALSE})}

\item{residuals}{Provide estimates of \code{residuals} (whatever we mean by that).
(default \code{FALSE})
Currently only residuals base on expected (saturated) deviance are available.
The sign of the residuals are only \verb{very likely} correct.
These residuals are not properly justified from a Bayesian point of
view, hence must be used with caution. It is provided in the hope
they would be useful. This feature is EXPERIMENTAL for the moment,
so changes can happen at any time.}

\item{q}{A boolean variable if binary images of the precision matrix, the
reordered precision matrix and the Cholesky triangle should be generated.
(Default FALSE.)}

\item{config}{A boolean variable if the internal GMRF approximations be
stored. (Default \code{FALSE}.)}

\item{smtp}{The sparse-matrix solver, one of 'default', 'taucs', 'band' or
(default \code{inla.getOption("smtp")}).}

\item{graph}{A boolean variable if the graph itself should be returned.
(Default FALSE.)}

\item{internal.opt}{A boolean variable, if to do internal online
optimisations or not. (Default \code{inla.getOption("internal.opt")})}

\item{save.memory}{A boolean variable, make choices which
saves memory over accuracy. (Default 'inla.getOption("save.memory")')}

\item{control.gcpo}{(For experts only!) Set control variables for the gcpo.
The intended use is to use \code{inla.group.cv}.
Refer to \link{control.gcpo}, \code{?inla.group.cv} and the vignette for details.}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\seealso{
Other control:
\code{\link[=control.bgev]{control.bgev()}},
\code{\link[=control.expert]{control.expert()}},
\code{\link[=control.family]{control.family()}},
\code{\link[=control.fixed]{control.fixed()}},
\code{\link[=control.gcpo]{control.gcpo()}},
\code{\link[=control.group]{control.group()}},
\code{\link[=control.hazard]{control.hazard()}},
\code{\link[=control.inla]{control.inla()}},
\code{\link[=control.lincomb]{control.lincomb()}},
\code{\link[=control.link]{control.link()}},
\code{\link[=control.lp.scale]{control.lp.scale()}},
\code{\link[=control.mix]{control.mix()}},
\code{\link[=control.mode]{control.mode()}},
\code{\link[=control.numa]{control.numa()}},
\code{\link[=control.pom]{control.pom()}},
\code{\link[=control.predictor]{control.predictor()}},
\code{\link[=control.scopy]{control.scopy()}},
\code{\link[=control.sem]{control.sem()}},
\code{\link[=control.stiles]{control.stiles()}},
\code{\link[=control.taucs]{control.taucs()}},
\code{\link[=control.update]{control.update()}},
\code{\link[=control.vb]{control.vb()}}
}
\concept{control}
