%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.compute}
\alias{control.compute}
\alias{inla.control.compute}
\alias{inla.set.control.compute.default}
\alias{set.control.compute.default}
\alias{control.compute.default}
\title{Control variables in control.compute}
\description{Control variables in \code{control.compute} for use in \code{inla}}
\usage{
inla.set.control.compute.default(...)
control.compute(config, control.gcpo, cpo, dic, graph, hyperpar, internal.opt, likelihood.info, mlik, openmp.strategy, po, q, residuals, return.marginals, return.marginals.predictor, save.memory, smtp, waic)
}
\arguments{
\item{...}{Possible arguments}
\item{openmp.strategy}{   The computational strategy to use: 'small', 'medium',  'large', 'huge', 'default' and 'pardiso'.}

\item{hyperpar}{   A boolean variable if the marginal for the hyperparameters  should be computed. Default TRUE.}

\item{return.marginals}{   A boolean variable if the marginals for the latent  field should be returned (although it is computed). Default TRUE}

\item{return.marginals.predictor}{   A boolean variable if the marginals for  the linear predictor should be returned (although it is computed).  Default FALSE}

\item{dic}{   A boolean variable if the DIC-value should be computed. Default  FALSE.}

\item{mlik}{   A boolean variable if the marginal likelihood should be computed.  Default \code{TRUE}.}

\item{cpo}{   A boolean variable if the cross-validated predictive measures  (cpo, pit) should be computed (default \code{FALSE})}

\item{po}{   A boolean variable if the predictive ordinate should be computed  (default \code{FALSE})}

\item{waic}{   A boolean variable if the Watanabe-Akaike information criteria  should be computed (default \code{FALSE})}

\item{residuals}{   Provide estimates of `residuals` (whatever we mean by that).  (default \code{FALSE})  Currently only residuals base on expected (saturated) deviance are available.  The sign of the residuals are only `very likely` correct.  These residuals are not properly justified from a Bayesian point of  view, hence must be used with caution. It is provided in the hope  they would be useful. This feature is EXPERIMENTAL for the moment,  so changes can happen at any time.}

\item{q}{   A boolean variable if binary images of the precision matrix, the  reordered precision matrix and the Cholesky triangle should be generated.  (Default FALSE.)}

\item{config}{   A boolean variable if the internal GMRF approximations be  stored. (Default FALSE. EXPERIMENTAL)}

\item{likelihood.info}{   A boolean variable to store likelihood-information or not.  This option requires \code{config=TRUE} (Default FALSE. EXPERIMENTAL)}

\item{smtp}{   The sparse-matrix solver, one of 'default', 'taucs', 'band' or  'pardiso' (default \code{inla.getOption("smtp")}). \code{smtp='pardiso'} implies  \code{openmp.strategy='pardiso'}.}

\item{graph}{   A boolean variable if the graph itself should be returned.  (Default FALSE.)}

\item{internal.opt}{   A boolean variable, if to do internal online  optimisations or not. (Default TRUE.)}

\item{save.memory}{   A boolean variable, make choices which  saves memory over accuracy. (Default 'inla.getOption("save.memory")')}

\item{control.gcpo}{   (For experts only!) Set control variables for the gcpo.  The intended use is to use \code{inla.group.cv}.  Refer to \code{?inla.group.cv} and the vignette for details.}

}
\value{
  
The function \code{control.compute} is used to TAB-complete arguments and returns a list of given arguments.
The function \code{inla.set.control.compute.default} returns a list with all the default values of all parameters within this control statement.
}
\seealso{
\code{\link{control.update}}, \code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.scopy}}, \code{\link{control.mix}}, \code{\link{control.pom}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.lp.scale}}, \code{\link{control.pardiso.default}}, \code{\link{control.bgev.default}}, \code{\link{control.family}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
