%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.compute}
\alias{control.compute}
\alias{inla.control.compute}
\alias{inla.set.control.compute.default}
\alias{set.control.compute.default}
\alias{control.compute.default}
\title{Control variables in control.compute}
\description{Control variables in \code{control.compute} for use in \code{inla}}
\usage{
inla.set.control.compute.default(...)
control.compute(config, cpo, dic, gdensity, graph, hyperpar, mlik, openmp.strategy, po, q, return.marginals, smtp, waic)
}
\arguments{
\item{...}{Possible arguments}
\item{openmp.strategy}{  The computational strategy to use: 'small', 'medium', 'large', 'huge' and 'default'. There are also two options for the pardiso solver: 'pardiso.serial' and 'pardiso.parallel'. The difference is how the parallelisation is done, and is tuned for 'small'-sized models, 'medium'-sized models, etc. The default option tries to make an educated guess, but this allows to overide this selection. Default is 'default'}
\item{hyperpar}{  A boolean variable if the marginal for the hyperparameters should be computed. Default TRUE.}
\item{return.marginals}{  A boolean variable if the marginals for the latent field should be returned (although it is computed). Default TRUE}
\item{dic}{  A boolean variable if the DIC-value should be computed. Default FALSE.}
\item{mlik}{  A boolean variable if the marginal likelihood should be computed. Default \code{TRUE}.}
\item{cpo}{  A boolean variable if the cross-validated predictive measures (cpo, pit) should be computed (default \code{FALSE})}
\item{po}{  A boolean variable if the predictive ordinate should be computed (default \code{FALSE})}
\item{waic}{  A boolean variable if the Watanabe-Akaike information criteria should be computed (default \code{FALSE})}
\item{q}{  A boolean variable if binary images of the precision matrix, the reordered precision matrix and the Cholesky triangle should be generated. (Default FALSE.)}
\item{config}{  A boolean variable if the internal GMRF approximations be stored. (Default FALSE. EXPERIMENTAL)}
\item{smtp}{  The sparse-matrix solver, one of 'default', 'taucs', 'band' or 'pardiso' (default \code{inla.getoption("smtp")})}
\item{graph}{  A boolean variable if the graph itself should be returned. (Default FALSE.)}
\item{gdensity}{  A boolean variable if the Gaussian-densities itself should be returned. (Default FALSE.)}
}
\value{
 
The function \code{control.compute} is used to TAB-complete arguments and returns a list of given arguments.
The function \code{inla.set.control.compute.default} returns a list with all the default values of all parameters within this control statement.
}
\seealso{
\code{\link{control.update}}, \code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.mix}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.gev2.default}}, \code{\link{control.family}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.results}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
