%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.data}
\alias{control.data}
\alias{inla.control.data}
\alias{inla.set.control.data.default}
\alias{set.control.data.default}
\alias{control.data.default}
\title{Control variables in control.data}
\description{Control variables in \code{control.data} for use in \code{inla}}
\usage{
inla.set.control.data.default(...)
}
\arguments{
\item{...}{Additional arguments; none for the moment}
}
\value{
 
The function \code{inla.set.control.data.default} returns a list with all the default values of all parameters within this control statement. These are as follows.
}
\seealso{
\code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.mix}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.family}}, \code{\link{control.data}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.results}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
