
\name{control.data}
\alias{control.data}
\alias{inla.control.data}
\title{Control variables in control.data}
\description{Control variables in \code{control.data} for use in \code{inla}}
\usage{
inla(formula, ..., control.data=list( <arg>=<value>[, <arg>=<value>]))
}
\arguments{
}
\seealso{
\code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.mix}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.family}}, \code{\link{control.data}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.results}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
