% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.expert}
\alias{control.expert}
\alias{inla.set.control.expert.default}
\title{control.expert}
\usage{
control.expert(
  cpo.manual = FALSE,
  cpo.idx = -1,
  disable.gaussian.check = FALSE,
  jp = NULL,
  globalconstr = list(A = NULL, e = NULL),
  opt.solve = FALSE,
  opt.storage = FALSE,
  opt.num.threads = TRUE
)

inla.set.control.expert.default(...)
}
\arguments{
\item{cpo.manual}{A boolean variable to decide if the inla-program is to be
runned in a manual-cpo-mode. (EXPERT OPTION: DO NOT USE)}

\item{cpo.idx}{The index/indices of the data point(s) to remove. (EXPERIMENTAL
OPTION: DO NOT USE)}

\item{disable.gaussian.check}{Disable the check for fast computations with a
Gaussian likelihood and identity link (default \code{FALSE})}

\item{jp}{An object of class \code{inla.jp} defining a joint prior}

\item{globalconstr}{Add a global constraint (see \code{?f} and argument
\code{extraconstr}). Note that a global constraint does NOT
correct the normalisation constant.
(EXPERIMENTAL OPTION)}

\item{opt.solve}{Store also \code{L^T} to optimize linear solves (TAUCS only).
Might benefite larger models. (EXPERIMENTAL OPTION)}

\item{opt.storage}{Make sure \code{L} is stored sequencially (TAUCS only).
Might benefite larger models. (EXPERIMENTAL OPTION)}

\item{opt.num.threads}{Reduce the number of threads adaptively if needed at
critical places to optimize wall-clock time}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.family}()},
\code{\link{control.fixed}()},
\code{\link{control.gcpo}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.mode}()},
\code{\link{control.numa}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.scopy}()},
\code{\link{control.sem}()},
\code{\link{control.stiles}()},
\code{\link{control.taucs}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
