%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.expert}
\alias{control.expert}
\alias{inla.control.expert}
\alias{inla.set.control.expert.default}
\alias{set.control.expert.default}
\alias{control.expert.default}
\title{Control variables in control.expert}
\description{Control variables in \code{control.expert} for use in \code{inla}}
\usage{
inla.set.control.expert.default(...)
control.expert(cpo.idx, cpo.manual, disable.gaussian.check, dot.product.gain, globalconstr, jp)
}
\arguments{
\item{...}{Possible arguments}
\item{cpo.manual}{   A boolean variable to decide if the inla-program is to be  runned in a manual-cpo-mode. (EXPERT OPTION: DO NOT USE)}

\item{cpo.idx}{   The index/indices of the data point(s) to remove. (EXPERT  OPTION: DO NOT USE)}

\item{disable.gaussian.check}{   Disable the check for fast computations with a  Gaussian likelihood and identity link (default \code{FALSE})}

\item{jp}{   An object of class \code{inla.jp} defining a joint prior}
\item{dot.product.gain}{   Output the gain in  optimizing dot-products? (Default \code{FALSE})}

\item{globalconstr}{   Add a global constraint (see \code{?f} and argument  \code{extraconstr}). Note that a global constraint does NOT  correct the normalisation constant.}

}
\value{
  
The function \code{control.expert} is used to TAB-complete arguments and returns a list of given arguments.
The function \code{inla.set.control.expert.default} returns a list with all the default values of all parameters within this control statement.
}
\seealso{
\code{\link{control.scopy}}, \code{\link{control.mix}}, \code{\link{control.pom}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.lp.scale}}, \code{\link{control.pardiso.default}}, \code{\link{control.bgev.default}}, \code{\link{control.family}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
