%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.family}
\alias{control.family}
\alias{inla.control.family}
\alias{inla.set.control.family.default}
\alias{set.control.family.default}
\alias{control.family.default}
\title{Control variables in control.family}
\description{Control variables in \code{control.family} for use in \code{inla}}
\usage{
inla.set.control.family.default(...)
control.family(beta.censor.value, cenpoisson.I, control.bgev, control.link, control.mix, control.pom, dummy, fixed, gev.scale.xi, hyper, initial, link, link.simple, param, prior, sn.shape.max, variant)
}
\arguments{
\item{...}{Possible arguments}
\item{dummy}{   A dummy argument that can be used as a workaround}
\item{hyper}{   Definition of the hyperparameters}
\item{initial}{   (OBSOLETE!) Initial value for the hyperparameter(s) of the  likelihood in the internal scale.}

\item{prior}{   (OBSOLETE!) The name of the prior distribution(s) for othe  hyperparameter(s).}

\item{param}{   (OBSOLETE!) The parameters for the prior distribution}
\item{fixed}{   (OBSOLETE!) Boolean variable(s) to say if the hyperparameter(s)  is fixed or random.}

\item{link}{   (OBSOLETE! Use \code{control.link=list(model=)} instead.) The  link function to use.}

\item{sn.shape.max}{   Maximum value for the shape-parameter for Skew Normal  observations (default 5.0)}

\item{gev.scale.xi}{   (Expert option, do not use unless you know what you are  doing.) The internal scaling of the shape-parameter for the GEV distribution.  (default 0.1)}

\item{control.bgev}{   See \code{?control.bgev}}
\item{cenpoisson.I}{   The censoring interval for the censored Poisson}
\item{beta.censor.value}{   The censor value for the Beta-likelihood \code{(0  <= beta.censor.value < 1/2)}}

\item{variant}{   This variable is used to give options for various variants of  the likelihood, like chosing different parameterisations for example. See the  relevant likelihood documentations for options (does only apply to some  likelihoods).}

\item{control.mix}{   See \code{?control.mix}}
\item{control.pom}{   See \code{?control.pom}}
\item{control.link}{   See \code{?control.link}}
\item{link.simple}{   See \code{inla.doc("0inflated")}}
}
\value{
  
The function \code{control.family} is used to TAB-complete arguments and returns a list of given arguments.
The function \code{inla.set.control.family.default} returns a list with all the default values of all parameters within this control statement.
}
\seealso{
\code{\link{control.update}}, \code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.scopy}}, \code{\link{control.mix}}, \code{\link{control.pom}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.lp.scale}}, \code{\link{control.pardiso.default}}, \code{\link{control.bgev.default}}, \code{\link{control.family}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
