% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.family}
\alias{control.family}
\alias{inla.set.control.family.default}
\title{control.family}
\usage{
control.family(
  dummy = 0,
  hyper = NULL,
  initial = NULL,
  prior = NULL,
  param = NULL,
  fixed = NULL,
  link = "default",
  sn.shape.max = 5,
  gev.scale.xi = 0.1,
  control.bgev = NULL,
  cenpoisson.I = c(-1L, -1L),
  beta.censor.value = 0,
  variant = 0L,
  link.simple = "default",
  control.mix = NULL,
  control.pom = NULL,
  control.link = INLA::control.link(),
  control.sem = NULL
)

inla.set.control.family.default(...)
}
\arguments{
\item{dummy}{A dummy argument that can be used as a workaround}

\item{hyper}{Definition of the hyperparameters}

\item{initial}{(OBSOLETE!) Initial value for the hyperparameter(s) of the
likelihood in the internal scale.}

\item{prior}{(OBSOLETE!) The name of the prior distribution(s) for othe
hyperparameter(s).}

\item{param}{(OBSOLETE!) The parameters for the prior distribution}

\item{fixed}{(OBSOLETE!) Boolean variable(s) to say if the hyperparameter(s)
is fixed or random.}

\item{link}{(OBSOLETE! Use \code{control.link=list(model=)} instead.) The
link function to use.}

\item{sn.shape.max}{Maximum value for the shape-parameter for Skew Normal
observations (default 5.0)}

\item{gev.scale.xi}{(Expert option, do not use unless you know what you are
doing.) The internal scaling of the shape-parameter for the GEV distribution.
(default 0.1)}

\item{control.bgev}{See \code{?control.bgev}}

\item{cenpoisson.I}{The censoring interval for the censored Poisson}

\item{beta.censor.value}{The censor value for the Beta-likelihood \verb{(0 <= beta.censor.value < 1/2)}}

\item{variant}{This variable is used to give options for various variants of
the likelihood,  like chosing different parameterisations for example. See the
relevant likelihood documentations for options (does only apply to some
likelihoods).}

\item{link.simple}{See \code{inla.doc("0inflated")}}

\item{control.mix}{See \code{?control.mix}}

\item{control.pom}{See \code{?control.pom}}

\item{control.link}{See \code{?control.link}}

\item{control.sem}{Parameters for likelihood \code{sem}}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.fixed}()},
\code{\link{control.gcpo}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.mode}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.scopy}()},
\code{\link{control.sem}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
