%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.family}
\alias{control.family}
\alias{inla.control.family}
\alias{inla.set.control.family.default}
\alias{set.control.family.default}
\alias{control.family.default}
\title{Control variables in control.family}
\description{Control variables in \code{control.family} for use in \code{inla}}
\usage{
inla.set.control.family.default(...)
control.family(cenpoisson.I, control.link, control.mix, fixed, gev.scale.xi, hyper, initial, link, param, prior, sn.shape.max, variant)
}
\arguments{
\item{...}{Possible arguments}
\item{hyper}{  Definition of the hyperparameters}
\item{initial}{  (OBSOLETE!) Initial value for the hyperparameter(s) of the likelihood in the internal scale.}
\item{prior}{  (OBSOLETE!) The name of the prior distribution(s) for othe hyperparameter(s).}
\item{param}{  (OBSOLETE!) The parameters for the prior distribution}
\item{fixed}{  (OBSOLETE!) Boolean variable(s) to say if the hyperparameter(s) is fixed or random.}
\item{link}{  (OBSOLETE! Use \code{control.link=list(model=)} instead.) The link function to use.}
\item{sn.shape.max}{  Maximum value for the shape-parameter for Skew Normal observations (default 5.0)}
\item{gev.scale.xi}{  The internal scaling of the shape-parameter for the GEV distribution. (default 0.01)}
\item{cenpoisson.I}{  The censoring interval for the censored Poisson}
\item{variant}{  This variable is used to give options for various variants of the likelihood, like chosing different parameterisations for example. See the relevant likelihood documentations for options (does only apply to some likelihoods).}
\item{control.mix}{  See \code{?control.mix}}
\item{control.link}{  See \code{?control.link}}
}
\value{
 
The function \code{control.family} is used to TAB-complete arguments and returns a list of given arguments.
The function \code{inla.set.control.family.default} returns a list with all the default values of all parameters within this control statement.
}
\seealso{
\code{\link{control.update}}, \code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.mix}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.family}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.results}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
