% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.fixed}
\alias{control.fixed}
\alias{inla.set.control.fixed.default}
\title{control.fixed}
\usage{
control.fixed(
  cdf = NULL,
  quantiles = NULL,
  expand.factor.strategy = "model.matrix",
  mean = 0,
  mean.intercept = 0,
  prec = 0.001,
  prec.intercept = 0,
  compute = TRUE,
  correlation.matrix = FALSE,
  remove.names = NULL
)

inla.set.control.fixed.default(...)
}
\arguments{
\item{cdf}{A list of values to compute the CDF for, for all fixed effects}

\item{quantiles}{A list of quantiles to compute for all fixed effects}

\item{expand.factor.strategy}{The strategy used to expand factors into fixed
effects based on their levels. The default strategy is us use the
\code{model.matrix}-function for which NA's are not allowed
(\code{expand.factor.strategy="model.matrix"}) and levels are possible removed.
The alternative option (\code{expand.factor.strategy="inla"}) use an
\code{inla}-specific expansion which expand a factor into one fixed effects for
each level, do allow for NA's and all levels are present in the model. In this
case, factors MUST BE factors in the data.frame/list and NOT added as
\code{.+factor(x1)+.} in the formula only.}

\item{mean}{Prior mean for all fixed effects except the intercept.
Alternatively, a named list with specific means where name=default applies to
unmatched names. For example \code{control.fixed=list(mean=list(a=1, b=2, default=0))} assign 'mean=1' to fixed effect 'a' , 'mean=2' to effect 'b' and
'mean=0' to all others. (default 0.0)}

\item{mean.intercept}{Prior mean for the intercept (default 0.0)}

\item{prec}{Default precision for all fixed effects except the intercept.
Alternatively, a named list with specific means where name=default applies to
unmatched names.  For example \code{control.fixed=list(prec=list(a=1, b=2, default=0.01))} assign 'prec=1' to fixed effect 'a' , 'prec=2' to effect 'b' and
'prec=0.01' to all others. (default 0.001)}

\item{prec.intercept}{Default precision the intercept (default 0.0)}

\item{compute}{Compute marginals for the fixed effects ? (default TRUE)}

\item{correlation.matrix}{Compute the posterior correlation matrix for all
fixed effects? (default FALSE) OOPS: This option will set up appropriate linear
combinations and the results are shown as the posterior correlation matrix of the
linear combinations. This option will imply
\code{control.inla=list(lincomb.derived.correlation.matrix=TRUE)}.}

\item{remove.names}{A vector of names of expanded fixed effects to remove
from the model-matrix. This is an expert option, and should only be used if you
know what you are doing.}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.family}()},
\code{\link{control.gcpo}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.mode}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.scopy}()},
\code{\link{control.sem}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
