% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.gcpo}
\alias{control.gcpo}
\alias{inla.set.control.gcpo.default}
\title{control.gcpo}
\usage{
control.gcpo(
  enable = FALSE,
  num.level.sets = -1,
  size.max = 32,
  strategy = c("posterior", "prior"),
  groups = NULL,
  selection = NULL,
  group.selection = NULL,
  friends = NULL,
  weights = NULL,
  verbose = FALSE,
  epsilon = 0.005,
  prior.diagonal = 1e-04,
  correct.hyperpar = TRUE,
  keep = NULL,
  remove = NULL,
  remove.fixed = TRUE,
  type.cv = "single"
)

inla.set.control.gcpo.default(...)
}
\arguments{
\item{num.level.sets}{Number of level.sets to use. The default value
\code{-1} corresponds to leave-one-out cross-validation. If argument \code{weights} is
used, then this is threshold for the sum of the weights defining a group.}

\item{size.max}{The maximum size (measure in the number of nodes) of a group. If the computed
group-size is larger, it will be truncated to \code{size.max}. Note that: If
\code{weights} are in use, then
this still corresponds to the number of nodes in the group, and not the sum of the
weights. This is ment as an emergency option to avoid the size of the
group to go nuts.}

\item{strategy}{One of \code{"posterior"} or \code{"prior"}. See the
vignette for details.}

\item{groups}{An (optional) predefined list of groups.  See the vignette for
details.}

\item{selection}{An optional list of data-indices to use. If not given, then
all data are used.}

\item{group.selection}{An optional list of data-indices to use when building the
groups. If given, each group beyond the observation itself, must be a subset of
\code{group.selection}. If not given, then all data are used.}

\item{friends}{An optional list of lists of indices to use a friends}

\item{weights}{An optional positive weight attached to each datapoint. The sum
of the weights define the size of the group. If \code{NULL}, then unit weight is
used.}

\item{verbose}{Run with \code{verbose} output of some of the internals in
the calculations. This option will also enable \code{inla(..., verbose=TRUE)} if its not enabled already.}

\item{epsilon}{Two correlations with a difference less than \code{epsilon},
will be classified as identical.}

\item{prior.diagonal}{When \code{strategy="prior"}, \code{prior.diagonal} is
added to the diagonal of the prior precision matrix to avoid singularities}

\item{keep}{For \code{strategy="prior"}, then this gives a vector of the
name of model-components TO USE when computing the groups. See the vignette
for details. Not both of \code{keep} and \code{remove} can be defined.}

\item{remove}{For \code{strategy="prior"}, then this gives a vector of the
name of model-components NOT TO USE when computing the groups. See the
vignette for details. Not both of \code{keep} and \code{remove} can be
defined.}

\item{remove.fixed}{For \code{strategy="prior"}, this is the default option
which is in effect if both \code{keep} and \code{remove} are \code{NULL}. If
\code{TRUE}, it will remove (or condition on) all fixed effects when
computing the groups. See the vignette for details.}

\item{type.cv}{Type of cv, either "single" (default) or "joint"}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\details{
(For experts only!) Set control variables for the gcpo in \link{control.compute}.
The intended use is to use \code{\link[=inla.group.cv]{inla.group.cv()}}.
Refer to \code{\link[=inla.group.cv]{inla.group.cv()}} and the vignette for details.
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.family}()},
\code{\link{control.fixed}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.mode}()},
\code{\link{control.numa}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.scopy}()},
\code{\link{control.sem}()},
\code{\link{control.stiles}()},
\code{\link{control.taucs}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
