% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.gcpo}
\alias{control.gcpo}
\alias{inla.set.control.gcpo.default}
\title{control.gcpo}
\usage{
control.gcpo(
  enable = FALSE,
  num.level.sets = -1,
  size.max = 32,
  strategy = c("posterior", "prior"),
  groups = NULL,
  selection = NULL,
  friends = NULL,
  verbose = FALSE,
  epsilon = 0.005,
  prior.diagonal = 1e-04,
  correct.hyperpar = TRUE,
  keep = NULL,
  remove = NULL,
  remove.fixed = TRUE
)

inla.set.control.gcpo.default(...)
}
\arguments{
\item{enable}{TODO}

\item{num.level.sets}{TODO}

\item{size.max}{TODO}

\item{strategy}{TODO}

\item{groups}{TODO}

\item{selection}{TODO}

\item{friends}{TODO}

\item{verbose}{TODO}

\item{epsilon}{TODO}

\item{prior.diagonal}{TODO}

\item{correct.hyperpar}{TODO}

\item{keep}{TODO}

\item{remove}{TODO}

\item{remove.fixed}{TODO}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\details{
(For experts only!) Set control variables for the gcpo in \link{control.compute}.
The intended use is to use \code{inla.group.cv}.
Refer to \code{?inla.group.cv} and the vignette for details.
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.family}()},
\code{\link{control.fixed}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.mode}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.scopy}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
