% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.lincomb}
\alias{control.lincomb}
\alias{inla.set.control.lincomb.default}
\title{control.lincomb}
\usage{
control.lincomb(verbose = FALSE)

inla.set.control.lincomb.default(...)
}
\arguments{
\item{verbose}{Use verbose mode for linear combinations if verbose model is
set globally. (Default FALSE). This option is only available for
the default \code{inla.mode} (\code{inla.mode="compact"}).}

\item{...}{Named arguments.}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\seealso{
Other control: 
\code{\link{control.group}()},
\code{\link{control.update}()}
}
\concept{control}
