%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.lincomb}
\alias{control.lincomb}
\alias{inla.control.lincomb}
\alias{inla.set.control.lincomb.default}
\alias{set.control.lincomb.default}
\alias{control.lincomb.default}
\title{Control variables in control.lincomb}
\description{Control variables in \code{control.lincomb} for use in \code{inla}}
\usage{
inla.set.control.lincomb.default(...)
control.lincomb(precision, verbose)
}
\arguments{
\item{...}{Possible arguments}
\item{precision}{  The precision for the artificial tiny noise. Default 1e09.}
\item{verbose}{  Use verbose mode for linear combinations if verbose model is set globally. (Default TRUE)}
}
\value{
 
The function \code{control.lincomb} is used to TAB-complete arguments and returns a list of given arguments.
The function \code{inla.set.control.lincomb.default} returns a list with all the default values of all parameters within this control statement.
}
\seealso{
\code{\link{control.update}}, \code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.mix}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.gev2.default}}, \code{\link{control.family}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.results}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
