% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.link}
\alias{control.link}
\alias{inla.set.control.link.default}
\title{control.link}
\usage{
control.link(
  model = "default",
  order = NULL,
  variant = NULL,
  hyper = NULL,
  quantile = NULL,
  a = 1,
  initial = NULL,
  fixed = NULL,
  prior = NULL,
  param = NULL
)

inla.set.control.link.default(...)
}
\arguments{
\item{model}{The name of the link function/model}

\item{order}{The \code{order} of the link function, where the interpretation
of \code{order} is model-dependent.}

\item{variant}{The \code{variant} of the link function, where the
interpretation of \code{variant} is model-dependent.}

\item{hyper}{Definition of the hyperparameter(s) for the link model chosen}

\item{quantile}{The quantile for quantile link function}

\item{a}{The parameter \code{a} in the LOGa link}

\item{initial}{(OBSOLETE!) The initial value(s) for the hyperparameter(s)}

\item{fixed}{(OBSOLETE!) A boolean variable if hyperparmater(s) is/are fixed
or random}

\item{prior}{(OBSOLETE!) The name of the prior distribution(s) for the
hyperparmater(s)}

\item{param}{(OBSOLETE!) The parameters for the prior distribution(s) for
the hyperparmater(s)}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\details{
The \code{control.link}-list is set within the corresponding \code{control.family}-list as the link is likelihood-family specific.
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.family}()},
\code{\link{control.fixed}()},
\code{\link{control.gcpo}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.mode}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.scopy}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
