%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.lp.scale}
\alias{control.lp.scale}
\alias{inla.control.lp.scale}
\alias{inla.set.control.lp.scale.default}
\alias{set.control.lp.scale.default}
\alias{control.lp.scale.default}
\title{Control variables in control.lp.scale}
\description{Control variables in \code{control.lp.scale} for use in \code{inla}}
\usage{
inla.set.control.lp.scale.default(...)
control.lp.scale(hyper)
}
\arguments{
\item{...}{Possible arguments}
\item{hyper}{   Definition of the hyperparameter(s)}
}
\value{
  
The function \code{control.lp.scale} is used to TAB-complete arguments and returns a list of given arguments.
The function \code{inla.set.control.lp.scale.default} returns a list with all the default values of all parameters within this control statement.
}
\seealso{
\code{\link{control.update}}, \code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.group}}, \code{\link{control.mix}}, \code{\link{control.pom}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.lp.scale}}, \code{\link{control.pardiso.default}}, \code{\link{control.bgev.default}}, \code{\link{control.family}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
