%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.mix}
\alias{control.mix}
\alias{inla.control.mix}
\alias{inla.set.control.mix.default}
\alias{set.control.mix.default}
\alias{control.mix.default}
\title{Control variables in control.mix}
\description{Control variables in \code{control.mix} for use in \code{inla}}
\usage{
inla.set.control.mix.default(...)
control.mix(fixed, hyper, initial, integrator, model, npoints, param, prior)
}
\arguments{
\item{...}{Possible arguments}
\item{model}{  The model for the random effect. Currently, only \code{model='gaussian'} is implemented}
\item{hyper}{  Definition of the hyperparameter(s) for the random effect model chosen}
\item{initial}{  (OBSOLETE!) The initial value(s) for the hyperparameter(s)}
\item{fixed}{  (OBSOLETE!) A boolean variable if hyperparmater(s) is/are fixed or random}
\item{prior}{  (OBSOLETE!) The name of the prior distribution(s) for the hyperparmater(s)}
\item{param}{  (OBSOLETE!) The parameters for the prior distribution(s) for the hyperparmater(s)}
\item{npoints}{  Number of points used to do the numerical integration (default 101)}
\item{integrator}{  The integration scheme to use (\code{default}, \code{quadrature}, \code{simpson})}
}
\value{
  The \code{control.mix} -list is set within the corresponding \code{control.family}-list a the mixture of the likelihood is likelihood spesific. (This option is EXPERIMENTAL.)
The function \code{control.mix} is used to TAB-complete arguments and returns a list of given arguments.
The function \code{inla.set.control.mix.default} returns a list with all the default values of all parameters within this control statement.
}
\seealso{
\code{\link{control.update}}, \code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.mix}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.gev2.default}}, \code{\link{control.family}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.results}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
