% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.mode}
\alias{control.mode}
\alias{inla.set.control.mode.default}
\title{control.mode}
\usage{
control.mode(
  result = NULL,
  theta = NULL,
  x = NULL,
  restart = TRUE,
  fixed = FALSE
)

inla.set.control.mode.default(...)
}
\arguments{
\item{result}{Previous result-object from inla(), a inla-state object or the name of a
state-file. Use the \code{theta}- and \code{x}-mode from this object}

\item{theta}{The theta-mode/initial values for theta. This option has
preference over result$mode$theta.}

\item{x}{The x-mode/initial values for x. This option has preference over
result$mode$x.}

\item{restart}{A boolean variable; should we restart the optimisation from
the given configuration? If \code{TRUE} (default),  then
use this configuration as the initial
values (both \code{theta} and \code{x}) and optimize. If \code{FALSE}, then use \code{theta} as
as the mode and \code{x} as the initial value. If \code{x != NULL},  \code{theta=NULL}
and \code{restart=TRUE}, then this will give an error unless number of
hyperparameters are zero.}

\item{fixed}{A boolean variable. If \code{TRUE} then treat all theta's as known and
fixed, and if FALSE (default) then treat all theta's as unknown and random.
If \code{fixed=TRUE} and \code{restart=TRUE},  then \code{restart} is assigned to \code{FALSE} and a
warning is issued.  Note that \code{fixed=TRUE} will change the model as the
corresponding hyperparmaeters will be defined as fixed.}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\details{
For internal use only and for algorithms built on to of INLA.
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.family}()},
\code{\link{control.fixed}()},
\code{\link{control.gcpo}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.scopy}()},
\code{\link{control.sem}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
