% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.mode}
\alias{control.mode}
\alias{inla.set.control.mode.default}
\title{control.mode}
\usage{
control.mode(
  result = NULL,
  theta = NULL,
  x = NULL,
  restart = FALSE,
  fixed = FALSE
)

inla.set.control.mode.default(...)
}
\arguments{
\item{result}{Previous result from inla(). Use the theta- and x-mode from
this run.}

\item{theta}{The theta-mode/initial values for theta. This option has
preference over result$mode$theta.}

\item{x}{The x-mode/initial values for x. This option has preference over
result$mode$x.}

\item{restart}{A boolean variable; should we restart the optimisation from
this configuration or fix the mode at this configuration? (Default FALSE.)}

\item{fixed}{A boolean variable. If TRUE then treat all thetas as known and
fixed, and if FALSE then treat all thetas as unknown and random (default).}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\details{
For internal use and for algorithms built on to of INLA.
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.family}()},
\code{\link{control.fixed}()},
\code{\link{control.gcpo}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.scopy}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
