% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.pardiso}
\alias{control.pardiso}
\alias{inla.set.control.pardiso.default}
\title{control.pardiso}
\usage{
control.pardiso(
  verbose = FALSE,
  debug = FALSE,
  parallel.reordering = TRUE,
  nrhs = -1
)

inla.set.control.pardiso.default(...)
}
\arguments{
\item{verbose}{Show detailed output (default FALSE)}

\item{debug}{Show internal debug output (default FALSE)}

\item{parallel.reordering}{Do reordering in parallel (default TRUE)}

\item{nrhs}{Number of right-hand sides to solve for in parallel (\code{-1}
will determine this adapative)}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\details{
Extra options controlling the PARDISO library
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.family}()},
\code{\link{control.fixed}()},
\code{\link{control.gcpo}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.mode}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.scopy}()},
\code{\link{control.sem}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
