% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.predictor}
\alias{control.predictor}
\alias{inla.set.control.predictor.default}
\title{control.predictor}
\usage{
control.predictor(
  hyper = NULL,
  fixed = NULL,
  prior = NULL,
  param = NULL,
  initial = NULL,
  compute = FALSE,
  cdf = NULL,
  quantiles = NULL,
  cross = NULL,
  A = NULL,
  precision = exp(15),
  link = NULL
)

inla.set.control.predictor.default(...)
}
\arguments{
\item{hyper}{Definition of the hyperparameters.}

\item{fixed}{(OBSOLETE!) If the precision for the artificial noise is fixed
or not (default TRUE)}

\item{prior}{(OBSOLETE!) The prior for the artificial noise}

\item{param}{(OBSOLETE!) Prior parameters for the artificial noise}

\item{initial}{(OBSOLETE!) The value of the log precision of the artificial
noise}

\item{compute}{A boolean variable; should the marginals for the linear
predictor be computed? (Default FALSE.)}

\item{cdf}{A list of values to compute the CDF for the linear predictor}

\item{quantiles}{A list of quantiles to compute for the linear predictor}

\item{cross}{Cross-sum-to-zero constraints with the linear predictor. All
linear predictors with the same level of 'cross' are constrained to have sum
zero. Use 'NA' for no contribution. 'Cross' has the same length as the linear
predictor (including the 'A' matrix extention). (THIS IS AN EXPERIMENTAL OPTION,
CHANGES MAY APPEAR.)}

\item{A}{The observation matrix (matrix or Matrix::sparseMatrix).}

\item{precision}{The precision for eta* - A*eta, (default \code{exp(15)})}

\item{link}{Define the family-connection for unobserved observations
(\code{NA}). \code{link} is integer values which defines the family connection;
\code{family[link[idx]]} unless \code{is.na(link[idx])} for which the
identity-link is used. The \code{link}-argument only influence the
\code{fitted.values} in the \code{result}-object. If \code{is.null(link)}
(default) then the identity-link is used for all missing observations. If the
length of \code{link} is 1, then this value is replicated with the length of the
response vector. If an element of the response vector is \code{!NA} then the
corresponding entry in \code{link} is not used (but must still be a legal value).
Setting this variable implies \code{compute=TRUE}.}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.family}()},
\code{\link{control.fixed}()},
\code{\link{control.gcpo}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.mode}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.scopy}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
