
\name{control.results}
\alias{control.results}
\alias{inla.control.results}
\alias{inla.set.control.results.default}
\alias{set.control.results.default}
\alias{control.results.default}
\title{Control variables in control.results}
\description{Control variables in \code{control.results} for use in \code{inla}}
\usage{
inla.set.control.results.default()
}
\arguments{}
\value{
The function \code{inla.set.control.results.default} returns a list with all the default values of all parameters within this control statement. These are as follows.
\item{return.marginals.random}{  A boolean variable; read the marginals for the fterms? (Default TRUE)}
\item{return.marginals.predictor}{  A boolean variable; read the marginals for the linear predictor? (Default TRUE)}
}
\seealso{
\code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.mix}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.family}}, \code{\link{control.data}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.results}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
