% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.scopy}
\alias{control.scopy}
\alias{inla.set.control.scopy.default}
\title{control.scopy}
\usage{
control.scopy(
  covariate = NULL,
  n = 5,
  model = "rw2",
  mean = 1,
  prec.mean = 1,
  prec.betas = 10
)

inla.set.control.scopy.default(...)
}
\arguments{
\item{covariate}{The covariate for the scopy function}

\item{n}{Number of betas}

\item{model}{scopy model (one of 'rw1' and 'rw2')}

\item{mean}{The prior mean for mean(betas)}

\item{prec.mean}{The prior precision for mean(betas)}

\item{prec.betas}{The prior precision prec(betas-mean(betas))}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.family}()},
\code{\link{control.fixed}()},
\code{\link{control.gcpo}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.mode}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
