%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.scopy}
\alias{control.scopy}
\alias{inla.control.scopy}
\alias{inla.set.control.scopy.default}
\alias{set.control.scopy.default}
\alias{control.scopy.default}
\title{Control variables in control.scopy}
\description{Control variables in \code{control.scopy} for use in \code{inla}}
\usage{
inla.set.control.scopy.default(...)
control.scopy(covariate, mean, model, n, precision, precision)
}
\arguments{
\item{...}{Possible arguments}
\item{covariate}{   The covariate for the scopy function}
\item{n}{   Number of betas}
\item{model}{   scopy model (one of 'rw1' and 'rw2')}
\item{mean}{   The prior mean for mean(betas)}
\item{precision}{   The prior precision for mean(betas)}
\item{precision}{   The prior precision prec(betas-mean(betas))}
}
\value{
  
The function \code{control.scopy} is used to TAB-complete arguments and returns a list of given arguments.
The function \code{inla.set.control.scopy.default} returns a list with all the default values of all parameters within this control statement.
}
\seealso{
\code{\link{control.update}}, \code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.scopy}}, \code{\link{control.mix}}, \code{\link{control.pom}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.lp.scale}}, \code{\link{control.pardiso.default}}, \code{\link{control.bgev.default}}, \code{\link{control.family}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
