% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.sem}
\alias{control.sem}
\alias{inla.set.control.sem.default}
\title{control.sem}
\usage{
control.sem(B = NULL, idx = 0)

inla.set.control.sem.default(...)
}
\arguments{
\item{B}{The symbolic B-matrix,  where each element is a string giving the
expression for that particular element, in terms of beta-parameters for copy}

\item{idx}{Which diagonal element to use for the variance.}
}
\description{
control.sem
}
\details{
Parameters to family \code{sem}
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.family}()},
\code{\link{control.fixed}()},
\code{\link{control.gcpo}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.mode}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.scopy}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
