% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.stiles}
\alias{control.stiles}
\alias{inla.set.control.stiles.default}
\title{control.stiles}
\usage{
control.stiles(verbose = FALSE, block.size = -1, param = rep(-1, 20))

inla.set.control.stiles.default(...)
}
\arguments{
\item{verbose}{Show detailed output (default FALSE)}

\item{block.size}{(integer) Preferred number of rhs's in each parallel solve. Used to
split many rhs's into threads. Default value (-1) is to use the 'tile.size'
from sTiles, which is a reasonable tradeoff between speed and memory use.
A higher value will increase memory usage.}

\item{param}{An integer vector of parameters (variable length).
Default values are given by '-1'.
See sTiles documentation for explaination of these parameters}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\details{
Extra options controlling the sTiles library
}
\seealso{
Other control:
\code{\link[=control.bgev]{control.bgev()}},
\code{\link[=control.compute]{control.compute()}},
\code{\link[=control.expert]{control.expert()}},
\code{\link[=control.family]{control.family()}},
\code{\link[=control.fixed]{control.fixed()}},
\code{\link[=control.gcpo]{control.gcpo()}},
\code{\link[=control.group]{control.group()}},
\code{\link[=control.hazard]{control.hazard()}},
\code{\link[=control.inla]{control.inla()}},
\code{\link[=control.lincomb]{control.lincomb()}},
\code{\link[=control.link]{control.link()}},
\code{\link[=control.lp.scale]{control.lp.scale()}},
\code{\link[=control.mix]{control.mix()}},
\code{\link[=control.mode]{control.mode()}},
\code{\link[=control.numa]{control.numa()}},
\code{\link[=control.pardiso]{control.pardiso()}},
\code{\link[=control.pom]{control.pom()}},
\code{\link[=control.predictor]{control.predictor()}},
\code{\link[=control.scopy]{control.scopy()}},
\code{\link[=control.sem]{control.sem()}},
\code{\link[=control.taucs]{control.taucs()}},
\code{\link[=control.update]{control.update()}},
\code{\link[=control.vb]{control.vb()}}
}
\concept{control}
