% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.stiles}
\alias{control.stiles}
\alias{inla.set.control.stiles.default}
\title{control.stiles}
\usage{
control.stiles(verbose = FALSE, tile.size = 0)

inla.set.control.stiles.default(...)
}
\arguments{
\item{verbose}{Show detailed output (default FALSE)}

\item{tile.size}{The size of the tile (default 0 will chose automatically)}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\details{
Extra options controlling the sTiles library
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.family}()},
\code{\link{control.fixed}()},
\code{\link{control.gcpo}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.mode}()},
\code{\link{control.numa}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.scopy}()},
\code{\link{control.sem}()},
\code{\link{control.taucs}()},
\code{\link{control.update}()},
\code{\link{control.vb}()}
}
\concept{control}
