% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.update}
\alias{control.update}
\alias{inla.set.control.update.default}
\title{control.update}
\usage{
control.update(result = NULL)

inla.set.control.update.default(...)
}
\arguments{
\item{result}{Update the joint posterior for the hyperparameters from result}

\item{...}{Named arguments.}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\seealso{
\code{\link[=inla]{inla()}}

Other control: 
\code{\link{control.group}()},
\code{\link{control.lincomb}()}
}
\concept{control}
