%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from set.default.arguments.R
%%
\name{control.update}
\alias{control.update}
\alias{inla.control.update}
\alias{inla.set.control.update.default}
\alias{set.control.update.default}
\alias{control.update.default}
\title{Control variables in control.update}
\description{Control variables in \code{control.update} for use in \code{inla}}
\usage{
inla.set.control.update.default(...)
control.update(result)
}
\arguments{
\item{...}{Possible arguments}
\item{result}{  Update the joint posterior for the hyperparameters from result}
}
\value{
 
The function \code{control.update} is used to TAB-complete arguments and returns a list of given arguments.
The function \code{inla.set.control.update.default} returns a list with all the default values of all parameters within this control statement.
}
\seealso{
\code{\link{control.update}}, \code{\link{control.lincomb}}, \code{\link{control.group}}, \code{\link{control.mix}}, \code{\link{control.link}}, \code{\link{control.expert}}, \code{\link{control.compute}}, \code{\link{control.gev2.default}}, \code{\link{control.family}}, \code{\link{control.fixed}}, \code{\link{control.inla}}, \code{\link{control.predictor}}, \code{\link{control.results}}, \code{\link{control.mode}}, \code{\link{control.hazard}}, 
\code{\link{inla}}
}
