% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.default.arguments.R
\name{control.vb}
\alias{control.vb}
\alias{inla.set.control.vb.default}
\title{control.fixed}
\usage{
control.vb(
  enable = "auto",
  strategy = c("mean", "variance"),
  verbose = TRUE,
  iter.max = 25,
  emergency = 25,
  f.enable.limit = c(30, 25),
  hessian.update = 2,
  hessian.strategy = c("default", "full", "partial", "diagonal")
)

inla.set.control.vb.default(...)
}
\arguments{
\item{enable}{Logical/Character Use this feature? If \code{"auto"} this will be selected automatically.}

\item{strategy}{Charactor What to correct, either "mean" or "variance".}

\item{verbose}{Logical Be verbose or not.}

\item{iter.max}{Integer Maximum number of interations.}

\item{emergency}{Numeric If the standarized correction for the mean is larger
than this value, then call the vb.correction off and issue a warning}

\item{f.enable.limit}{Vector of length 2. The size limit to correct for a \code{f()}.
First element is for \code{strategy="mean"}.
Second element is for \code{strategy="mean"}.}

\item{hessian.update}{How many times the Hessian is updated for each
correction (\code{strategy="variance"} only).}

\item{hessian.strategy}{Select strategy for computing the Hessian
matrix for \code{strategy="variance"}, one of \code{"full"},
\code{"diagonal"}, \code{"partial"} and \code{"default"}.}

\item{...}{Named arguments passed on to the main function}
}
\description{
Control variables in \verb{control.*} for use with \code{\link[=inla]{inla()}}.
The functions can be used to TAB-complete arguments, and
returns a list of the default control arguments, unless overridden by
specific input arguments.
}
\details{
control.vb List of arguments for various VB corrections.
Used for \link{control.inla} \code{control.vb} specifications.
}
\seealso{
Other control: 
\code{\link{control.bgev}()},
\code{\link{control.compute}()},
\code{\link{control.expert}()},
\code{\link{control.family}()},
\code{\link{control.fixed}()},
\code{\link{control.gcpo}()},
\code{\link{control.group}()},
\code{\link{control.hazard}()},
\code{\link{control.inla}()},
\code{\link{control.lincomb}()},
\code{\link{control.link}()},
\code{\link{control.lp.scale}()},
\code{\link{control.mix}()},
\code{\link{control.mode}()},
\code{\link{control.pardiso}()},
\code{\link{control.pom}()},
\code{\link{control.predictor}()},
\code{\link{control.scopy}()},
\code{\link{control.update}()}
}
\concept{control}
