% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.graph.R
\name{debug.graph}
\alias{debug.graph}
\alias{inla.debug.graph}
\title{Debug a graph-file}
\usage{
inla.debug.graph(graph.file)
}
\arguments{
\item{graph.file}{The filename of the graph (ascii-mode)}
}
\value{
If an error is found, then an error message is shows, otherwise the
graph-object returned by \code{inla.read.graph()} is returned.
}
\description{
Debug a graph specification on file (ascii-mode only), by checking the
specification along the way.
}
\examples{

\dontrun{
cat("3\n 1 1 2n\ 2 1 1\n 3 4\n", file="g.dat")
g = inla.debug.graph("g.dat")
}

}
\seealso{
inla.read.graph
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
