%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from dev.new.R
%%
 \name{inla.dev.new}
 \alias{inla.dev.new}

 \title{Opens a new device}

 \description{Open a new device using \code{\link{dev.new}} unless using RStudio}
 \usage{
     inla.dev.new(...)
 }
 \arguments{
   \item{...}{Optional arguments to \code{\link{dev.new}}} 
}
 \value{
   The value of \code{\link{dev.new}} if not running RStudio, otherwise \code{NULL} 
 }
 \author{Havard Rue \email{hrue@r-inla.org}}
