% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{inla.doc}
\alias{inla.doc}
\alias{inla.inla.doc}
\title{View documentation}
\usage{
inla.doc(what, section, verbose = FALSE)
}
\arguments{
\item{what}{What to view documentation about; name of latent model, name of
prior, etc. (A regular expression.)}

\item{section}{An optional section, like \code{names(inla.models())}, to
look for the documentation. If missing, all sections are used.}

\item{verbose}{Logical If \code{TRUE} then run in verbose mode}
}
\description{
View documentation of latent, prior and likelihood models.
}
\examples{

\dontrun{inla.doc("rw2")}
\dontrun{inla.doc("gaussian", section = "prior")}

}
\seealso{
\code{www.r-inla.org}
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
