%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from doc.R
%%
\name{inla.doc}
\alias{inla.doc}
\alias{inla.doc}
\alias{inla.inla.doc}
\title{View documentation}
\description{View documentation of latent, prior and likelihood models.}
\usage{
inla.doc(what, sec, verbose=FALSE)
}
\arguments{
\item{what}{What to view documentation about;
            name of latent model,  name of prior,  etc. (A regular expression.)}
\item{sec}{An optional section to look for the documentation. If missing, all sections are used.}
\item{verbose}{Logical if \code{TRUE} then run in verbose mode}
}
\author{Havard Rue \email{hrue@r-inla.org}}
\seealso{\code{www.r-inla.org}}
\examples{
\dontrun{inla.doc("rw2")}
\dontrun{inla.doc("gaussian")}
}
