% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryrun.R
\name{dryrun}
\alias{dryrun}
\alias{inla.dryrun}
\title{Do a dryrun to extract some internal information upfront}
\usage{
inla.dryrun(...)
}
\arguments{
\item{...}{Same arguments as \code{inla}()}
}
\value{
A list of start-index and length for each latent component and a
list of the hyperparameters in the model
}
\description{
Do a dryrun to get information about the internal storage and the list (and
ordering) of the hyperparameters
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
