% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{extract.groups}
\alias{extract.groups}
\alias{extract.groups.inla.mesh.segment}
\title{Extract tagged boundary/internal segments.}
\usage{
extract.groups(segm, groups, groups.new = groups, ...)
}
\arguments{
\item{segm}{An \code{\link[=inla.mesh.segment]{inla.mesh.segment()}} object.}

\item{groups}{The segment groups id:s to extract.}

\item{groups.new}{Optional vector of group id remapping; \code{groups[k]} in
the input will be replaced by \code{groups.new[k]} in the output.}

\item{\dots}{Additional arguments, passed on to other methods.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[fmesher:fm_segm]{fmesher::fm_segm()}} instead.

Extract boundary or internal segments tagged by group id:s.
}
\seealso{
\code{\link[=inla.mesh.segment]{inla.mesh.segment()}}
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
