% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{Define general Gaussian models in the INLA formula}
\usage{
f(
  ...,
  model = "iid",
  copy = NULL,
  scopy = NULL,
  same.as = NULL,
  n = NULL,
  nrep = NULL,
  replicate = NULL,
  ngroup = NULL,
  group = NULL,
  control.group = inla.set.control.group.default(),
  control.scopy = inla.set.control.scopy.default(),
  hyper = NULL,
  initial = NULL,
  prior = NULL,
  param = NULL,
  fixed = NULL,
  season.length = NULL,
  constr = NULL,
  extraconstr = list(A = NULL, e = NULL),
  values = NULL,
  cyclic = NULL,
  diagonal = NULL,
  graph = NULL,
  graph.file = NULL,
  cdf = NULL,
  quantiles = NULL,
  Cmatrix = NULL,
  rankdef = NULL,
  Z = NULL,
  nrow = NULL,
  ncol = NULL,
  nu = NULL,
  bvalue = NULL,
  spde.prefix = NULL,
  spde2.prefix = NULL,
  spde2.transform = c("logit", "log", "identity"),
  spde3.prefix = NULL,
  spde3.transform = c("logit", "log", "identity"),
  mean.linear = inla.set.control.fixed.default()$mean,
  prec.linear = inla.set.control.fixed.default()$prec,
  compute = TRUE,
  of = NULL,
  precision = 10^8,
  range = NULL,
  adjust.for.con.comp = TRUE,
  order = NULL,
  scale = NULL,
  rgeneric = NULL,
  cgeneric = NULL,
  scale.model = NULL,
  args.slm = list(rho.min = NULL, rho.max = NULL, X = NULL, W = NULL, Q.beta = NULL),
  args.ar1c = list(Z = NULL, Q.beta = NULL),
  args.intslope = list(subject = NULL, strata = NULL, covariates = NULL),
  vb.correct = TRUE,
  locations = NULL,
  debug = FALSE
)
}
\arguments{
\item{...}{Name of the covariate and, possibly of the
weights vector. NB: order counts!!!! The first specified
term is the covariate and the second one is the vector of
weights (which can be negative).}

\item{model}{A string indicating the chosen model. The
default is \code{iid}. See
\code{names(inla.models()$latent)} for a list of possible
alternatives and \code{\link[=inla.doc]{inla.doc()}} for detailed docs.}

\item{copy}{The name of the model-component to copy}

\item{scopy}{The name of the model-component to smooth-copy (where the
copy-function is a spline)}

\item{same.as}{Can be used with \code{copy=".."}. \code{same.as="A"} says
that this \code{copy} should use the same scaling parameter as another
\code{copy} "A"}

\item{n}{An optional argument which defines the dimension
of the model if this is different from
\code{length(sort(unique(covariate)))}}

\item{nrep}{Number of replications, if not given,  then \code{nrep=max(replications)}}

\item{replicate}{A vector of which replications to use.}

\item{ngroup}{Number of groups,  if not given,  then \code{ngroup=max(group)}}

\item{group}{A vector of which groups to use.}

\item{control.group}{Controls the use of \code{group}}

\item{control.scopy}{Controls the use of \code{scopy}}

\item{hyper}{Specification of the hyperparameter, fixed or
random, initial values, priors and its parameters. See
\code{?inla.models} for the list of hyparameters for each
model and its default options or
use \code{inla.doc()} for
detailed info on the family and
supported prior distributions.}

\item{initial}{THIS OPTION IS OBSOLETE,  DO NOT USE}

\item{prior}{THIS OPTION IS OBSOLETE,  DO NOT USE}

\item{param}{THIS OPTION IS OBSOLETE,  DO NOT USE}

\item{fixed}{THIS OPTION IS OBSOLETE; DO NOT USE}

\item{season.length}{Length of the seasonal component for \code{model="seasonal"}}

\item{constr}{A boolean variable indicating whater to set
a sum to 0 constraint on the term. By default the sum to 0
constraint is imposed on all intrinsic models
("iid","rw1","rw1","besag", etc..).}

\item{extraconstr}{This argument defines extra linear
constraints. The argument is a list with two elements, a
matrix \code{A} and a vector \code{e}, which defines the
extra constraint \code{Ax = e}; for example
\code{extraconstr = list(A = A, e=e)}. The number of
columns of \code{A} must correspond to the length of this
\code{f}-model.  Note that this constraint comes
additional to the sum-to-zero constraint defined if
\code{constr = TRUE}.}

\item{values}{An optional vector giving all values
assumed by the covariate for which we want estimated the
effect. It must be a numeric vector, a vector of factors
or \code{NULL}.}

\item{cyclic}{A boolean specifying wheather the model is
cyclical. Only valid for "rw1" and "rw2" models, is
cyclic=T then the sum to 0 constraint is removed. For the
correct form of the grah file see \cite{Martino and Rue
(2008)}.}

\item{diagonal}{An extra constant added to the diagonal of
the precision matrix to prevent numerical issues.}

\item{graph}{Defines the graph-object either as a file with
a graph-description, an \code{inla.graph}-object, or as a
(sparse) symmetric matrix .}

\item{graph.file}{THIS OPTION IS OBSOLETE,  DO NOT USE}

\item{cdf}{THIS OPTION IS OBSOLETE,  DO NOT USE}

\item{quantiles}{A vector of maximum 10 quantiles,
\eqn{p(0), p(1),\dots}{p(0), p(1),\ldots} to compute for
each posterior marginal. The function returns, for each
posterior marginal, the values
\eqn{x(0), x(1),\dots}{x(0), x(1),\ldots} such that
\deqn{\mbox{Prob}(X<x(p))=p}{Prob(X<x)=p}}

\item{Cmatrix}{The specification of the precision matrix
for the generic,  generic3 or z models (up to a scaling constant).
\code{Cmatrix} is either a
(dense) matrix, a matrix created using
\code{Matrix::sparseMatrix()}, or a filename which stores the
non-zero elements of \code{Cmatrix}, in three columns:
\code{i}, \code{j} and \code{Qij}. In case of the generic3 model,
it is a list of such specifications.}

\item{rankdef}{A number \strong{defining} the rank
deficiency of the model, with sum-to-zero constraint and
possible extra-constraints taken into account. See
details.}

\item{Z}{The matrix for the z-model}

\item{nrow}{Number of rows for 2d-models}

\item{ncol}{Number of columns for 2d-models}

\item{nu}{Smoothing parameter for the Matern2d-model,
possible values are \code{c(0, 1, 2, 3)}}

\item{bvalue}{The boundary conditions for model \code{rw2d},  0 means use
the correct subspace (default),  while 1 means condition on 0's outside}

\item{spde.prefix}{Internal use only}

\item{spde2.prefix}{Internal use only}

\item{spde2.transform}{Internal use only}

\item{spde3.prefix}{Internal use only}

\item{spde3.transform}{Internal use only}

\item{mean.linear}{Prior mean for \code{model="linear"}}

\item{prec.linear}{Prior precision for \code{model="linear"}}

\item{compute}{A boolean variable indicating whether the
marginal posterior distribution for the nodes in the
\code{f()} model should be computed or not. This is
usefull for large models where we are only interested in
some posterior marginals.}

\item{of}{Internal use only}

\item{precision}{The precision for the artificial noise added when creating a copy of a model and others.}

\item{range}{A vector of size two giving the lower and
upper range for the scaling parameter \code{beta} in the
model \code{COPY}, \code{CLINEAR}, \code{MEC} and \code{MEB}.
If \code{low = high} then the identity mapping
is used.}

\item{adjust.for.con.comp}{If TRUE (default), adjust some
of the models (currently: besag, bym, bym2 and besag2) if the
number of connected components in graph is larger than 1.
If FALSE, do nothing.}

\item{order}{Defines the \code{order} of the model: for
model \code{ar} this defines the order p, in AR(p). Not
used for other models at the time being.}

\item{scale}{A scaling vector. Its meaning depends on the model.}

\item{rgeneric}{A object of class \code{inla.rgeneric} which defines the model. (EXPERIMENTAL!)}

\item{cgeneric}{A object of class \code{inla.cgeneric} which defines the model. (EXPERIMENTAL!)}

\item{scale.model}{Logical. If \code{TRUE} then scale the RW1 and RW2 and BESAG and BYM and BESAG2 and RW2D models so the their (generlized) variance is 1. Default value is \code{inla.getOption("scale.model.default")}}

\item{args.slm}{Required arguments to the model="slm"; see the documentation for further details.}

\item{args.ar1c}{Required arguments to the model="ar1c"; see the documentation for further details.}

\item{args.intslope}{A list with the \code{subject} (factor),  \code{strata} (factor) and \code{covariates} (numeric) for the \code{intslope} model; see the documentation for further details,}

\item{vb.correct}{Add this model component to the list of nodes to be used for the (potential) vb correction?  If \code{TRUE} do,  and do not if \code{FALSE}. Can also be a vector of nodes to add in the correction-set.}

\item{locations}{A matrix with locations for the model \code{dmatern}. This also defines \code{n}.}

\item{debug}{Enable local debug output}
}
\value{
TODO
}
\description{
Function used for defining of smooth and spatial terms within \code{inla} model
formulae. The function does not evaluate anything - it
exists purely to help set up a model. The function specifies one
smooth function in the linear predictor (see \code{\link[=inla.list.models]{inla.list.models()}}) as
\deqn{w\ f(x)}{weight*f(var)}
}
\details{
There is no default value for \code{rankdef}, if it
is not defined by the user then it is computed by the rank
deficiency of the prior model (for the generic model, the
default is zero), plus 1 for the sum-to-zero constraint if the
prior model is proper, plus the number of extra
constraints. \strong{Oops:} This can be wrong, and then the user
must define the \code{rankdef} explicitly.
}
\seealso{
\code{\link[=inla]{inla()}}, \code{\link[=hyperpar.inla]{hyperpar.inla()}}
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
