%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from fgn.R
%%
 \name{fgn}
 \alias{inla.fgn}
 \alias{fgn}

 \title{Extract coefficients of the 3-component AR(1) mixture representing FGN}

 \description{Extract coefficients of the 3-component AR(1) mixture representing FGN}

 \usage{
     inla.fgn(H)
 } 

 \arguments{
   \item{H}{The Hurst coeffcient (0<H<1)}
  }
 \value{
  A named vector of length 7, where the first element is \code{H},
  the next three are the lag one
  correlations (or phi's),
  and the last three are the weights.

  This function is EXPERIMENTAL!!!
 }
 \author{Havard Rue \email{hrue@r-inla.org}}

 \examples{
     inla.fgn(0.7)
 } 
