%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from fgn.R
%%
 \name{fgn}
 \alias{inla.fgn}
 \alias{fgn}

 \title{Lookup coefficients in the 3-component AR(1) mixture representing FGN(H)}

 \description{This function will lookup the coefficients in the 3-component AR(1)
              mixture representing FGN(H)}

 \usage{
     inla.fgn(H, K=3L)
 } 

 \arguments{
   \item{H}{The Hurst coeffcient (0<H<1),  or a vector of those}
   \item{K}{The number of components in representation,  must be 3L or 4L}
  }
 \value{
  A named matrix with 7 columns, where the first column is \code{H},
  column 2,  3 and 4 are lag one correlations (or phi's),
  and column 5,  6 and 7 are the weights.

  This function is EXPERIMENTAL!!!
 }
 \author{Havard Rue \email{hrue@r-inla.org}}

 \examples{
     r = c(inla.fgn(0.7))
     r_m = inla.fgn(seq(0.6, 0.8, by=0.01))
 } 
