\name{Germany}
\alias{Germany}
\docType{data}
\title{ Disease Mapping}
\description{
Cases of Oral cavity cancer in Germany from 1986-1990
}
\usage{data(Germany)}
\format{
  A data frame with 544 observations on the following 4 variables.
  \describe{
    \item{\code{region}}{Region of Germany}
    \item{\code{E}}{Fixed quantity which accounts for number of
        people in the district (offset)}
    \item{\code{Y}}{Number of cases}
    \item{\code{x}}{covariate measuring smoking consumption}
  }
}


\references{
 Rue, H and Held, L. (2005)
  \emph{Gaussian Markov Random Fields - Theory and Applications}
  Chapman and Hall
  

}
\examples{
data(Germany)

}
\keyword{datasets}
