% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.matrix.R
\name{graph.matrix}
\alias{graph.matrix}
\alias{inla.matrix2graph}
\alias{graph2matrix}
\alias{inla.graph2matrix}
\alias{spy}
\alias{inla.spy}
\title{Construct a neighbour-matrix from a \code{graph}}
\usage{
inla.matrix2graph(graph, ...)

inla.graph2matrix(graph, ...)

inla.spy(graph, ..., reordering = NULL)
}
\arguments{
\item{graph}{An \code{inla.graph}-object, a (sparse) symmetric matrix, a
filename containing the graph, or a list or collection of characters and/or
numbers defining the graph.}

\item{...}{Additional arguments to \code{inla.read.graph()}}

\item{reordering}{A possible reordering. Typical the one obtained from a
\code{inla}-call, \code{result$misc$reordering}, or the result of
\code{inla.qreordering}.}
}
\value{
\code{inla.graph2matrix} returns a sparse symmetric matrix where the
non-zero pattern is defined by the \code{graph}.  The \code{inla.spy}
function, plots the associated sparse matrix defined by the \code{graph}. The \code{reordering}
argument is typically the reordering found by \code{inla.qreordering()}.
}
\description{
Construct a neighbour-matrix from a \code{graph} and disaply it
}
\examples{

n = 50
Q = matrix(0, n, n)
idx = sample(1:n, 2*n, replace=TRUE)
Q[idx, idx] = 1
diag(Q) = 1
g = inla.read.graph(Q)
QQ = inla.graph2matrix(g)
inla.spy(QQ)
print(all.equal(as.matrix(Q), as.matrix(QQ)))

g.file = inla.write.graph(g, filename = tempfile())
inla.dev.new()
inla.spy(g.file)
inla.spy(g.file,  reordering = inla.qreordering(g))

g = inla.read.graph(g.file)
inla.dev.new()
inla.spy(g)

## Old examples that don't work with the inla.spy call syntax:
# inla.dev.new()
# inla.spy(3, 1, "1 2 2 1 1 3 0")
# inla.dev.new()
# inla.spy(3, 1, "1 2 2 1 1 3 0", reordering = 3:1)

}
\seealso{
\code{\link[=inla.read.graph]{inla.read.graph()}}, \code{\link[=inla.qreordering]{inla.qreordering()}}
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
