% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.cv.R
\name{inla.group.cv}
\alias{inla.group.cv}
\title{Compute group.cv-values}
\usage{
inla.group.cv(
  result,
  group.cv = NULL,
  num.level.sets = -1,
  strategy = c("posterior", "prior"),
  size.max = 32,
  groups = NULL,
  selection = NULL,
  group.selection = NULL,
  friends = NULL,
  weights = NULL,
  verbose = FALSE,
  epsilon = 0.005,
  prior.diagonal = 1e-04,
  build.radius = 1,
  keep = NULL,
  remove = NULL,
  remove.fixed = TRUE,
  type.cv = "single"
)
}
\arguments{
\item{result}{An object of class \code{inla}, ie a result of a call to
\code{inla()}.}

\item{group.cv}{If given, the groups are taken from this argument.
\code{group.cv} must be the output of previous call to
\code{inla.group.cv()}.}

\item{num.level.sets}{Number of level.sets to use. The default value
\code{-1} corresponds to leave-one-out cross-validation. If argument \code{weights} is
used, then this is threshold for the sum of the weights defining a group.}

\item{strategy}{One of \code{"posterior"} or \code{"prior"}. See the
vignette for details.}

\item{size.max}{The maximum size (measure in the number of nodes) of a group. If the computed
group-size is larger, it will be truncated to \code{size.max}. Note that: If
\code{weights} are in use, then
this still corresponds to the number of nodes in the group, and not the sum of the
weights. This is ment as an emergency option to avoid the size of the
group to go nuts.}

\item{groups}{An (optional) predefined list of groups.  See the vignette for
details.}

\item{selection}{An optional list of data-indices to use. If not given, then
all data are used.}

\item{group.selection}{An optional list of data-indices to use when building the
groups. If given, each group beyond the observation itself, must be a subset of
\code{group.selection}. If not given, then all data are used.}

\item{friends}{An optional list of lists of indices to use a friends}

\item{weights}{An optional positive weight attached to each datapoint. The sum
of the weights define the size of the group. If \code{NULL}, then unit weight is
used.}

\item{verbose}{Run with \code{verbose} output of some of the internals in
the calculations. This option will also enable \code{inla(..., verbose=TRUE)} if its not enabled already.}

\item{epsilon}{Two correlations with a difference less than \code{epsilon},
will be classified as identical.}

\item{prior.diagonal}{When \code{strategy="prior"}, \code{prior.diagonal} is
added to the diagonal of the prior precision matrix to avoid singularities}

\item{build.radius}{Radius, in hops on the latent graph, of the local
candidate search used to build the groups: the groups are formed from
the candidates within the radius and certified complete by a separator
bound, and nodes that cannot be certified fall back to the exact full
solve. \code{0} disables the local build (always the full solve).}

\item{keep}{For \code{strategy="prior"}, then this gives a vector of the
name of model-components TO USE when computing the groups. See the vignette
for details. Not both of \code{keep} and \code{remove} can be defined.}

\item{remove}{For \code{strategy="prior"}, then this gives a vector of the
name of model-components NOT TO USE when computing the groups. See the
vignette for details. Not both of \code{keep} and \code{remove} can be
defined.}

\item{remove.fixed}{For \code{strategy="prior"}, this is the default option
which is in effect if both \code{keep} and \code{remove} are \code{NULL}. If
\code{TRUE}, it will remove (or condition on) all fixed effects when
computing the groups. See the vignette for details.}

\item{type.cv}{Type of cv, either "single" (default) or "joint"}
}
\value{
The object returned is list related to leave-group-out
cross-validation. See the vignette for details.
}
\description{
From a fitted model, compute and add the \code{group.cv}-values
}
\seealso{
\code{\link[=control.compute]{control.compute()}}
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
