%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from group.cv.R
%%
\name{inla.group.cv}
\alias{inla.group.cv}

\title{Compute group.cv-values}
\description{
 From a fitted model, compute and add the \code{group.cv}-values
}
\usage{
inla.group.cv(result,
          num.level.sets = -1, 
          size.max = 32, 
          strategy = c("posterior", "prior"), 
          groups = NULL, 
          selection = NULL, 
          friends = NULL, 
          verbose = FALSE, 
          epsilon = 0.0025, 
          prior.diagonal = 1e-04, 
          keep = NULL, 
          remove = NULL, 
          remove.fixed = TRUE)
}
\arguments{
\item{result}{An object of class \code{inla}, ie a result
of a call to \code{inla()}.}
\item{num.level.sets}{Number of level.sets to use. The default value
\code{-1} corresponds to leave-one-out cross-validation.}
\item{strategy}{One of \code{"posterior"} or \code{"prior"}.
See the  vignette for details.}
\item{size.max}{The maximum size of a group. If the computed
group-size is larger, it will be truncated to \code{size.max}.}
\item{groups}{An (optional) predefined list of groups. 
See the vignette for details.}
\item{selection}{An optional list of data-indices to use.
If not given, then all data are used.}
\item{friends}{An optional list of lists of indices to use a friends}
\item{verbose}{Run with \code{verbose} output of some of the internals
in the  calculations. This option will also
enable \code{inla(...,  verbose=TRUE)} if its not enabled already.}
\item{epsilon}{Two correlations with a difference less than
\code{epsilon}, will be classified as identical.}
\item{prior.diagonal}{When \code{strategy="prior"}, \code{prior.diagonal}
is added to the diagonal of the prior precision matrix to avoid singularities}
\item{keep}{For \code{strategy="prior"}, then this gives a vector of 
the name of model-components TO USE when computing the groups.
See the  vignette for details.
Not both of \code{keep} and \code{remove} can be defined.}
\item{remove}{For \code{strategy="prior"}, then this gives a vector of 
the name of model-components NOT TO USE when computing the groups.
See the  vignette for details.
Not both of \code{keep} and \code{remove} can be defined.}
\item{remove.fixed}{For \code{strategy="prior"}, this is the default
option which is in effect if both \code{keep} and \code{remove} are 
\code{NULL}. If \code{TRUE}, it will remove (or condition on) all fixed effects
when computing the groups.
See the  vignette for details.}
}
\value{The object returned is list related to leave-group-out cross-validation. See the vignette for details.}
\author{Havard Rue \email{hrue@r-inla.org}}
\seealso{\code{\link{control.compute}}}
