% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx.R
\name{idx}
\alias{idx}
\alias{inla.idx}
\title{Convert indexes}
\usage{
inla.idx(
  idx,
  n = max(idx),
  group = rep(1, length(idx)),
  ngroup = max(group),
  replicate = rep(1, length(idx)),
  nrep = max(replicate)
)
}
\arguments{
\item{idx}{The index within the basic model. (Legal values from \verb{1' to }n'.)}

\item{n}{The length `n' of the basic model.}

\item{group}{The index within group. (Legal values from \verb{1' to }ngroup'.)}

\item{ngroup}{Number of groups.}

\item{replicate}{The index within replication. (Legal values from \verb{1' to }nrep'.)}

\item{nrep}{Number of replications.}
}
\value{
\code{inla.idx} returns indexes in the range \verb{1' to }n\emph{ngroup}nrep' representing where the triplet `(idx,group,replicate)' is
stored internally in the full grouped and replicated model.
}
\description{
Convert indexes given by to triplet `(idx, group, replicate)' to the
(one-dimensional) index used in the grouped and replicated model
}
\examples{

##TODO

}
\author{
Havard Rue \email{hrue@r-inla.org}
}
