% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iidkd.R
\name{inla.iidkd.sample}
\alias{inla.iidkd.sample}
\alias{iidkd.sample}
\title{Provide samples from the \code{iidkd} component (experimental)}
\usage{
inla.iidkd.sample(n = 10^4, result, name, return.cov = FALSE)
}
\arguments{
\item{n}{Integer Number of samples to use}

\item{result}{inla-object An object of class \code{inla}, ie a result of a
call to \code{inla()}}

\item{name}{Character The name of the \code{iidkd} component}

\item{return.cov}{Logical Return samples of the covariance matrix instead of
stdev/correlation matrix described below?}
}
\value{
A list of sampled matrices, with (default) correlations on the
off-diagonal and standard-deviations on the diagonal
}
\description{
This function provide samples of the \code{iidkd} component using more
interpretable parameters
}
\seealso{
\code{inla.doc("iidkd")}
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
