% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{inla-control}
\alias{inla-control}
\alias{ctrl_class}
\alias{ctrl_type}
\alias{ctrl_type.default}
\alias{ctrl_type.character}
\alias{ctrl_check}
\alias{ctrl_object}
\alias{ctrl_default}
\alias{ctrl_default.character}
\alias{ctrl_default.inla_ctrl_object}
\alias{ctrl_update}
\alias{ctrl_update.inla_ctrl_object}
\alias{ctrl_update.ctrl_family}
\alias{ctrl_update.ctrl_mix}
\alias{ctrl_update.ctrl_link}
\alias{ctrl_update.ctrl_lp_scale}
\alias{ctrl_update.ctrl_predictor}
\alias{ctrl_update.ctrl_compute}
\alias{ctrl_update.ctrl_group}
\alias{ctrl_update.ctrl_scopy}
\alias{ctrl_update.ctrl_hazard}
\alias{ctrl_update.ctrl_mode}
\title{inla control argument handlers}
\usage{
ctrl_class(x)

ctrl_type(x)

\method{ctrl_type}{default}(x)

\method{ctrl_type}{character}(x)

ctrl_check(x, the_type, default = NULL)

ctrl_object(x, the_type, data = NULL, check = TRUE)

ctrl_default(x)

\method{ctrl_default}{character}(x)

\method{ctrl_default}{inla_ctrl_object}(x)

ctrl_update(x, ...)

\method{ctrl_update}{inla_ctrl_object}(x, ..., default = NULL)

\method{ctrl_update}{ctrl_family}(x, ..., model)

\method{ctrl_update}{ctrl_mix}(x, ...)

\method{ctrl_update}{ctrl_link}(x, ...)

\method{ctrl_update}{ctrl_lp_scale}(x, ...)

\method{ctrl_update}{ctrl_predictor}(x, ..., control.compute, control.inla)

\method{ctrl_update}{ctrl_compute}(x, ...)

\method{ctrl_update}{ctrl_group}(x, ...)

\method{ctrl_update}{ctrl_scopy}(x, ...)

\method{ctrl_update}{ctrl_hazard}(x, ...)

\method{ctrl_update}{ctrl_mode}(x, ...)
}
\arguments{
\item{default}{A default object of matching \verb{ctrl_*} class.
If \code{NULL}, uses \code{ctrl_default(x)} instead.}

\item{data}{An optional data object (\code{data.frame}, \code{list}, or \code{environment},
see the \code{what} argument of \code{\link[=base:attach]{base:attach()}}).}

\item{check}{logical; If \code{TRUE} (default), calls \code{ctrl_check()} to verify
compliance with the default object of the target control type. Must be set
to \code{FALSE} by code calling \code{ctrl_object()} to construct default objects.}

\item{model}{character; a model specifier}

\item{control.compute}{A \code{ctrl_compute} object.}

\item{control.inla}{A \code{ctrl_inla} object.}

\item{.envir}{The enclosing parent frame for resolving variables in \code{x}}
}
\description{
Methods for controlling \code{\link[=inla]{inla()}} and \code{\link[=f]{f()}} \code{control} arguments
}
\section{Functions}{
\itemize{
\item \code{ctrl_type()}: Returns the type-part of the control class name, i.e.
without the "ctrl_" prefix.

\item \code{ctrl_check()}: Checks that the elements are named and match
the available ones in the corresponding default control object.

\item \code{ctrl_object()}: Construct a control object and check that it's compatible
with the corresponding defaults.

\item \code{ctrl_update(inla_ctrl_object)}: Merges the \code{x} object with the defaults for control
type \code{ctrl_type(x)}. Recursively handles \verb{control.*} elements, with \code{NULL}
defaults also being replaced by their corresponding defaults before merging.

\item \code{ctrl_update(ctrl_predictor)}: Requires \code{control.compute} and \code{control.inla} to be provided,
in order to determine if \code{compute} needs to be forced to \code{TRUE}. To ignore,
use \code{ctrl_update(x, control.compute = list(), control.inla = list())}.

\item \code{ctrl_update(ctrl_compute)}: If \code{residuals} is \code{TRUE}, also sets \code{dic = TRUE}.

}}
\keyword{internal}
