\name{INLA-package}
\alias{INLA-package}
\alias{INLA}
\docType{package}
\title{
Integrated Nested Laplace Approximation 
}

\description{ Package to perform full Bayesian analysis on generalised
additive mixed models using Integrated Nested Laplace Approximations.
}


\details{
\tabular{ll}{
Package: \tab INLA\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab TODAY\cr
License: \tab GPL\cr
NOTE: \tab This package has no version number yet; it is to heavily developed
at the moment; see \code{inla.googlecode.com} and \code{www.r-inla.org}
}

Please see the web-site \cite{www.r-inla.org} for further details.

} 
%%

\author{Håvard Rue, Finn Lindgren and Sara Martino.
}

\references{
  Rue, H. and Martino, S. and Chopin, N. (2009)
  \emph{Approximate Bayesian Inference for latent Gaussian models using
  Integrated Nested Laplace Approximations}, Journal of Royal
  Statistical Society, Series B, with discussion. (To appear.)

  Rue, H and Held, L. (2005) \emph{Gaussian Markov Random Fields -
  Theory and Applications}, Monographs on Statistics and Applied
  Probability, no 104, Chapman \& Hall, London.
  
  F. Lindgren, H. Rue, and J.Lindstrom.
  \emph{An explicit link between Gaussian fields and Gaussian Markov
  random fields: The SPDE approach (with discussion).}
  Journal of the Royal Statistical Society, Series B,
  73 (4): 423--498, 2011.
}

\keyword{}
