% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{package}
\name{INLA-package}
\alias{INLA-package}
\alias{INLA}
\alias{R-INLA}
\title{Integrated Nested Laplace Approximation}
\description{
Package to perform full Bayesian analysis on generalised additive mixed
models using Integrated Nested Laplace Approximations.
}
\details{
\describe{
\item{Package}{INLA}
\item{Version}{Currently, this package uses a YY.MM.DD versioning
system, and is in heavy development.
See \url{https://github.com/hrue/r-inla/} and
\url{https://www.r-inla.org}}
\item{License}{GPL2}
}

See the web-site \url{https://www.r-inla.org} for further details.
}
\author{
Havard Rue, Sara Martino, Finn Lindgren, ...
}
