\name{INLA-package}
\alias{INLA-package}
\alias{INLA}
\docType{package}
\title{
Integrated Nested Laplace Approximation 
}

\description{ Package to perform full Bayesian analysis on generalised
additive mixed models using Integrated Nested Laplace Approximations.
}


\details{
\tabular{ll}{
Package: \tab INLA\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab TODAY\cr
License: \tab GPL2\cr
NOTE: \tab This package has no version number yet; it is to heavily developed
at the moment; see \code{bitbucket.org/hrue/r-inla} and \code{www.r-inla.org}
}

See the web-site \cite{www.r-inla.org} for further details.
} 
%%

\author{Havard Rue, Sara Martino, Finn Lindgren, Daniel Simpson and
  Andrea Riebler}

%%\references{}

%%\keyword{}
