% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmesh.R
\name{inla.CRS}
\alias{inla.CRS}
\alias{inla.wkt_predef}
\title{Create a coordinate reference system object}
\usage{
inla.CRS(...)

inla.wkt_predef()
}
\arguments{
\item{\dots}{Arguments passed on to \code{inlabru::fm_CRS(...)}.}
}
\value{
Either an \code{sp::CRS} object or an \code{inla.CRS} object,
depending on if the coordinate reference system described by the parameters
can be expressed with a pure \code{sp::CRS} object or not.

An S3 \code{inla.CRS} object is a list, usually (but not necessarily)
containing at least one element: \item{crs }{The basic \code{sp::CRS}
object}

\code{inla.wkt_predef} returns a WKT2 string defining a projection

\code{inla.wkt_predef} returns a WKT2 string defining a projection
}
\description{
Creates either a CRS object or an inla.CRS object, describing a coordinate
reference system. Deprecated in favour of \code{inlabru::fm_CRS()}.
}
\examples{


if (require("sf") && require("sp")) {
    crs1 <- inla.CRS("longlat_globe")
    crs2 <- inla.CRS("lambert_globe")
    crs3 <- inla.CRS("mollweide_norm")
    crs4 <- inla.CRS("hammer_globe")
    crs5 <- inla.CRS("sphere")
    crs6 <- inla.CRS("globe")
}
\dontrun{
names(inla.wkt_predef())
}

\dontrun{
names(inla.wkt_predef())
}
}
\seealso{
\code{\link[sp:CRS-class]{sp::CRS()}}, \code{\link[=crs_wkt]{crs_wkt()}},
\code{\link[=inla.sp_get_crs]{inla.sp_get_crs()}} \code{\link[=plot.CRS]{plot.CRS()}},
\code{\link[=inla.identical.CRS]{inla.identical.CRS()}}
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
