% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmesh.R
\name{inla.CRS}
\alias{inla.CRS}
\alias{inla.wkt_predef}
\title{Create a coordinate reference system object}
\usage{
inla.CRS(..., args = NULL)

inla.wkt_predef()
}
\arguments{
\item{\dots}{Arguments passed on to \code{fmesher::fm_CRS(...)}.}

\item{args}{list of named proj4 arguments.}
}
\value{
Either an \code{sp::CRS} object or an \code{inla.CRS} object,
depending on if the coordinate reference system described by the parameters
can be expressed with a pure \code{sp::CRS} object or not.

An S3 \code{inla.CRS} object is a list, usually (but not necessarily)
containing at least one element: \item{crs }{The basic \code{sp::CRS}
object}

\code{inla.wkt_predef} returns a WKT2 string defining a projection

\code{inla.wkt_predef} returns a WKT2 string defining a projection
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} in favour of
\code{\link[fmesher:fm_CRS_sp]{fmesher::fm_CRS()}}

Creates either a CRS object or an inla.CRS object, describing a coordinate
reference system.
}
\section{Functions}{
\itemize{
\item \code{inla.wkt_predef()}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} in favour of
\code{\link[fmesher:fm_crs]{fmesher::fm_wkt_predef()}}

}}
\examples{
if (require("sf")) {
    crs1 <- fmesher::fm_crs("longlat_globe")
    crs2 <- fmesher::fm_crs("lambert_globe")
    crs3 <- fmesher::fm_crs("mollweide_norm")
    crs4 <- fmesher::fm_crs("hammer_globe")
    crs5 <- fmesher::fm_crs("sphere")
    crs6 <- fmesher::fm_crs("globe")
}
\dontrun{
names(inla.wkt_predef())
}

\dontrun{
names(inla.wkt_predef())
}
}
\seealso{
\code{\link[fmesher:fm_crs]{fmesher::fm_crs()}}, \code{\link[fmesher:fm_crs_wkt]{fmesher::fm_wkt()}},
\code{\link[fmesher:fm_crs_is_identical]{fmesher::fm_crs_is_identical()}}
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
