% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{inla.diameter}
\alias{inla.diameter}
\alias{inla.diameter.default}
\alias{inla.diameter.inla.mesh}
\alias{inla.diameter.inla.mesh.segment}
\alias{inla.diameter.inla.mesh.lattice}
\alias{inla.diameter.inla.mesh.1d}
\title{Diameter of a point set}
\usage{
inla.diameter(x, ...)
}
\arguments{
\item{x}{A point set as an \eqn{n\times d}{n x d} matrix, or an
\code{inla.mesh} related object.}

\item{\dots}{Additional parameters passed on to other methods.}

\item{manifold}{Character string specifying the manifold type. Default is to
treat the point set with Euclidean \eqn{R^d} metrics. Use
\code{manifold="S2"} for great circle distances on the unit sphere (this is
set automatically for \code{inla.mesh} objects).}
}
\value{
A scalar, upper bound for the diameter of the convex hull of the
point set.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[fmesher:fm_diameter]{fmesher::fm_diameter()}} instead.

Find an upper bound to the convex hull of a point set
}
\examples{

inla.diameter(matrix(c(0, 1, 1, 0, 0, 0, 1, 1), 4, 2))
}
\author{
Finn Lindgren \href{mailto:finn.lindgren@gmail.com}{finn.lindgren@gmail.com}
}
