% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmesh.R
\name{inla.identical.CRS}
\alias{inla.identical.CRS}
\title{Test CRS and inla.CRS for equality}
\usage{
inla.identical.CRS(...)
}
\description{
Wrapper for identical, optionally testing only the CRS part of two objects
Deprecated in favour of \code{inlabru::fm_identical_CRS()}
}
