\name{inla.matrix2graph}
\alias{inla.matrix2graph}
\title{
  These are old functions; please use \code{inla.read.graph} and \code{inla.write.graph}.
}
\description{
  These are old functions; please use \code{inla.read.graph} and \code{inla.write.graph}.
}
\usage{
}
\arguments{
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}

\seealso{
  \code{\link{inla.read.graph}},
  \code{\link{inla.write.graph}}
}
