% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{inla.mesh.1d.bary}
\alias{inla.mesh.1d.bary}
\alias{inla.mesh.1d.A}
\title{Mapping matrix for 1D meshes}
\usage{
inla.mesh.1d.bary(mesh, loc, method = c("linear", "nearest"))

inla.mesh.1d.A(mesh, loc, weights = NULL, derivatives = NULL, method = NULL)
}
\arguments{
\item{mesh}{An \code{\link[=inla.mesh.1d]{inla.mesh.1d()}} object.}

\item{loc}{Coordinate values.}

\item{method}{Interpolation method.  If not specified for
\code{inla.mesh.1d.A} (recommended), it is determined by the \code{mesh}
basis function properties.}

\item{weights}{Weights to be applied to the \code{A} matrix rows.}

\item{derivatives}{If \code{TRUE}, also compute derivative weight matrices
\code{dA} and \code{d2A}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[fmesher:fm_basis]{fmesher::fm_basis()}} instead.

Calculates barycentric coordinates and weight matrices for
\code{\link[=inla.mesh.1d]{inla.mesh.1d()}} objects.
}
\section{Functions}{
\itemize{
\item \code{inla.mesh.1d.bary()}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use \code{\link[fmesher:fm_bary]{fmesher::fm_bary()}} instead.

}}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
