% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{inla.mesh.create.helper}
\alias{inla.mesh.create.helper}
\title{High-quality triangulations}
\usage{
inla.mesh.create.helper(points = NULL, points.domain = NULL, ...)
}
\arguments{
\item{points}{Matrix of point locations to be used as initial triangulation
nodes.}

\item{points.domain}{Matrix of point locations used to determine the domain
extent.}

\item{\dots}{Additional arguments passed on to \code{\link[=inla.mesh.2d]{inla.mesh.2d()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[fmesher:fm_mesh_2d]{fmesher::fm_mesh_2d_inla()}} instead.

Create a triangle mesh based on initial point locations, specified or
automatic boundaries, and mesh quality parameters.
}
\note{
Obsolete function kept for backwards compatibility.  Use
\code{\link[fmesher:fm_mesh_2d]{fmesher::fm_mesh_2d_inla()}} instead.
}
\seealso{
\code{\link[=inla.mesh.2d]{inla.mesh.2d()}}
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
\keyword{internal}
