% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{inla.mesh.project}
\alias{inla.mesh.project}
\alias{inla.mesh.projector}
\alias{inla.mesh.projector.inla.mesh}
\alias{inla.mesh.project.inla.mesh}
\alias{inla.mesh.project.inla.mesh.projector}
\alias{inla.mesh.project.inla.mesh.1d}
\alias{inla.mesh.projector.inla.mesh.1d}
\title{Methods for projecting to/from an inla.mesh}
\usage{
inla.mesh.project(...)

inla.mesh.projector(...)
}
\arguments{
\item{\dots}{Arguments passed on to \code{\link[fmesher:fm_evaluate]{fmesher::fm_evaluate()}} and
\code{\link[fmesher:fm_evaluate]{fmesher::fm_evaluator()}}.}
}
\value{
For \code{inla.mesh.project(mesh, ...)}, a list with projection
information.  For \code{inla.mesh.projector(mesh, ...)}, an
\code{inla.mesh.projector} object.  For \code{inla.mesh.project(projector, field, ...)}, a field projected from the mesh onto the locations given by
the projector object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[fmesher:fm_evaluate]{fmesher::fm_evaluate()}} and
\code{\link[fmesher:fm_evaluate]{fmesher::fm_evaluator()}} instead.

Calculate a lattice projection to/from an \code{\link[=inla.mesh]{inla.mesh()}}.

The call \code{inla.mesh.project(mesh, loc, field=..., ...)}, is a shortcut
to \code{inla.mesh.project(inla.mesh.projector(mesh, loc), field)}.
}
\examples{

n <- 20
loc <- matrix(runif(n * 2), n, 2)
mesh <- inla.mesh.create(loc, refine = list(max.edge = 0.05))
proj <- inla.mesh.projector(mesh)
field <- cos(mesh$loc[, 1] * 2 * pi * 3) * sin(mesh$loc[, 2] * 2 * pi * 7)
image(proj$x, proj$y, inla.mesh.project(proj, field))
\donttest{
if (require(rgl)) {
  plot(mesh, rgl = TRUE, col = field, draw.edges = FALSE, draw.vertices = FALSE)
}
}

}
\seealso{
\code{\link[=inla.mesh]{inla.mesh()}}, \code{\link[=inla.mesh.1d]{inla.mesh.1d()}},
\code{\link[=inla.mesh.lattice]{inla.mesh.lattice()}}
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
