% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{inla.mesh.query}
\alias{inla.mesh.query}
\title{High-quality triangulations}
\usage{
inla.mesh.query(mesh, ...)
}
\arguments{
\item{mesh}{An \code{inla.mesh} object.}

\item{\dots}{Query arguments.
\itemize{
\item tt.neighbours Compute
neighbour triangles for triangles; list of vectors: list(triangles, orders)

\item vt.neighbours Compute neighbour triangles for vertices; list of
vectors: list(vertices, orders)
}}
}
\value{
A list of query results.
}
\description{
Query information about an inla.mesh object.
}
\examples{

loc <- matrix(c(0.1, 0.15), 1, 2)
lattice <- inla.mesh.lattice(
  seq(0, 1, length.out = 10),
  seq(0, 1, length.out = 10)
)
mesh <- inla.mesh.create(loc = loc, lattice = lattice, extend = FALSE)

vt <- which(inla.mesh.query(mesh,
  vt.neighbours = list(
    mesh$idx$loc,
    4:6
  )
)$vt.neighbours)

mesh2 <- inla.mesh.create(mesh$loc,
  tv = mesh$graph$tv[vt, , drop = FALSE],
  refine = FALSE, extend = FALSE
)
}
\seealso{
\code{\link[=inla.mesh.create]{inla.mesh.create()}}, \code{\link[=inla.mesh.segment]{inla.mesh.segment()}},
\code{\link[=inla.mesh.lattice]{inla.mesh.lattice()}}
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
