% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{inla.mesh.segment}
\alias{inla.mesh.segment}
\alias{inla.contour.segment}
\title{Constraint segments for inla.mesh}
\usage{
inla.mesh.segment(...)

inla.contour.segment(...)
}
\arguments{
\item{...}{Parameters passed on to \code{\link[fmesher:fm_segm]{fmesher::fm_segm()}} and other
replacement \code{fmesher} functions.}
}
\value{
An \code{fm_segm} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{\link[fmesher:fm_segm]{fmesher::fm_segm()}} instead.

Constructs \code{inla.mesh.segment} objects that can be used to specify
boundary and interior constraint edges in calls to \code{\link[=inla.mesh]{inla.mesh()}}.
}
\section{Functions}{
\itemize{
\item \code{inla.contour.segment()}: \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use
\code{\link[fmesher:fm_segm_contour_helper]{fmesher::fm_segm_contour_helper()}} instead.

}}
\examples{
require("fmesher")
## Create a square boundary and a diagonal interior segment
loc.bnd <- matrix(c(0, 0, 1, 0, 1, 1, 0, 1), 4, 2, byrow = TRUE)
loc.int <- matrix(c(0.9, 0.1, 0.1, 0.6), 2, 2, byrow = TRUE)
segm.bnd <- fm_segm(loc.bnd)
segm.int <- fm_segm(loc.int, is.bnd = FALSE)

## Points to be meshed
loc <- matrix(runif(10 * 2), 10, 2) * 0.9 + 0.05
mesh <- fm_rcdt_2d_inla(loc,
  boundary = segm.bnd,
  interior = segm.int,
  refine = list()
)
plot(mesh)

mesh <- fm_rcdt_2d_inla(loc, interior = fm_segm_join(segm.bnd, segm.int))
plot(mesh)

}
\seealso{
\code{\link[=inla.mesh.create]{inla.mesh.create()}}, \code{\link[=inla.mesh.2d]{inla.mesh.2d()}}
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
