% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\name{INLA}
\alias{INLA}
\title{INLA}
\usage{
INLA()
}
\description{
Package to perform full Bayesian analysis on latent Gaussian models using
Integrated Nested Laplace Approximations.

See r-inla.org for further details.
}
