% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmesh.R
\name{inla.spTransform}
\alias{inla.spTransform}
\title{Wrapper method for \code{inlabru::fm_transform}}
\usage{
inla.spTransform(x, CRSobj, ...)
}
\arguments{
\item{x}{The object that should be transformed from it's current CRS to a
new CRS}

\item{CRSobj}{The target crs object}

\item{\dots}{Potential additional arguments}

\item{crs0}{The source crs object, only needed for raw coordinate inputs}

\item{passthrough}{default FALSE. Setting to TRUE allows objects with no CRS
information to be passed through without transformation.}
}
\value{
The object is returned with its coordinates transformed
}
\description{
Handles transformation of various inla objects accorting to coordinate
reference systems of \code{sf::crs}, \code{sp::CRS} or \code{inla.CRS} class.
}
\examples{

if (require("sf") && require("sp") && require("inlabru")) {
    latt <- inla.mesh.lattice(-10:10, 40:60)
    mesh1 <- inla.mesh.create(
        lattice = latt, extend = FALSE, refine = FALSE,
        crs = fm_CRS("longlat_norm")
    )
    mesh2 <- fm_transform(mesh1, fm_crs("lambert_globe"))
    summary(mesh1)
    summary(mesh2)
}
}
\seealso{
\code{\link[=inla.CRS]{inla.CRS()}}
}
\author{
Finn Lindgren \href{mailto:finn.lindgren@gmail.com}{finn.lindgren@gmail.com}
}
