% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlaEnv.R
\name{inla.get.inlaEnv}
\alias{inla.get.inlaEnv}
\alias{get.inlaEnv}
\title{Return the internal environment used by INLA}
\usage{
inla.get.inlaEnv()
}
\value{
This function returns the internal environment used by INLA to
keep internal variables.
}
\description{
A function which return the internal environment used by INLA
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
