%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from jmarginal.R
%%
 \name{joint.marginal}
 \alias{inla.joint.marginal}
 \alias{joint.marginal}
 \alias{rjmarginal}
 \alias{inla.rjmarginal}

 \title{Sample from a joint marginal approximation}
 
 \description{Sample from from a joint marginal approximation
 as returned using argument \code{selection} in \code{inla}.
 }
 
 \usage{
 inla.rjmarginal(n, jmarginal)
 }
 \arguments{
   \item{n}{The number of samples}
   \item{jmarginal}{A marginal object given  either by a \code{inla} object
     or \code{result$selection}}
 }
 
 \value{%%
 A list with the samples in \code{samples} (matrix) and the corresponding log-densities
 in \code{log.density} (vector).
 Each column in \code{samples} contains one sample.
 }
 
 \author{Havard Rue \email{hrue@r-inla.org}}
 \seealso{\code{\link{inla}}}

 \examples{
 n = 10
 x = 1+rnorm(n)
 xx = 3 + rnorm(n)
 y = 1 + x + xx + rnorm(n)
 selection = list(xx=1, Predictor = 3:4, x=1)
 r = inla(y ~ 1 + x + xx,
          data = data.frame(y, x, xx),
          selection = selection)
 ns = 100
 xx = inla.rjmarginal(ns, r)
 
 print(cbind(mean = r$selection$mean, sample.mean = rowMeans(xx$samples)))
 print("cov matrix")
 print(round(r$selection$cov.matrix, dig=3))
 print("sample cov matrix")
 print(round(cov(t(xx$samples)), dig=3))
 
 skew = function(z) mean((z-mean(z))^3)/var(z)^1.5
 print(round(cbind(skew = r$selection$skewness,
                   sample.skew = apply(xx$sample, 1, skew)), dig=3))
}
 
