% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp.R
\name{jp}
\alias{jp}
\alias{inla.jp.define}
\alias{jp.define}
\title{Joint-prior models}
\usage{
inla.jp.define(jp = NULL, ...)
}
\arguments{
\item{jp}{The \code{jp}-function which returns the joint log-prior as a
function of argument \code{theta}. There is an optional second argument that
is a vector of \code{theta}-names. If second argument is not present,
argument \code{.theta.desc} will be added.}

\item{...}{Named list of variables that defines the environment of \code{jp}}
}
\value{
This allows joint priors to be defined in \code{R}.

This function is for internal use only.
}
\description{
A framework for defining joint priors in R
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
