%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from likelihood.R
%%
 \name{inla.likelihood}
 \alias{likelilhood}

 \title{Providing functions for sampling new data, evaluating pdf, cdf, and quantiles for new data.}

 \description{This function return function to compute the pdf,cdf,quantiles, or samples for new data
 using the likelihood from a inla-object.}
 \usage{
     inla.likelihood(type = c(d,r,p,q),args)
 }

 \arguments{
   \item{type}{The returned function type. The definition is similar to "rnorm","dnorm","pnorm",and "dnorm".}
   \item{args}{It is usually a return value from "inla.likelihood.parser", which specifies parameters, link function and transformation function of hyperparameters.}
}
\details{Details goes here}
\value{value goes here}
\author{Havard Rue \email{hrue@r-inla.org}}

\examples{
 }
