% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincomb.R
\name{make.lincomb}
\alias{make.lincomb}
\alias{inla.make.lincomb}
\alias{make.lincombs}
\alias{inla.make.lincombs}
\title{Create linear combinations}
\usage{
inla.make.lincomb(...)

inla.make.lincombs(...)
}
\arguments{
\item{...}{Arguments; see examples}
}
\value{
A structure to be passed on to \code{\link[=inla]{inla()}} argument
\code{lincomb}
}
\description{
Create a linear combination or several linear combinations, as input to
\verb{inla(..., lincomb = <lincomb>)}
}
\examples{

##See the worked out examples and description in the FAQ
##section on {www.r-inla.org}

}
\seealso{
TODO
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
