%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from lincomb.R
%%
\name{make.lincomb}
\alias{make.lincomb}
\alias{make.lincombs}
\alias{inla.make.lincomb}
\alias{inla.make.lincombs}
\title{Create linear combinations}
\description{Create a linear combination or several
linear combinations, as input to \code{inla(...,  lincomb = <lincomb>)}}
\usage{
inla.make.lincomb(...)
inla.make.lincombs(...)
}
\arguments{
   \item{...}{Arguments; see examples}
}
\value{
 A structure to be passed on to \code{\link{inla}} argument \code{lincomb} 
}
\author{Havard Rue \email{hrue@r-inla.org}}
\seealso{TODO}
\examples{
##See the worked out examples and description in the FAQ
##section on {www.r-inla.org}
}
