%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from list-models.R
%%
\name{inla.list.models}
\alias{inla.list.models}
\alias{list.models}
\title{List available model components, likelihoods,  priors,  etc}
\description{
    List available model components, likelihoods,  priors,  etc.
}
\usage{
    inla.list.models(section = names(inla.models()), ...)
}
\arguments{
   \item{section}{The section(s) to list, missing \code{section} will list all sections.
                  \code{names(inla.models())} lists available sections.}
   \item{...}{Additional argument to \code{cat}}
}
\details{
   The list is \code{cat}'ed with \code{...} arguments.

   This function is EXPERIMENTAL.
}
\value{
   Nothing is returned
}
\author{Havard Rue} 
\examples{
\dontrun{
inla.list.models("likelihood")
inla.list.models(c("prior", "group"))
inla.list.models(file=file("everything.txt"))
}
}
