%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from load.R
%%
 \name{inla.load}
 \alias{inla.load}
 \title{Load or source a file}
 \description{Load or source a file: (internal use)}
 \usage{
     inla.load(filename, debug = TRUE)
 }
 \arguments{
   \item{filename}{The name of the file to be loaded, alternatively, sourced. }
   \item{debug}{Logical. Turn on/off debug information.}
  }
 \value{
   None
 }
 \details{
   Try to \code{load} the file into the global environment,
   if that fail,  try to \code{source} the file into the
   global environment.
 }    
 \author{Havard Rue \email{hrue@r-inla.org}}
