% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdata.R
\name{inla.mdata}
\alias{inla.mdata}
\alias{mdata}
\alias{is.inla.mdata}
\alias{as.inla.mdata}
\alias{print.inla.mdata}
\title{Create an mdata-object for INLA}
\usage{
inla.mdata(y, ...)

\method{print}{inla.mdata}(object, ...)

as.inla.mdata(object)

is.inla.mdata(object)
}
\arguments{
\item{y}{The response vector/matrix}

\item{...}{Additional vectors/matrics of same length as \code{y}}

\item{object}{Any \code{R}-object}

\item{x}{An mdata object}
}
\value{
An object of class \code{inla.mdata}.  There is method for
\code{print}.

\code{is.inla.mdata} returns \code{TRUE} if \code{object} inherits from
class \code{inla.mdata}, otherwise \code{FALSE}.

\code{as.inla.mdata} returns an object of class \code{inla.mdata}
}
\description{
This defines an mdata-object for matrix valued response-families
}
\note{
It is often required to set \code{Y=inla.mdata(...)} and then define
the formula as \code{Y~...}, especially when used with \code{inla.stack}.
}
\seealso{
\code{\link{inla}}
}
\author{
Havard Rue
}
