%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from mdata.R
%%
\name{inla.mdata}
\alias{inla.mdata}
\alias{is.inla.mdata}
\alias{as.inla.mdata}
\alias{print.inla.mdata}

\title{
Create an mdata-object for INLA
}

\description{
This defines an mdata-object for matrix valued response-families
}
\usage{
inla.mdata(y, ...)
is.inla.mdata(object)
as.inla.mdata(object)
}

\arguments{
  \item{y}{The response vector/matrix}
  \item{...}{Additional vectors/matrics of same length as \code{y}}
  \item{object}{Any \code{R}-object}
  \item{x}{An mdata object}
  }

\value{
     An object of class \code{inla.mdata}.  There is method for \code{print}.

     \code{is.inla.mdata} returns \code{TRUE} if \code{object}
     inherits from class \code{inla.mdata}, otherwise \code{FALSE}.

     \code{as.inla.mdata} returns an object of class \code{inla.mdata}
}

\author{
 Havard Rue
}

\seealso{
\code{\link{inla}}
}
