%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from merge.R
%%
 \name{inla.merge}
 \alias{inla.merge}
 \title{Merge a mixture of \code{inla}-objects}
 \description{Merge a mixture of \code{inla}-objects}
 \usage{
     inla.merge(loo, prob = rep(1,  length(loo)), verbose = FALSE)
 }
 \arguments{
   \item{loo}{A list of \code{inla}-objects to be merged}
   \item{prob}{The mixture of (possibly unnormalized) probabilities}
   \item{verbose}{Turn on verbose-output or not}
  }
 \value{
   A merge \code{inla}-object.
 }
 \details{
    This function is intented for merging a mixture of \code{inla}-object,
    each run with the same formula and settings,  except for a set of
    hyperparameters that are fixed to different values. Using this function,
    we can then integrate over these hyperparameters using integration weights
    \code{prob}. Not all entries in the object can be merged, so they are
    inheritated from the first object in the list. Those objectes that are merged,
    can be seen setting \code{verbose=TRUE}.
  
    Note that merging hyperparameter in the user-scale can be prone to
    discretization error, so more stable results can be found converting
    the marginal of the hyperparameter in the internal scale to the user-scale.
 }    
 \author{Havard Rue \email{hrue@r-inla.org}}
