% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshassessment.R
\name{inla.mesh.assessment}
\alias{inla.mesh.assessment}
\title{Interactive mesh building and diagnostics}
\usage{
inla.mesh.assessment(mesh, spatial.range, alpha = 2, dims = c(500, 500))
}
\arguments{
\item{mesh}{An \code{inla.mesh}}

\item{spatial.range}{numeric; the spatial range parameter to use for the assessment}

\item{alpha}{numeric; A valid \code{inla.spde2.pcmatern} \code{alpha} parameter}

\item{dims;}{2-numeric; the grid size}
}
\description{
Assess the finite element approximation errors in a mesh for interactive R
sessions.  More detailed assessment tools are in \code{\link[=meshbuilder]{meshbuilder()}}.
}
\examples{

bnd <- inla.mesh.segment(cbind(
    c(0, 10, 10, 0, 0),
    c(0, 0, 10, 10, 0)
), bnd = TRUE)
mesh <- inla.mesh.2d(boundary = bnd, max.edge = 1)
out <- inla.mesh.assessment(mesh, spatial.range = 3, alpha = 2)

}
\seealso{
inla.mesh.2d, inla.mesh.create, meshbuilder
}
\author{
Finn Lindgren \email{finn.lindgren@gmail.com}
}
