%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from meshbuilder.R
%%
\name{meshbuider}
\alias{meshbuilder}
\title{Interactive mesh building and diagnostics}
\description{Interactively design and build a triangle mesh
  for use with SPDE models, and assess the finite element
  approximation errors. The R code needed to recreate the mesh
  outside the interactive Shiny app is also generated. Spatial
  objects can be imported from the global workspace. }
\usage{
meshbuilder()
}
\author{Finn Lindgren \email{finn.lindgren@gmail.com}}
\seealso{inla.mesh.2d, inla.mesh.create}
\examples{
\dontrun{
meshbuilder()
}
}
