\name{Munich}
\alias{Munich}
\docType{data}
\title{The Munich rent data}
\description{
  The Munich rent data
}
\usage{data(Munich)}
\format{
  A data frame with 2035 observations on the following 17 variables.
  \describe{

\item{\code{rent}}{Net rent per square meter.}
\item{\code{floor.size}}{Size of the flat in square meters.}
\item{\code{year}}{Year of construction of the building in which the flat is located.}
\item{\code{location}}{Location index (in terms of subquarters).}
\item{\code{Gute.Wohnlage}}{Dummy variable for good locations / good neighborhoods.}
\item{\code{Beste.Wohnlage}}{Dummy variable for very good locations / very good neighborhoods.}
\item{\code{Keine.Wwv}}{Dummy for absence of warm water supply.}
\item{\code{Keine.Zh}}{Dummy for absence of central heating system.}
\item{\code{Kein.Badkach}}{Dummy for absence of flagging in the bathroom.}
\item{\code{Besond.Bad}}{Dummy for special features of the bathroom.}
\item{\code{Gehobene.Kueche}}{Dummy for more refined kitchen equipment.}
\item{\code{zim1}}{Dummy for a flat with 1 room.}
\item{\code{zim2}}{Dummy for a flat with 2 rooms.}
\item{\code{zim3}}{Dummy for a flat with 3 rooms.}
\item{\code{zim4}}{Dummy for a flat with 4 rooms.}
\item{\code{zim5}}{Dummy for a flat with 5 rooms.}
\item{\code{zim6}}{Dummy for a flat with 6 rooms.}
  }
}

\source{
  See Rue and Held (2005), Chapter 4.
}
\references{
  Rue, H and Held, L. (2005) \emph{Gaussian Markov Random
    Fields - Theory and Applications} Chapman and Hall
}

\keyword{datasets}
