% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA-package.R
\docType{data}
\name{nwEngland}
\alias{nwEngland}
\title{The North West England map}
\format{
A SpatialPolygons object.
}
\source{
This map are used to analyse the Leukaemia dataset from

Henderson, R. and Shimakura, S. and Gorst, D., 2002, Modeling spatial
variation in leukemia survival data, JASA, 97, 460, 965--972.
}
\description{
This map is used in association to the Leukemia data from Henderson et al
(2003); see source.
}
\examples{

data(Leuk)
plot(nwEngland)
}
\keyword{datasets}
