% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{inla.option}
\alias{inla.option}
\alias{inla.options}
\alias{inla.setOption}
\alias{inla.getOption}
\title{Set and get global options for INLA}
\usage{
inla.getOption(option = NULL, list.out = NULL)

inla.setOption(...)
}
\arguments{
\item{option}{The option to get. If \code{option = NULL} then
\code{inla.getOption} will return a named list of
current values, otherwise, \code{option} must be one (or a vector of several) of
\describe{
\item{inla.call}{The path to the inla-program.}

\item{inla.arg}{Additional arguments to \code{inla.call}}

\item{num.threads}{Character string with the number of threads to use as
\code{A:B:C}, see \code{?inla}}

\item{blas.num.threads}{If set to integer '> 0',  then this fix the number of threads for blas
functions. If set to 0,  the the number of threads will be determined adapatively. This value
is overrided if the value is set in 'control.expert',  hence 'blas.num.threads' provides the
default value for 'control.expert$blas.num.threads'. }

\item{smtp}{Sparse matrix library to use, one of \code{band}, \code{taucs}, \code{stiles}
or \code{default}}

\item{safe}{Run in safe-mode (ie try to automatically fix convergence errors)
(default \code{TRUE})}

\item{keep}{Keep temporary files?}

\item{verbose}{Verbose output?}

\item{save.memory}{Save memory at the cost of (minor) accuracy and computing time?}

\item{internal.opt}{Do internal online optimisations or not}

\item{working.directory}{The name of the working directory.}

\item{silent}{Run the inla-program in a silent mode?}

\item{debug}{Run the inla-program in a debug mode?}

\item{cygwin}{The home of the Cygwin installation (default "C:/cygwin") (Remote
computing for Windows only) (No longer in use!)}

\item{ssh.auth.sock}{The ssh bind-adress (value of $SSH_AUTH_SOCK int the
Cygwin-shell). (Remote computing for Windows only)}

\item{show.warning.graph.file}{Give a warning for using the obsolete argument
\code{graph.file} instead of \code{graph}}

\item{scale.model.default}{The default value of argument \code{scale.model} which
optionally scale intrinisic models to have generalized unit average variance}

\item{short.summary}{Use a less verbose output for \code{summary}. Useful for
Markdown documents.}

\item{inla.timeout}{The timeout limit, in whole seconds, for calls to the inla
binary. Default is 0, meaning no timeout limit.  Set to a positive integer
to terminate inla calls if they run to long.  Fractional seconds are rounded
up to the nearest integer. This feature is EXPERIMENTAL and might change at
a later stage.}

\item{inla.mode}{Which mode to use in INLA? Default is \code{"compact"}. Other
options are \code{"classic"} and \code{"twostage"}.}

\item{malloc.lib}{Which malloc library to use: \code{"je"}, \code{"tc"}, \code{"mi"}, \code{"compiler"} or \code{"default"}.
Option \code{"compiler"} use the compiler's implementation. The library
is loaded using \code{LD_PRELOAD} and similar functionality. Loosely, \code{jemalloc} is from Facebook,
\code{tcmalloc} is from Google and \code{mimalloc} is from Microsoft. This option is not available for
Windows and not all options might be available for every arch.
If \code{malloc.lib} is a complete path to an external library, that file will be used
instead of one of the supported ones.}

\item{fmesher.evolution}{Control use of fmesher methods during the transition
to a separate fmesher package. Only supported value is \code{2L}, for using the
full range of \code{fmesher} package methods. A higher value may be used in the future
to force use of newer methods, but this is not currently implemented.
Default is \code{2L}.}

\item{fmesher.evolution.warn}{logical; whether to show warnings about deprecated
use of legacy INLA methods with fmesher package replacements. When \code{TRUE},
shows deprecation messages for many CRS and mesh
related methods, pointing to their \verb{fm_*} replacements. Default
since July 2025 is \code{TRUE}.}

\item{fmesher.evolution.verbosity}{logical or character; at what minimum
severity to show warnings about deprecated
use of legacy INLA methods with fmesher package replacements.
When set to "default" (default), "soft", "warn", or "stop", indicates the
minimum warning level used when \code{fmesher.evolution.warn} is \code{TRUE}.}

\item{INLAjoint.features}{logical Do not use. By purpose left undocumented}

\item{numa}{logical Enable NUMA features (Linux only)}

\item{disable.values.warning}{logical Disable warning about missing 'values'}
}}

\item{list.out}{logical; If \code{TRUE}, then \code{inla.getOption} will return a named
list of values, even if only one option is requested.  Default is \code{NULL},
which means that a named list will be returned if more than one option is
requested, and a single value will be returned if only one option is
requested. Using \code{list.out = FALSE} will return a plain vector of values,
without names, even if more than one option is requested; the data type may
depend on which options are requested, and may be unpredictable, so use only
if you know what you're doing.}

\item{...}{Option and value, like \code{option=value} or \verb{option, value}; see the Examples}
}
\description{
Set and get global options for INLA
}
\examples{

 ## set number of threads
 inla.setOption("num.threads", "4:1:1")
 inla.setOption(num.threads="4:1:2")
 ## check it
 inla.getOption("num.threads")

}
\author{
Havard Rue \email{hrue@r-inla.org}
}
