% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{inla.option}
\alias{inla.option}
\alias{inla.options}
\alias{inla.setOption}
\alias{inla.getOption}
\title{Set and get global options for INLA}
\usage{
inla.getOption(
  option = c("inla.call", "inla.arg", "fmesher.call", "fmesher.arg", "num.threads",
    "blas.num.threads", "smtp", "mkl", "safe", "vecLib", "vecLibPath", "pardiso.license",
    "keep", "verbose", "save.memory", "working.directory", "silent", "debug", "cygwin",
    "ssh.auth.sock", "cygwin.home", "show.warning.graph.file", "scale.model.default",
    "short.summary", "inla.timeout", "fmesher.timeout", "inla.mode")
)

inla.setOption(...)
}
\arguments{
\item{option}{The option to get. If \code{option = NULL} then
\code{inla.getOption} then \code{inla.getOption} will return a named list of
current values, otherwise, \code{option} must be one of
\describe{
\item{inla.call}{The path to the inla-program.}

\item{inla.arg}{Additional arguments to \code{inla.call}}

\item{fmesher.call}{The path to the fmesher-program}

\item{fmesher.arg}{Additional arguments to \code{fmesher.call}}

\item{num.threads}{Character string with the number of threads to use as
\code{A:B}, see \code{?inla}}

\item{blas.num.threads}{Number of threads to use for openblas and mklblas (see
\code{inla} for details)}

\item{smtp}{Sparse matrix library to use, one of \code{band}, \code{taucs}
(\code{default}) or \code{pardiso}}

\item{mkl}{Use binaries buildt with Intel MKL?  (If possible)}

\item{safe}{Run in safe-mode (ie try to automatically fix convergence errors)
(default \code{TRUE})}

\item{vecLib}{This option applies to Mac only. If TRUE and mkl=FALSE, link with
vecLib BLAS and LAPACK libs (if available)}

\item{vecLibPath}{This option applies to Mac only. Path to vecLib-libraries. If
empty, use default.}

\item{pardiso.license}{The full path to the PARDISO license file or a
newline-separated string with license key(s)}

\item{keep}{Keep temporary files?}

\item{verbose}{Verbose output?}

\item{save.memory}{Save memory at the cost of (minor) accuracy and computing time?}

\item{working.directory}{The name of the working directory.}

\item{silent}{Run the inla-program in a silent mode?}

\item{debug}{Run the inla-program in a debug mode?}

\item{cygwin}{The home of the Cygwin installation (default "C:/cygwin") (Remote
computing for Windows only)}

\item{ssh.auth.sock}{The ssh bind-adress (value of $SSH_AUTH_SOCK int the
Cygwin-shell). (Remote computing for Windows only)}

\item{show.warning.graph.file}{Give a warning for using the obsolete argument
\code{graph.file} instead of \code{graph}}

\item{scale.model.default}{The default value of argument \code{scale.model} which
optionally scale intrinisic models to have generalized unit average variance}

\item{short.summary}{Use a less verbose output for \code{summary}. Useful for
Markdown documents.}

\item{inla.timeout}{The timeout limit, in whole seconds, for calls to the inla
binary. Default is 0, meaning no timeout limit.  Set to a positive integer
to terminate inla calls if they run to long.  Fractional seconds are rounded
up to the nearest integer. This feature is EXPERIMENTAL and might change at
a later stage.}

\item{fmesher.timeout}{The timeout limit, in whole seconds, for calls to the
fmesher binary. Default is 0, meaning no timeout limit.  Set to a positive
integer to terminate fmesher calls that may enter infinite loops due to
special geometry regularity. Fractional seconds are rounded up to the
nearest integer.}

\item{inla.mode}{Which mode to use in INLA? Default is \code{"compact"}. Other
options are \code{"classic"} and \code{"twostage"}.}
}}

\item{...}{Option and value, like \code{option=value} or \code{option,
value}; see the Examples}
}
\description{
Set and get global options for INLA
}
\examples{

 ## set number of threads
 inla.setOption("num.threads", "4:1")
 ## alternative format
 inla.setOption(num.threads="4:1")
 ## check it
 inla.getOption("num.threads")

}
\author{
Havard Rue \email{hrue@r-inla.org}
}
