% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{inla.option}
\alias{inla.option}
\alias{inla.options}
\alias{inla.setOption}
\alias{inla.getOption}
\title{Set and get global options for INLA}
\usage{
inla.getOption(
  option = c("inla.call", "inla.arg", "fmesher.call", "fmesher.arg", "num.threads",
    "smtp", "safe", "keep", "verbose", "save.memory", "internal.opt",
    "working.directory", "silent", "debug", "show.warning.graph.file",
    "scale.model.default", "short.summary", "inla.timeout", "fmesher.timeout",
    "inla.mode", "malloc.lib", "fmesher.evolution", "fmesher.evolution.warn",
    "fmesher.evolution.verbosity")
)

inla.setOption(...)
}
\arguments{
\item{option}{The option to get. If \code{option = NULL} then
\code{inla.getOption} then \code{inla.getOption} will return a named list of
current values, otherwise, \code{option} must be one of
\describe{
\item{inla.call}{The path to the inla-program.}

\item{inla.arg}{Additional arguments to \code{inla.call}}

\item{fmesher.call}{The path to the fmesher-program}

\item{fmesher.arg}{Additional arguments to \code{fmesher.call}}

\item{num.threads}{Character string with the number of threads to use as
\code{A:B}, see \code{?inla}}

\item{smtp}{Sparse matrix library to use, one of \code{band}, \code{taucs}
(\code{default}) or \code{pardiso}}

\item{safe}{Run in safe-mode (ie try to automatically fix convergence errors)
(default \code{TRUE})}

\item{keep}{Keep temporary files?}

\item{verbose}{Verbose output?}

\item{save.memory}{Save memory at the cost of (minor) accuracy and computing time?}

\item{internal.opt}{Do internal online optimisations or not}

\item{working.directory}{The name of the working directory.}

\item{silent}{Run the inla-program in a silent mode?}

\item{debug}{Run the inla-program in a debug mode?}

\item{cygwin}{The home of the Cygwin installation (default "C:/cygwin") (Remote
computing for Windows only) (No longer in use!)}

\item{ssh.auth.sock}{The ssh bind-adress (value of $SSH_AUTH_SOCK int the
Cygwin-shell). (Remote computing for Windows only)}

\item{show.warning.graph.file}{Give a warning for using the obsolete argument
\code{graph.file} instead of \code{graph}}

\item{scale.model.default}{The default value of argument \code{scale.model} which
optionally scale intrinisic models to have generalized unit average variance}

\item{short.summary}{Use a less verbose output for \code{summary}. Useful for
Markdown documents.}

\item{inla.timeout}{The timeout limit, in whole seconds, for calls to the inla
binary. Default is 0, meaning no timeout limit.  Set to a positive integer
to terminate inla calls if they run to long.  Fractional seconds are rounded
up to the nearest integer. This feature is EXPERIMENTAL and might change at
a later stage.}

\item{fmesher.timeout}{The timeout limit, in whole seconds, for calls to the
fmesher binary. Default is 0, meaning no timeout limit.  Set to a positive
integer to terminate fmesher calls that may enter infinite loops due to
special geometry regularity. Fractional seconds are rounded up to the
nearest integer.}

\item{inla.mode}{Which mode to use in INLA? Default is \code{"compact"}. Other
options are \code{"classic"} and \code{"twostage"}.}

\item{malloc.lib}{Which malloc library to use: \code{"je"}, \code{"tc"}, \code{"mi"}, \code{"compiler"} or \code{"default"}.
Option \code{"compiler"} use the compiler's implementation. The library
is loaded using \code{LD_PRELOAD} and similar functionality. Loosely, \code{jemalloc} is from Facebook,
\code{tcmalloc} is from Google and \code{mimalloc} is from Microsoft. This option is not available for
Windows and not all options might be available for every arch.
If \code{malloc.lib} is a complete path to an external library, that file will be used
instead of one of the supported ones.}

\item{fmesher.evolution}{Control use of fmesher methods during the transition
to a separate fmesher package. Levels of
\code{fmesher.evolution}:

\code{1L} uses the intermediate \verb{fm_*} methods in \code{fmesher} that were already
available via \code{inlabru} from 2.8.0.

\code{2L} (current default) uses the full range of \code{fmesher} package methods.

Further levels may be added as the package development progresses.}

\item{fmesher.evolution.warn}{logical; whether to show warnings about deprecated
use of legacy INLA methods with fmesher package replacements. When \code{TRUE},
shows deprecation messages for many CRS and mesh
related methods, pointing to their \verb{fm_*} replacements. Default is currently \code{FALSE}.}

\item{fmesher.evolution.verbosity}{logical or character; at what minimum
severity to show warnings about deprecated
use of legacy INLA methods with fmesher package replacements.
When set to "default" (default), "soft", "warn", or "stop", indicates the
minimum warning level used when \code{fmesher.evolution.warn} is \code{TRUE}.}
}}

\item{...}{Option and value, like \code{option=value} or \verb{option, value}; see the Examples}
}
\description{
Set and get global options for INLA
}
\examples{

 ## set number of threads
 inla.setOption("num.threads", "4:1")
 ## alternative format
 inla.setOption(num.threads="4:1")
 ## check it
 inla.getOption("num.threads")

}
\author{
Havard Rue \email{hrue@r-inla.org}
}
