%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from pardiso.R
%%
\name{inla.pardiso}
\alias{inla.pardiso}
\title{PARDISO support in R-INLA}
\description{
Describe the PARDISO support in R-INLA
}
\usage{
inla.pardiso()
}
\arguments{
}
\details{
Describe the PARDISO support in R-INLA, how to get the
license key and enable it in the R-INLA package.
}
\value{
}
\author{Havard Rue \email{hrue@r-inla.org}}
