%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from pardiso.R
%%
\name{inla.pardiso}
\alias{inla.pardiso}
\alias{inla.pardiso.check}
\title{PARDISO support in R-INLA}
\description{
Describe and check the PARDISO support in R-INLA
}
\usage{
inla.pardiso()
inla.pardiso.check()
}
\arguments{
}
\details{
\code{inla.pardiso()} describes the \code{PARDISO} support in R-INLA, how to get the license
key and enable it in the \code{R-INLA} package. \code{inla.pardiso.check()} check if the
\code{PARDISO} support is working.
}
\value{
}
\author{Havard Rue \email{hrue@r-inla.org}}
