%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from plot.inla.R
%%
 \name{plot.inla}
 \alias{plot.inla}
 \alias{inla.plot}
 \title{Default INLA plotting}
 \description{
   Takes am \code{inla} object produced by \code{inla} and plot the results
 }
 \usage{
 \method{plot}{inla}(x, 
              plot.fixed.effects = TRUE,
              plot.lincomb = TRUE,
              plot.random.effects = TRUE,
              plot.hyperparameters = TRUE,
              plot.predictor = TRUE,
              plot.q = TRUE,
              plot.cpo = TRUE,
              postscript = FALSE,
              pdf = FALSE, 
              ...)
 }
 \arguments{
   \item{x}{A fitted  \code{inla} object produced by \code{inla} }
   \item{plot.fixed.effects}{Boolean indicating if posterior marginals
     for the fixed effects in the model should be plotted }
   \item{plot.lincomb}{Boolean indicating if posterior marginals
     for the linear combinations should be plotted }
   \item{plot.random.effects}{Boolean indicating if posterior mean and quantiles
     for the random effects in the model should be plotted  }
   \item{plot.hyperparameters}{Boolean indicating if posterior marginals
     for the hyperparameters in the model should be plotted }
   \item{plot.predictor}{Boolean indicating if posterior mean and quantiles
     for the linear predictor in the model should be plotted }
   \item{plot.q}{Boolean indicating if precision matrix should be displayed}
   \item{plot.cpo}{Boolean indicating if CPO/PIT valuesshould be plotted}
   \item{single}{Boolean indicating if there should be more than one plot per page
                 (FALSE) or just one (TRUE)}
   \item{postscript}{Boolean indicating if postscript files should be produced instead.
                     The return value is the directory where the files are stored.}
   \item{pdf}{Boolean indicating if PDF files should be produced instead.
                     The return value is the directory where the files are stored.}
   \item{...}{Additional arguments}
 }
 \author{Havard Rue \email{hrue@math.ntnu.no} }
 \seealso{\code{\link{inla}}}
 \examples{
     result = inla(...)
     plot(result)
     plot(result, single=TRUE)
     plot(result, single=TRUE, pdf=TRUE, paper = "a4")
 }
 \keyword{plot}
