%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from posterior.sample.R
%%
 \name{inla.sample}
 \alias{inla.posterior.sample}
 \alias{posterior.sample}
 
 \title{Generate samples from an approximated posterior of a fitted model}
 
 \description{This function generate samples from an approximated posterior
              of a fitted model (an inla-object}
 \usage{
     inla.posterior.sample(n = 1L, result, intern = FALSE, use.improved.mean = TRUE,
                           add.names = TRUE)
 }
 
 \arguments{
   \item{n}{Number of samples.}
   \item{result}{The inla-object, ie the output from an \code{inla}-call.
                 The \code{inla}-object must be created with
                 \code{control.compute=list(config=TRUE)}.}
   \item{use.improved.mean}{Logical. If \code{TRUE} then use the
   marginal mean values when constructing samples. If \code{FALSE}
   then use the mean in the Gaussian approximations.}
  \item{intern}{Logical. If \code{TRUE} then produce samples in the
  internal scale for the hyperparmater, if \code{FALSE} then produce
  samples in the user-scale. (For example log-precision (intern)
  and precision (user-scale))}
   \item{add.names}{Logical. If \code{TRUE} then add name for each elements of each
                sample. If \code{FALSE}, only add name for the first sample. 
                (This save space.)}
}
\value{ A list of the samples, where each sample is a list with
  names \code{hyperpar} and \code{latent}, and with their marginal
  densities in \code{logdens$hyperpar} and \code{logdens$latent}
  and the joint density is in \code{logdens$joint}.  THIS IS AN
  EXPERIMENTAL FUNCTION AND CHANGES MAY APPEAR AT ANY TIME!  }
\author{Havard Rue \email{hrue@math.ntnu.no}}
 
\examples{
  r = inla(y ~ 1 ,data = data.frame(y=rnorm(1)), control.compute = list(config=TRUE))
  samples = inla.posterior.sample(2,r)
}
