\name{PRborder}
\alias{PRborder}
\alias{PRborder}
\docType{data}
\title{The PRborder data}
\description{
  A data matrix with Longitude and Latitude coordinates for the boundary
  of Parana State. 
}
\usage{data(PRborder)}
\format{
  \describe{
    \item{Longtiude}{The Longtiude coordinate}
    \item{Latitude}{The Latitude coordinate}
  }
}

\seealso{PRprec}

%%\examples{}
\keyword{datasets}
