%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from print.R
%%
\name{print.inla}
\alias{print.inla}
\title{Print a INLA fit }
\description{
Print a INLA fit
}
\usage{
\method{print}{inla}(x,...)
}
\arguments{
  \item{x}{An inla-object (output from an \code{\link{inla}}-call).}
  \item{...}{ other arguments.}
}
\details{
 None
}
\value{
 None
}
\author{Havard Rue}

\seealso{ \code{\link{inla}} }
\examples{
## None
}
