% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{inla.prune}
\alias{inla.prune}
\alias{prune}
\title{Prune the INLA-package}
\usage{
inla.prune(ask = TRUE)
}
\arguments{
\item{ask}{Logical. If TRUE, then ask for user confirmation before deleting.
If FALSE, then delete without user confirmation.}
}
\value{
No value is returned.
}
\description{
Prune the INLA-package by deleting binary files not supported by the running
OS
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
