% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qreordering.R
\name{qreordering}
\alias{qreordering}
\alias{inla.qreordering}
\title{Compute the reordering using the GMRFLib implementation}
\usage{
inla.qreordering(graph, reordering = inla.reorderings())
}
\arguments{
\item{graph}{A \verb{(inla-)graph} object}

\item{reordering}{The name of the reordering algorithm to be used; either
one of the names listed in \code{inla.reorderings()}.  The default is "auto"
which try several reordering algorithm and use the best one for this
particular matrix.}
}
\value{
\code{inla.qreordering} returns a list with the name of the
reordering algorithm used or found, the reordering code for the reordering
algorithm, the actual reordering and its inverse.
}
\description{
This function compute the reordering (or find the best reordering) using the
GMRFLib implementation
}
\examples{

 g = system.file("demodata/germany.graph", package="INLA")
 r = inla.qreordering(g)
 m = inla.graph2matrix(g)
 r = inla.qreordering(m)

}
\author{
Havard Rue \email{hrue@r-inla.org}
}
