%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from qsample.R
%%
 \name{qsample}
 \alias{inla.qsample}
 \alias{qsample}
 
 \title{Generate samples from a GMRF using the GMRFLib implementation}
 
 \description{This function generate samples from a GMRF using the GMRFLib implementation}
 \usage{
     inla.qsample(n, Q, reordering = "auto",  seed = 0L)
 }
 
 \arguments{
   \item{n}{Number of samples}
   \item{Q}{The precision matrix or a filename containing it.}
   \item{reordering}{The type of reordering algorithm to be used; either one of the names listed in \code{inla.reorderings()}
        or the output from \code{inla.qreordering(Q)}.
        The default is "auto" which try several reordering algorithm and use the best one for this particular matrix.}
   \item{seed}{The seed to be used,  where \code{seed=0L} means that GMRFLib should decide the seed.}
}
\value{
  \code{inla.qsample} returns a matrix where each column is a sample.
}
\author{Havard Rue \email{hrue@math.ntnu.no}}
 
\examples{
 g = system.file("demodata/germany.graph", package="INLA")
 G = inla.graph2matrix(g)
 diag(G) = dim(G)[1L]
 x = inla.qsample(10, G)
 matplot(x)
}
