%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from read.graph.R
%%
\name{read.graph}
\alias{read.graph}
\alias{write.graph}
\alias{inla.read.graph}
\alias{inla.write.graph}
\alias{inla.graph}
\title{Read and write a graph-object}
\description{Reads a graph-object to a file and write graph-object to file}
\usage{
g = inla.read.graph(graph)
inla.write.graph(graph, filename, mode = c("binary", "ascii""))
summary(graph)
plot(graph)
\arguments{
    \item{filename}{The filename of the graph.}
    \item{graph}{An \code{inla.graph}-object, a (sparse) symmetric matrix or a filename containing the graph.}
    \item{mode}{The mode of the file; ascii-file or a (gzip-compressed) binary. Default value depends on 
                the inla.option \code{internal.binary.mode} which is default \code{TRUE}; see \code{inla.setOption}.}
}
\value{
    The output of \code{inla.read.graph}, is an \code{inla.graph} object, with elements
    \item{n}{is the size of the graph}
    \item{nnbs}{is a vector with the number of neigbours}
    \item{nbs}{is a list-list with the neigbours}
    \item{filename}{is the name of the file if this \code{inla.graph}-object was read from a file}
    \item{cc}{list with connected component information (this entry can be auto-generated; see below)
        \itemize{
            \item{\code{id}}{is a vector with the connected component id for each node (starting from 1)}
            \item{\code{n}}{is the number of connected components}
            \item{\code{nodes}}{is a list-list of nodes belonging to each connected component}
        }
    }
    The connected component information,  can be generated from the rest of the graph-structure,
    using \code{graph = inla.add.graph.cc(graph)}.
    Methods implemented for \code{inla.graph} are \code{summary} and \code{plot}.
    The method \code{plot} require the libraries \code{Rgraphviz} and \code{graph} from the Bioconductor-project,
    see \code{http://www.bioconductor.org}.

    The \code{graph} is either a filename containing the \code{inla.graph}-object (previously written using
    \code{inla.write.graph()},  or an \code{inla.graph}-object,  or a (sparse) symmetric matrix whos non-zero pattern
    determines the graph.
}
\author{Havard Rue \email{hrue@math.ntnu.no}}
\seealso{
    \code{\link{inla.matrix2graph}}
    \code{\link{inla.graph2matrix}} 
    \code{\link{inla.spy}}
}
\examples{
cat("3 1 1 2 2 1 1 3 0\n", file="g.dat")
g = inla.read.graph("g.dat")
g.file = inla.write.graph(g, mode="binary")
gg = inla.read.graph(g.file)
summary(g)
summary(gg)
plot(gg)
inla.spy(gg)
ggg = inla.graph2matrix(gg)
ggg.file = inla.write.graph(ggg)
inla.spy(ggg.file)
}
