%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from remote.R
%%
\name{inla.ssh.copy.id}
\alias{inla.ssh.copy.id}
\alias{ssh.copy.id}
\alias{inla.remote}

\title{Setup remote computing}

\description{Initialize the definition file and print the path to the internal script to transfer ssh-keys}

\usage{
inla.remote()
inla.ssh.copy.id()
}
\arguments{
 None
}

\value{%%
 \code{inla.remote} is used once to setup the remote host information
 file (definition file) in the users home directory; see the FAQ entry on this
 issue for more information. 
 \code{inla.ssh.copy.id} will return the path to the internal script to transfer ssh-keys.
}
\author{Havard Rue \email{hrue@r-inla.org}}
\examples{
##See the FAQ entry on this issue on r-inla.org.
}
