%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from reorderings.R
%%
\name{inla.reorderings}
\alias{inla.reorderings}
\alias{reorderings}

\title{Reorderings methods for sparse matrices}

\description{Provide the names of all implemented reordering schemes}

\usage{
 inla.reorderings()
}

\arguments{
 None
}

\value{The names of all available reorderings}
%%

\author{Havard Rue \email{hrue@r-inla.org}}

\examples{
inla.reorderings()
}
