% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeneric.R
\name{rgeneric.define}
\alias{rgeneric.define}
\alias{rgeneric}
\alias{inla.rgeneric.define}
\alias{rgeneric.ar1.model}
\alias{inla.rgeneric.ar1.model}
\alias{rgeneric.iid.model}
\alias{inla.rgeneric.iid.model}
\alias{rgeneric.wrapper}
\alias{inla.rgeneric.wrapper}
\alias{rgeneric.q}
\alias{inla.rgeneric.q}
\alias{inla.rgeneric.ar1.model.opt}
\title{rgeneric models}
\usage{
inla.rgeneric.ar1.model(
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  theta = NULL
)

inla.rgeneric.ar1.model.opt(
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  theta = NULL
)

inla.rgeneric.iid.model(
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  theta = NULL
)

inla.rgeneric.define(
  model = NULL,
  debug = FALSE,
  compile = TRUE,
  optimize = FALSE,
  ...
)

inla.rgeneric.wrapper(
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  model,
  theta = NULL
)

inla.rgeneric.q(
  rmodel,
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  theta = NULL
)
}
\arguments{
\item{cmd}{An allowed request}

\item{theta}{Values of theta}

\item{model}{The definition of the model; see \code{inla.rgeneric.ar1.model}}

\item{debug}{Logical. Enable debug output}

\item{compile}{Logical. Compile the definition of the model or not.}

\item{optimize}{Logical. With this option \code{TRUE}, then \code{model} pass
only the values of \code{Q} and not the whole matrix.  Please see the
vignette for details and \code{inla.rgeneric.ar1.model.opt} for an example.}

\item{...}{Named list of variables that defines the environment of
\code{model}}

\item{rmodel}{The rgeneric model-object, the output of
\code{inla.rgeneric.define}}
}
\value{
This allows a latent model to be defined in \code{R}.  See
\code{inla.rgeneric.ar1.model} and \code{inla.rgeneric.iid.model} and the
documentation for worked out examples of how to define latent models in this
way.  This will be somewhat slow and is intended for special cases and
protyping. The function \code{inla.rgeneric.wrapper} is for internal use
only.
}
\description{
A framework for defining latent models in R
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
