%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from rgeneric.R
%%
\name{rgeneric.define}
\alias{rgeneric}
\alias{rgeneric.define}
\alias{inla.rgeneric.define}
\alias{rgeneric.ar1.model}
\alias{inla.rgeneric.ar1.model}

\title{rgeneric models}

\description{A framework for defining latent models in R}

\usage{
inla.rgeneric.define(model.def = NULL, ...)
inla.rgeneric.ar1.model(
        cmd = c("graph", "Q", "initial", "log.norm.const", "log.prior", "quit"),
        theta = NULL,
        args = NULL)
}

\arguments{

  \item{model.def}{The definition of the model; see \code{inla.rgeneric.ar1.model}}
  \item{cmd}{The allowed commands}
  \item{theta}{Values of theta}
  \item{args}{Further args}
  \item{...}{Further args}
}

\value{%%
  This is a test-implementation of how a latent model can be
  defined within R, and details would likely change in the future.
  See \code{inla.rgeneric.ar1.model} and the documentation for a
  worked through example of how to define the AR1 model this way.
  Using this function require the \code{multicore}-package, and
  only runs smoothly under Linux.}
\author{Havard Rue \email{hrue@math.ntnu.no}}
