\name{SPDEtoy}
\alias{SPDEtoy}
\docType{data}
\title{toy simulated data set for the SPDE tutorial}
\description{
  Simulated data set on 200 location points. 
  The simulation process is made at the introduction 
  of the SPDE tutorial.
}
\usage{data(SPDEtoy)}
\format{
  A data frame with 200 observations on the following 3 variables.
  \describe{
    \item{\code{s1}}{First element of the coordinates}
    \item{\code{s2}}{Second element of the coordinates}
    \item{\code{y}}{data simulated at the locations}
  }
}

\source{
SPDE tutorial 
}

\examples{
data(SPDEtoy)

}
\keyword{datasets}
