% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.scopy.R
\name{summary.scopy}
\alias{summary.scopy}
\alias{inla.summary.scopy}
\title{Computes the mean and stdev for the spline from \code{scopy}}
\usage{
inla.summary.scopy(result, name, by = 0.05, range = c(0, 1))
}
\arguments{
\item{result}{An \code{inla}-object, ie the output from an \code{inla()}
call}

\item{name}{The name of the \code{scopy} model component see \code{?inla::f}
and argument \code{extraconstr}}

\item{by}{The resolution of the results, in the scale where distance between
two nearby locations is 1}

\item{range}{The range of the locations, in \verb{(from, to)}}
}
\value{
A \code{data.frame} with locations, mean and stdev.  if \code{name}
is not found, NULL is returned.
}
\description{
This function computes the mean and stdev for the spline function that is
implicite from an \code{scopy} model component
}
\examples{

 ## see example in inla.doc("scopy")
 
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
