\name{Surg}
\alias{Surg}
\docType{data}
\title{ Surgical: Institutional ranking}
\description{
This example considers mortality rates in 12 hospitals performing cardiac surgery in babies
}
\usage{data(Surg)}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{n}}{Number of deaths}
    \item{\code{r}}{Total number of cases}
    \item{\code{hospital}}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{E} \code{F} \code{G} \code{H} \code{I} \code{J} \code{K} \code{L}}
  }
}

\source{
WinBUGS/OpenBUGS Manual Examples Vol. I
}

\examples{
data(Surg)

}
\keyword{datasets}
