\name{SurvSim}
\alias{SurvSim}
\docType{data}
\title{Survival data}
\description{
  Simulated data set for Weibull survival model

}
\usage{data(SurvSim)}
\format{
  A data frame with 100 observations on the following 3 variables.
  \describe{
    \item{\code{y}}{a numeric vector of survival times}
    \item{\code{cens}}{a numeric vector of event indicator (0=censured 1=failure)}
    \item{\code{x}}{a numeric vector of covariate }
  }
}
\keyword{datasets}
