\name{Tokyo}
\alias{Tokyo}
\docType{data}
\title{Binomial time series}
\description{
 Recorded days of rain above 1 mm in Tokyo for 2 years, 1983:84
}
\usage{data(Tokyo)}
\format{
  A data frame with 366 observations on the following 3 variables.
  \describe{
    \item{\code{y}}{number of days with rain}
    \item{\code{n}}{total number of days}
    \item{\code{time}}{day of the year}
  }
}

\source{
\url{http://www.math.ntnu.no/~hrue/GMRF-book/tokyo.rainfall.data.dat}
}
\references{
Rue, H and Held, L. (2005)
  \emph{Gaussian Markov Random Fields - Theory and Applications}
  Chapman and Hall
}
\examples{
data(Tokyo)

}
\keyword{datasets}
