% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{inla.update}
\alias{inla.update}
\alias{inla.upgrade}
\title{Upgrade the INLA-package}
\usage{
inla.update(lib = NULL, testing = FALSE, ask = TRUE)

inla.upgrade(lib = NULL, testing = FALSE, ask = TRUE)
}
\arguments{
\item{lib}{Location to install the library.}

\item{testing}{If \code{TRUE}, then look for a test-version if the
INLA-package.}

\item{ask}{same argument as in \code{update.packages}}
}
\value{
\code{inla.upgrade} will update the INLA package to the current
version, and \code{inla.update} do the same for backward compatibility. This
function is simple wrapper for \code{update.packages} using the INLA
repository.
}
\description{
Functions to upgrade the \code{INLA}-package to the current version.
}
\seealso{
\code{update.packages}
}
\author{
Havard Rue \email{hrue@r-inla.org}
}
