%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from upgrade.R
%%
\name{inla.upgrade}
\alias{inla.upgrade}
\alias{inla.update}

\title{Upgrade the INLA-package}

\description{Functions to upgrade the \code{INLA}-package to the
  current version.}

\usage{
inla.upgrade(lib = NULL, testing=FALSE, ask = TRUE)
inla.update(lib = NULL, testing=FALSE, ask = TRUE)
}

\arguments{
  \item{lib}{Location to install the library.}
  \item{testing}{If \code{TRUE}, then look for a test-version if the
    INLA-package.}
  \item{ask}{same argument as in \code{update.packages}}
}
\value{\code{inla.upgrade} will update the INLA package to the current version, 
 and \code{inla.update} do the same 
  for backward compatibility. This function is simple wrapper for
  \code{update.packages} using the INLA repository.}
\author{Havard Rue \email{hrue@r-inla.org}}
\seealso{\code{update.packages}}
