% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{inla.version}
\alias{inla.version}
\alias{version}
\title{Show the version of the INLA-package}
\usage{
inla.version(what = c("default", "version", "date"))
}
\arguments{
\item{what}{What to show version of}
}
\value{
\code{inla.version} display the current version information using
\code{cat} with \code{default} or \code{info}, or return other specific
requests through the call.
}
\description{
Show the version of the INLA-package
}
\examples{

## Summary of all
inla.version()
## The building date
inla.version("date")

}
\author{
Havard Rue \email{hrue@r-inla.org}
}
