%%
%% WARNING! DO NOT EDIT!
%% This file is automatically generated from version.R
%%
\name{inla.version}
\alias{version}
\alias{inla.version}

\title{Show the version of the INLA-package}

\description{Show the version of the INLA-package}

\usage{
inla.version(what = c("default",
                      "version", 
                      "info", 
                      "hgid",
                      "rinla",
                      "inla",
                      "date",
                      "bdate"))
}

\arguments{
  \item{what}{What to show version of}
}

\value{%%
  \code{inla.version} either display the current version information using \code{cat}
  with
  \code{default} or \code{info},  or return the version number/information
  for other spesific requests through the call.
}
%%

\author{Havard Rue \email{hrue@r-inla.org}}

\examples{
## Summary of all
inla.version()
## The building date
inla.version("bdate")
}
